% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars2.R, R/options.R
\name{gg_bar_DatNum}
\alias{gg_bar_DatNum}
\alias{chart_options}
\title{Bar Chart Date Numeric}
\usage{
gg_bar_DatNum(data, ...)
}
\arguments{
\item{data}{data.frame with the proper frtype}

\item{agg}{defaults to "sum"}

\item{agg_text}{defaults to NULL}

\item{caption}{defaults to NULL}

\item{colors}{defaults to NULL}

\item{color_scale}{defaults to discrete"}

\item{drop_na}{defaults to FALSE}

\item{highlight_value}{defaults to NULL}

\item{highlight_value_color}{defaults to '#F9B233'}

\item{hor_label}{defaults to NULL}

\item{hor_line}{defaults to NULL}

\item{label_ratio}{defaults to 1}

\item{label_wrap}{defaults to 12}

\item{marks}{defaults to c("." "")}

\item{n_digits}{defaults to NULL}

\item{order}{defaults to NULL}

\item{orientation}{defaults to "ver"}

\item{percentage}{defaults to FALSE}

\item{prefix}{defaults to NULL}

\item{slice_n}{defaults to NULL}

\item{sort}{defaults to "no"}

\item{subtitle}{defaults to NULL}

\item{suffix}{defaults to NULL}

\item{text_color}{defaults to "#5A6B72"}

\item{text_show}{defaults to TRUE}

\item{text_size}{defaults to 3}

\item{theme}{defaults to NULL}

\item{title}{defaults to NULL}

\item{ver_label}{defaults to NULL}

\item{ver_line}{defaults to NULL}

\item{opts}{defaults to NULL}
}
\value{
ggplot2 object
}
\description{
Bar Chart Date Numeric

Chart Options
}
\section{ctypes}{

Cat-Num, Dat-Num, Yea-Num
}

\examples{
gg_bar_CatNum(sampleData("Cat-Num", nrow = 10))
}
