% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Nu.R
\name{circleAreaPlotCCN}
\alias{circleAreaPlotCCN}
\title{circleAreaPlotCCN
circle area plot}
\usage{
circleAreaPlotCCN(data, titleLabel = "Report", xLabel = "Category",
  yLabel = "Category", leg_pos = "right")
}
\arguments{
\item{x}{A category.}

\item{y}{A category.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
circleAreaPlotCCN
circle area plot
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

