% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Nu.R
\name{flip_circleAreaPlotCCN}
\alias{flip_circleAreaPlotCCN}
\title{flip_circleAreaPlotCCN
flip circle area plot}
\usage{
flip_circleAreaPlotCCN(data, titleLabel = "Report", xLabel = "Category",
  yLabel = "Category", leg_pos = "top")
}
\arguments{
\item{x}{A category.}

\item{y}{A category.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
flip_circleAreaPlotCCN
flip circle area plot
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

