% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{gg_area_Cat}
\alias{gg_area_Cat}
\title{Area (categories)}
\usage{
gg_area_Cat(data = NULL, agg_text = NULL, caption = NULL,
  colors = NULL, color_opacity = 0.7, drop_na = FALSE,
  hor_label = NULL, hor_line = NULL, label_ratio = 1,
  label_wrap = 12, marks = c(".", ","), n_digits = NULL,
  order = NULL, orientation = "ver", percentage = FALSE,
  prefix = NULL, shape_size = 3, shape_type = 19, slice_n = NULL,
  sort = "no", start_zero = TRUE, subtitle = NULL, suffix = NULL,
  text_color = "#5A6B72", text_show = TRUE, text_size = 3,
  theme = NULL, title = NULL, ver_label = NULL, ver_line = NULL,
  opts = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Ggplot2 visualization
}
\description{
Compare category's levels
}
\section{ctypes}{

Cat
}

\examples{
gg_area_Cat(sampleData("Cat", nrow = 10))
}
