% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_area_CatDatNum .R
\name{gg_area_CatDatNum}
\alias{gg_area_CatDatNum}
\title{Area Chart Categories Yea Numeric}
\usage{
gg_area_CatDatNum(data, ...)
}
\arguments{
\item{data}{A data.frame}

\item{...}{
  Arguments passed on to \code{\link[dsvizopts:chart_viz_options]{dsvizopts::dsviz_default_opts}}
  \describe{
    \item{\code{}}{}
  }}

\item{orientation}{Doesn't do anything for this type of chart.}

\item{order}{doesn't do anything}
}
\description{
This chart does not allow for chaning orientation
}
\section{ctypes}{

Yea-Num, Yea-Num
}

\examples{
gg_area_CatDatNum(sample_data("Cat-Dat-Num", nrow = 10))
}
