% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{gg_area_CatNum}
\alias{gg_area_CatNum}
\title{Area (categories, numbers)}
\usage{
gg_area_CatNum(data = NULL, opts = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Ggplot visualization
}
\description{
Compare aggregations among category's levels
}
\section{ctypes}{

Cat-Num
}

\examples{
gg_area_CatNum(sampleData("Cat-Num", nrow = 10))
}
