% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_area_DatNum.R
\name{gg_area_Dat}
\alias{gg_area_Dat}
\title{Area Chart Cat}
\usage{
gg_area_Dat(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
This chart does not allow for chaning orientation
}
\examples{
gg_area_Dat(sample_data("Dat", nrow = 10))
}
