% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possible_viz.R
\name{gg_area_density_CatNum}
\alias{gg_area_density_CatNum}
\title{Filled density distribution (categories, numbers)}
\usage{
gg_area_density_CatNum(data, title = NULL, subtitle = NULL,
  caption = NULL, horLabel = NULL, verLabel = NULL, yLine = NULL,
  yLineLabel = NULL, dropNa = FALSE, theme = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Ggplot visualization
}
\description{
Compare categories distributions
}
\section{ctypes}{

Cat-Num
}

\examples{
gg_area_density_CatNum(sampleData("Cat-Num", nrow = 10))
}
