% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{gg_area_hor_facet_CatNum.}
\alias{gg_area_hor_facet_CatNum.}
\title{Horizontal area facet
Facet Horizontal Area}
\usage{
gg_area_hor_facet_CatNum.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal area facet
Facet Horizontal Area
}
\section{ftypes}{
 Cat-Num
}

\examples{
add(1, 1)
add(10, 1)
}
