% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possible_viz.R
\name{gg_area_hor_facet_CatNum}
\alias{gg_area_hor_facet_CatNum}
\title{Horizontal area facet
Facet Horizontal Area}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal area facet
Facet Horizontal Area
}
\section{ftypes}{
 Cat-Num
}

\examples{
gg_area_hor_facet_CatNum(sampleData("Cat-Num"))
}
