% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_area_multi_density_dist_CaNu.}
\alias{gg_area_multi_density_dist_CaNu.}
\title{gg_area_multi_density_dist_CaNu.
Filled Density Distribution}
\usage{
gg_area_multi_density_dist_CaNu.(data, titleLabel = "Report", xLabel = NULL,
  yLabel = "Count", fillLabel = NULL, leg_pos = "right", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_area_multi_density_dist_CaNu.
Filled Density Distribution
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

