% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{gg_area_Oca}
\alias{gg_area_Oca}
\title{Vertical area (ordered categories)}
\usage{
gg_area_Oca(data, title = NULL, subtitle = NULL, caption = NULL,
  horLabel = NULL, verLabel = NULL, yLine = NULL, yLineLabel = NULL,
  dropNa = FALSE, order = NULL, theme = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Ggplot visualization
}
\description{
Compare category's levels
}
\section{ctypes}{

Oca
}

\examples{
gg_area_Oca(sampleData("Cat", nrow = 10))
}
