% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{gg_area_OcaNum}
\alias{gg_area_OcaNum}
\title{Vertical area (ordered categories, numbers)}
\usage{
gg_area_OcaNum(data, title = NULL, subtitle = NULL, caption = NULL,
  horLabel = NULL, verLabel = NULL, yLine = NULL,
  yLineLabel = NULL, agg = "sum", dropNa = FALSE, order = NULL,
  percentage = FALSE, theme = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Ggplot visualization
}
\description{
Compare quantities among categories
}
\section{ctypes}{

Oca-Num
}

\examples{
gg_area_OcaNum(sampleData("Cat-Num", nrow = 10))
}
