% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca.R
\name{gg_area_stacked_100_hor_CaCa.}
\alias{gg_area_stacked_100_hor_CaCa.}
\title{gg_area_stacked_100_hor_CaCa.
Stacked horizontal Area 100pct}
\usage{
gg_area_stacked_100_hor_CaCa.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, leg_pos = "right",
  l_ncol = 1, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_area_stacked_100_hor_CaCa.
Stacked horizontal Area 100pct
}
\section{ftypes}{
 Ca-Ca
}
\examples{
add(1, 1)
add(10, 1)
}

