% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_area_stacked_100_hor_CaNu.}
\alias{gg_area_stacked_100_hor_CaNu.}
\title{Horizontal 100% stacked area
Stacked Horizontal Area 100}
\usage{
gg_area_stacked_100_hor_CaNu.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = "Index", yLabel = NULL, leg_pos = "right",
  angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal 100% stacked area
Stacked Horizontal Area 100
}
\section{ftypes}{
 Ca-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
