% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{gg_area_stacked_100_hor_CatCat}
\alias{gg_area_stacked_100_hor_CatCat}
\title{Horizontal 100% stacked area
Stacked horizontal Area 100pct}
\usage{
gg_area_stacked_100_hor_CatCat(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, fillLabel = NULL,
  leg_pos = "right", angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal 100% stacked area
Stacked horizontal Area 100pct
}
\section{ftypes}{
 Cat-Cat
}

\examples{
gg_area_stacked_100_hor_CatCat(sampleData("Cat-Cat"))
}
