% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{gg_area_stacked_100_hor_CatCatNum.}
\alias{gg_area_stacked_100_hor_CatCatNum.}
\title{Horizontal 100% stacked area
Stacked area 100}
\usage{
gg_area_stacked_100_hor_CatCatNum.(data, titleLabel = "",
  subtitle = "", caption = "", xLabel = NULL, yLabel = NULL,
  fillLabel = NULL, leg_pos = "right", aggregation = "sum",
  angle_x = 0, ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A category.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal 100% stacked area
Stacked area 100
}
\section{ftypes}{
 Cat-Cat-Num
}

\examples{
add(1, 1)
add(10, 1)
}
