% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{gg_area_stacked_100_hor_CatDatNum.}
\alias{gg_area_stacked_100_hor_CatDatNum.}
\title{Horizontal 100% stacked area
Stacked Horizontal Area 100}
\usage{
gg_area_stacked_100_hor_CatDatNum.(data, titleLabel = "",
  subtitle = "", caption = "", xLabel = NULL, yLabel = NULL,
  fillLabel = NULL, leg_pos = "right", angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal 100% stacked area
Stacked Horizontal Area 100
}
\section{ftypes}{
 Cat-Dat-Num
}

\examples{
add(1, 1)
add(10, 1)
}
