% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Da-Nu.R
\name{gg_area_stacked_100_ver_CaDaNu.}
\alias{gg_area_stacked_100_ver_CaDaNu.}
\title{Vertical 100% stacked area
Stacked Vertical Area 100}
\usage{
gg_area_stacked_100_ver_CaDaNu.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, fillLabel = NULL,
  leg_pos = "top", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical 100% stacked area
Stacked Vertical Area 100
}
\section{ftypes}{
 Ca-Da-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
