% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Nu.R
\name{gg_area_stacked_100_ver_CatCatNum.}
\alias{gg_area_stacked_100_ver_CatCatNum.}
\title{Vertical 100% stacked area
Stacked vertical Area 100}
\usage{
gg_area_stacked_100_ver_CatCatNum.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, fillLabel = NULL,
  leg_pos = "right", aggregation = "sum", angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical 100% stacked area
Stacked vertical Area 100
}
\section{ftypes}{
 Cat-Cat-Num
}
\examples{
add(1, 1)
add(10, 1)
}

