% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{gg_area_stacked_100_ver_CatDatNum}
\alias{gg_area_stacked_100_ver_CatDatNum}
\title{Vertical 100% stacked area
Stacked Vertical Area 100}
\usage{
gg_area_stacked_100_ver_CatDatNum(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, fillLabel = NULL,
  aggregation = "sum", leg_pos = "right", angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical 100% stacked area
Stacked Vertical Area 100
}
\section{ftypes}{
 Cat-Dat-Num
}

\examples{
gg_area_stacked_100_ver_CatDatNum(sampleData("Cat-Dat-Num"))
}
