% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca.R
\name{gg_area_stacked_hor_CaCa.}
\alias{gg_area_stacked_hor_CaCa.}
\title{gg_area_stacked_hor_CaCa.
Stacked horizontal Area}
\usage{
gg_area_stacked_hor_CaCa.(data, titleLabel = "Report", xLabel = NULL,
  yLabel = "Count", fillLabel = NULL, leg_pos = "top")
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_area_stacked_hor_CaCa.
Stacked horizontal Area
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

