% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Nu.R
\name{gg_area_stacked_hor_CaCaNu.}
\alias{gg_area_stacked_hor_CaCaNu.}
\title{Horizontal stacked area
Stacked horizontal Area}
\usage{
gg_area_stacked_hor_CaCaNu.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, leg_pos = "right",
  aggregation = "sum", angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal stacked area
Stacked horizontal Area
}
\section{ftypes}{
 Ca-Ca-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
