% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{gg_area_stacked_hor_CatCatNum.}
\alias{gg_area_stacked_hor_CatCatNum.}
\title{Horizontal stacked area
Stacked area}
\usage{
gg_area_stacked_hor_CatCatNum.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, leg_pos = "right",
  aggregation = "sum", angle_x = 0, ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A category.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal stacked area
Stacked area
}
\section{ftypes}{
 Cat-Cat-Num
}

\examples{
add(1, 1)
add(10, 1)
}
