% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Nu.R
\name{gg_area_stacked_ver_CaCaNu.}
\alias{gg_area_stacked_ver_CaCaNu.}
\title{Vertical stacked area
Stacked area}
\usage{
gg_area_stacked_ver_CaCaNu.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, leg_pos = "right",
  aggregation = "sum", angle_x = 0, ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A category.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical stacked area
Stacked area
}
\section{ftypes}{
 Ca-Ca-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
