% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_area_stacked_ver_CaNu.}
\alias{gg_area_stacked_ver_CaNu.}
\title{Vertical stacked area
Stacked Vertical Area}
\usage{
gg_area_stacked_ver_CaNu.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, leg_pos = "right",
  text = TRUE, color_text = "black", type = "count", angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical stacked area
Stacked Vertical Area
}
\section{ftypes}{
 Ca-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
