% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{gg_area_stacked_ver_CatCat.}
\alias{gg_area_stacked_ver_CatCat.}
\title{Vertical stacked area
Stacked vertical Area}
\usage{
gg_area_stacked_ver_CatCat.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, fillLabel = NULL,
  leg_pos = "right", angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical stacked area
Stacked vertical Area
}
\section{ftypes}{
 Cat-Cat
}

\examples{
add(1, 1)
add(10, 1)
}
