% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Da-Nu.R
\name{gg_area_stepped_DatNum.}
\alias{gg_area_stepped_DatNum.}
\title{Stepped area
stepped stacked area.}
\usage{
gg_area_stepped_DatNum.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, angle_x = 0, ...)
}
\arguments{
\item{x}{A data.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Stepped area
stepped stacked area.
}
\section{ftypes}{
 Dat-Num
}

\examples{
add(1, 1)
add(10, 1)
}
