% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{gg_area_stepped_DatNum}
\alias{gg_area_stepped_DatNum}
\title{Stepped area
Stepped stacked area}
\usage{
gg_area_stepped_DatNum(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, angle_x = 0, ...)
}
\arguments{
\item{x}{A data.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Stepped area
Stepped stacked area
}
\section{ftypes}{
 Dat-Num
}

\examples{
gg_area_stepped_DatNum(sampleData("Dat-Num"))
}
