% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_area_ver_facet_CaNu.}
\alias{gg_area_ver_facet_CaNu.}
\title{gg_area_ver_facet_CaNu.
Facet Vertical Area}
\usage{
gg_area_ver_facet_CaNu.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = "Index", yLabel = NULL, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_area_ver_facet_CaNu.
Facet Vertical Area
}
\section{ftypes}{
 Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

