% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{gg_area_ver_facet_CatNum}
\alias{gg_area_ver_facet_CatNum}
\title{Vertical area facet
Facet Vertical Area}
\usage{
gg_area_ver_facet_CatNum(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical area facet
Facet Vertical Area
}
\section{ftypes}{
 Cat-Num
}

\examples{
gg_area_ver_facet_CatNum(sampleData("Cat-Num"))
}
