% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_area_YeaNum.R
\name{gg_area_Yea}
\alias{gg_area_Yea}
\title{Area Chart Cat}
\usage{
gg_area_Yea(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
This chart does not allow for chaning orientation
}
\examples{
gg_area_Cat(sample_data("Cat", nrow = 10))
}
