% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_bar_CatCatNum.R
\name{gg_bar_CatCatNum}
\alias{gg_bar_CatCatNum}
\title{Bar Chart Cat Cat Num}
\usage{
gg_bar_CatCatNum(data, ...)
}
\arguments{
\item{data}{A data.frame}

\item{...}{
  Arguments passed on to \code{\link[dsvizopts:chart_viz_options]{dsvizopts::dsviz_default_opts}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
This chart does not allow for chaning orientation
}
\section{ctypes}{

Cat-Cat-Num
}

\examples{
gg_bar_CatNum(sample_data("Cat-Num", nrow = 10))
}
