% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{gg_bar_CatDatNum}
\alias{gg_bar_CatDatNum}
\title{Bar (dates, numbers)}
\usage{
gg_bar_CatDatNum(data, title = NULL, subtitle = NULL, caption = NULL,
  horLabel = NULL, verLabel = NULL, horLine = NULL, verLine = NULL,
  agg = "sum", colors = NULL, colorText = "black",
  colorScale = "discrete", dropNaV = c(FALSE, FALSE), format = c("",
  ""), graphType = "grouped", labelRatio = 0.5, labelWrapV = c(12,
  12), legendPosition = "right", legendTitle = NULL, marks = c(".",
  ","), nDigits = 0, order1 = NULL, order2 = NULL,
  orientation = "ver", percentage = FALSE, showText = TRUE,
  theme = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Ggplot2 visualization
}
\description{
Compare quantities among categories over dates
}
\section{ctypes}{

Cat-Dat-Num
}

\examples{
gg_bar_CatDatNum(sampleData("Cat-Dat-Num", nrow = 10))
}
