% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{gg_bar_CatNumP}
\alias{gg_bar_CatNumP}
\title{Bar (ordered category, n numbers)}
\usage{
gg_bar_CatNumP(data = NULL, opts = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Ggplot2 visualization
}
\description{
Compare n quantities among category's levels
}
\section{ctypes}{

Cat-NumP
}

\examples{
gg_bar_CatNumP(sampleData("Cat-NumP", nrow = 10))
}
