% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_bar_CatYeaNum.R
\name{gg_bar_CatYea}
\alias{gg_bar_CatYea}
\title{bar Chart Cat Yea}
\usage{
gg_bar_CatYea(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
This chart does not allow for chaning orientation
}
\examples{
gg_bar_CatYea(sample_data("Cat-Yea", nrow = 10))
}
