% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca.R
\name{gg_bar_circular_Ca.}
\alias{gg_bar_circular_Ca.}
\title{gg_bar_circular_Ca.
Circular Bar}
\usage{
gg_bar_circular_Ca.(data, titleLabel = "", fillLabel = NULL,
  leg_pos = "right", width = 0.85)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_bar_circular_Ca.
Circular Bar
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

