% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_bar_circular_CaNu.}
\alias{gg_bar_circular_CaNu.}
\title{gg_bar_circular_CaNu.
Circular Bar}
\usage{
gg_bar_circular_CaNu.(data, titleLabel = "", subtitle = "", caption = "",
  leg_pos = "right", width = 0.85, aggregation = "sum", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_bar_circular_CaNu.
Circular Bar
}
\section{ftypes}{
 Ca-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
