% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Nu-bars.R
\name{gg_bar_circular_facet_CaCaNu.}
\alias{gg_bar_circular_facet_CaCaNu.}
\title{Circular bar facet
Circular Bar}
\usage{
gg_bar_circular_facet_CaCaNu.(data, titleLabel = "", subtitle = "",
  caption = "", fillLabel = NULL, leg_pos = "right",
  aggregation = "sum", width = 0.85, text = TRUE, type = "count",
  color_text = "black", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Circular bar facet
Circular Bar
}
\section{ftypes}{
 Ca-Ca-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
