% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possible_viz.R
\name{gg_bar_coloured_facet_x_ver_CatCatNum}
\alias{gg_bar_coloured_facet_x_ver_CatCatNum}
\title{Vertical facet bar coloured by first variable
vertical bar}
\usage{
gg_bar_coloured_facet_x_ver_CatCatNum(data, titleLabel = "",
  subtitle = "", caption = "", xLabel = NULL, yLabel = NULL,
  fillLabel = NULL, leg_pos = "right", angle_x = 0,
  aggregation = "sum", text = TRUE, color_text = "black",
  type = "count", ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical facet bar coloured by first variable
vertical bar
}
\section{ftypes}{
 Cat-Cat-Num
}

\examples{
gg_bar_coloured_facet_x_ver_CatCatNum(sampleData("Cat-Cat-Num"))
}
