% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{gg_bar_coloured_facet_y_hor_CatCatNum}
\alias{gg_bar_coloured_facet_y_hor_CatCatNum}
\title{Horizontal facet bar coloured by second variable
horizontal bar}
\usage{
gg_bar_coloured_facet_y_hor_CatCatNum(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, fillLabel = NULL,
  leg_pos = "right", angle_x = 0, aggregation = "sum", text = TRUE,
  color_text = "black", type = "count", ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal facet bar coloured by second variable
horizontal bar
}
\section{ftypes}{
 Cat-Cat-Num
}

\examples{
gg_bar_coloured_facet_y_hor_CatCatNumsampleData("Cat-Cat-Num"))
}
