% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possible_viz.R
\name{gg_bar_coloured_facet_y_ver_CatCat.}
\alias{gg_bar_coloured_facet_y_ver_CatCat.}
\title{Vertical bar facet coloured by second variable
Facet Vertical coloured bars}
\usage{
gg_bar_coloured_facet_y_ver_CatCat.(data, titleLabel = "",
  subtitle = "", caption = "", xLabel = NULL, yLabel = NULL,
  fillLabel = NULL, leg_pos = "right", angle_x = 0, text = TRUE,
  type = "count", color_text = "black", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical bar facet coloured by second variable
Facet Vertical coloured bars
}
\section{ftypes}{
 Cat-Cat
}

\examples{
gg_bar_coloured_facet_y_ver_CatCat.(sampleData("Cat-Cat"))
}
