% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca.R
\name{gg_bar_coloured_hor_Ca.}
\alias{gg_bar_coloured_hor_Ca.}
\title{Horizontal coloured bar
Horizontal coloured Bars}
\usage{
gg_bar_coloured_hor_Ca.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, fillLabel = NULL,
  text = TRUE, type = "count", color_text = "black", leg_pos = "right",
  angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal coloured bar
Horizontal coloured Bars
}
\section{ftypes}{
 Ca
}

\examples{
add(1, 1)
add(10, 1)
}
