% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{gg_bar_coloured_hor_Cat.}
\alias{gg_bar_coloured_hor_Cat.}
\title{Horizontal coloured bar
Horizontal coloured Bars}
\usage{
gg_bar_coloured_hor_Cat.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, text = TRUE,
  type = "count", color_text = "black", leg_pos = "right", angle_x = 0,
  ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal coloured bar
Horizontal coloured Bars
}
\section{ftypes}{
 Cat
}

\examples{
add(1, 1)
add(10, 1)
}
