% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{gg_bar_coloured_parameter_facet_hor_CatCat}
\alias{gg_bar_coloured_parameter_facet_hor_CatCat}
\title{Horizontal bar facet highlighting some parameter
Facet Horizontal coloured by parameter Bars}
\usage{
gg_bar_coloured_parameter_facet_hor_CatCat(data, titleLabel = "",
  subtitle = "", caption = "", xLabel = NULL, yLabel = NULL,
  parameter1 = NULL, parameter2 = NULL, leg_pos = "right", angle_x = 0,
  text = TRUE, type = "count", color_text = "black", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal bar facet highlighting some parameter
Facet Horizontal coloured by parameter Bars
}
\section{ftypes}{
 Cat-Cat
}

\examples{
gg_bar_coloured_parameter_facet_hor_CatCat(sampleData("Cat-Cat"))
}
