% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possible_viz.R
\name{gg_bar_coloured_parameter_facet_hor_CatCatNum.}
\alias{gg_bar_coloured_parameter_facet_hor_CatCatNum.}
\title{Horizontal facet bar coloured highliting some parameter
Facet Horizontal coloured by parameter Bars}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal facet bar coloured highliting some parameter
Facet Horizontal coloured by parameter Bars
}
\section{ftypes}{
 Cat-Cat-Num
}

\examples{
gg_bar_coloured_parameter_facet_hor_CatCatNum.(sampleData("Cat-Cat-Num"))
}
