% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_bar_coloured_parameter_hor_CaNu.}
\alias{gg_bar_coloured_parameter_hor_CaNu.}
\title{Horizontal bar highlighting some parameter
Horizontal coloured by parameter Bars}
\usage{
gg_bar_coloured_parameter_hor_CaNu.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, order = FALSE,
  parameter = NULL, line_mean = FALSE, text = TRUE, type = "count",
  color_text = "black", leg_pos = "right", aggregation = "sum",
  angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal bar highlighting some parameter
Horizontal coloured by parameter Bars
}
\section{ftypes}{
 Ca-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
