% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{gg_bar_coloured_parameter_hor_Cat}
\alias{gg_bar_coloured_parameter_hor_Cat}
\title{Horizontal bar highlighting some parameter
Horizontal coloured by parameter Bars}
\usage{
gg_bar_coloured_parameter_hor_Cat(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, parameter = NULL,
  text = TRUE, type = "count", color_text = "black", leg_pos = "right",
  angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal bar highlighting some parameter
Horizontal coloured by parameter Bars
}
\section{ftypes}{
 Cat
}

\examples{
gg_bar_coloured_parameter_hor_Cat(sampleData("Cat"))
}
