% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca.R
\name{gg_bar_coloured_parameter_hor_facet_CaCa.}
\alias{gg_bar_coloured_parameter_hor_facet_CaCa.}
\title{Horizontal bar facet highlighting some parameter
Facet Horizontal coloured by parameter Bars}
\usage{
gg_bar_coloured_parameter_hor_facet_CaCa.(data, titleLabel = "",
  subtitle = "", caption = "", xLabel = NULL, yLabel = "Count",
  parameter1 = NULL, parameter2 = NULL, leg_pos = "right", angle_x = 0,
  text = TRUE, type = "count", color_text = "black", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal bar facet highlighting some parameter
Facet Horizontal coloured by parameter Bars
}
\section{ftypes}{
 Ca-Ca
}

\examples{
add(1, 1)
add(10, 1)
}
