% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Nu-bars.R
\name{gg_bar_coloured_parameter_hor_facet_CatCatNum.}
\alias{gg_bar_coloured_parameter_hor_facet_CatCatNum.}
\title{Horizontal facet bar coloured highliting some parameter
Facet Horizontal coloured by parameter Bars}
\usage{
gg_bar_coloured_parameter_hor_facet_CatCatNum.(data, titleLabel = "",
  subtitle = "", caption = "", xLabel = NULL, yLabel = NULL,
  parameter1 = NULL, parameter2 = NULL, leg_pos = "right", angle_x = 0,
  aggregation = "sum", text = TRUE, type = "count",
  color_text = "black", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal facet bar coloured highliting some parameter
Facet Horizontal coloured by parameter Bars
}
\section{ftypes}{
 Cat-Cat-Num
}

\examples{
add(1, 1)
add(10, 1)
}
