% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca.R
\name{gg_bar_coloured_parameter_ver_Ca.}
\alias{gg_bar_coloured_parameter_ver_Ca.}
\title{Vertical bar highlighting some parameter
Vertical coloured by parameter bars}
\usage{
gg_bar_coloured_parameter_ver_Ca.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = "Count", parameter = NULL,
  text = TRUE, type = "count", color_text = "black", leg_pos = "right",
  ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical bar highlighting some parameter
Vertical coloured by parameter bars
}
\section{ftypes}{
 Ca
}

\examples{
add(1, 1)
add(10, 1)
}
