% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_bar_coloured_parameter_ver_CatNum.}
\alias{gg_bar_coloured_parameter_ver_CatNum.}
\title{Vertical bar highlighting some parameter
Vertical coloured by parameter bars}
\usage{
gg_bar_coloured_parameter_ver_CatNum.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, order = FALSE,
  parameter = NULL, line_mean = FALSE, text = TRUE, type = "count",
  color_text = "black", leg_pos = "right", aggregation = "sum",
  angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical bar highlighting some parameter
Vertical coloured by parameter bars
}
\section{ftypes}{
 Cat-Num
}
\examples{
add(1, 1)
add(10, 1)
}

