% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Nu-bars.R
\name{gg_bar_coloured_parameter_ver_facet_CaCaNu.}
\alias{gg_bar_coloured_parameter_ver_facet_CaCaNu.}
\title{Vertical facet bar coloured highliting some parameter
Facet Vertical coloured by parameter bars}
\usage{
gg_bar_coloured_parameter_ver_facet_CaCaNu.(data, titleLabel = "",
  subtitle = "", caption = "", xLabel = NULL, yLabel = NULL,
  parameter1 = NULL, parameter2 = NULL, leg_pos = "right", angle_x = 0,
  ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical facet bar coloured highliting some parameter
Facet Vertical coloured by parameter bars
}
\section{ftypes}{
 Ca-Ca-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
