% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possible_viz.R
\name{gg_bar_coloured_ver_Cat}
\alias{gg_bar_coloured_ver_Cat}
\title{Vertical coloured bar
Vertical coloured bars}
\usage{
gg_bar_coloured_ver_Cat(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, text = TRUE,
  type = "count", color_text = "black", leg_pos = "right", angle_x = 0,
  ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical coloured bar
Vertical coloured bars
}
\section{ftypes}{
 Cat
}

\examples{
gg_bar_coloured_ver_Cat(sampleData("Cat"))
}
