% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_bar_coloured_x_hor_CaNu.}
\alias{gg_bar_coloured_x_hor_CaNu.}
\title{gg_bar_coloured_x_hor_CaNu.
horizontal bar}
\usage{
gg_bar_coloured_x_hor_CaNu.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, line_mean = FALSE,
  text = TRUE, type = "percent", text_size = 3, leg_pos = "right",
  aggregation = "sum", angle = 0, ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_bar_coloured_x_hor_CaNu.
horizontal bar
}
\section{ftypes}{
 Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

