% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_bar_coloured_x_hor_CatNum.}
\alias{gg_bar_coloured_x_hor_CatNum.}
\title{Horizontal bar coloured by first variable
horizontal bar}
\usage{
gg_bar_coloured_x_hor_CatNum.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, line_mean = FALSE,
  text = TRUE, type = "count", leg_pos = "right", aggregation = "sum",
  angle_x = 0, color_text = "black", ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal bar coloured by first variable
horizontal bar
}
\section{ftypes}{
 Cat-Num
}

\examples{
add(1, 1)
add(10, 1)
}
