% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{gg_bar_coloured_x_hor_YeaNum}
\alias{gg_bar_coloured_x_hor_YeaNum}
\title{Horizontal bar coloured by first variable
Horizontal coloured bar}
\usage{
gg_bar_coloured_x_hor_YeaNum(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, text = TRUE, type = "percent",
  color_text = "black", yLabel = NULL, leg_pos = "right",
  aggregation = "sum", angle_x = 0, ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal bar coloured by first variable
Horizontal coloured bar
}
\section{ftypes}{
 Yea-Num
}

\examples{
gg_bar_coloured_x_hor_YeaNum(sampleData("Yea-Num"))
}
