% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ye-Nu.R
\name{gg_bar_coloured_x_hor_YeNu.}
\alias{gg_bar_coloured_x_hor_YeNu.}
\title{Horizontal bar coloured by first variable
Horizontal coloured bar}
\usage{
gg_bar_coloured_x_hor_YeNu.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, text = TRUE, type = "percent",
  color_text = "black", yLabel = NULL, fillLabel = NULL,
  leg_pos = "right", aggregation = "sum", angle_x = 0, ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal bar coloured by first variable
Horizontal coloured bar
}
\section{ftypes}{
 Ye-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
