% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ye-Nu.R
\name{gg_bar_coloured_x_ver_YeNu.}
\alias{gg_bar_coloured_x_ver_YeNu.}
\title{gg_bar_coloured_x_ver_YeNu.
vertical coloured bar}
\usage{
gg_bar_coloured_x_ver_YeNu.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, text = TRUE, type = "percent",
  size_text = 3, yLabel = NULL, fillLabel = NULL, leg_pos = "right",
  ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_bar_coloured_x_ver_YeNu.
vertical coloured bar
}
\section{ftypes}{
 Ca-Ye-Nu, Ye-Nu, Nu-Nu, Da-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

