% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_bar_coloured_y_hor_CatNum.}
\alias{gg_bar_coloured_y_hor_CatNum.}
\title{Horizontal bar density by first numeric variable
horizontal bar}
\usage{
gg_bar_coloured_y_hor_CatNum.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, fillLabel = NULL,
  reverse = FALSE, line_mean = FALSE, text = TRUE, type = "count",
  color_text = "black", leg_pos = "right", aggregation = "sum",
  angle_x = 0, ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal bar density by first numeric variable
horizontal bar
}
\section{ftypes}{
 Cat-Num
}
\examples{
add(1, 1)
add(10, 1)
}

