% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca.R
\name{gg_bar_coloured_y_hor_facet_CatCat.}
\alias{gg_bar_coloured_y_hor_facet_CatCat.}
\title{Horizontal bar facet coloured by second variable
Facet Horizontal coloured Bars}
\usage{
gg_bar_coloured_y_hor_facet_CatCat.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, fillLabel = NULL,
  leg_pos = "right", angle_x = 0, text = TRUE, type = "count",
  color_text = "black", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal bar facet coloured by second variable
Facet Horizontal coloured Bars
}
\section{ftypes}{
 Cat-Cat
}
\examples{
add(1, 1)
add(10, 1)
}

