% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possible_viz.R
\name{gg_bar_density_facet_z_hor_CatCatNum.}
\alias{gg_bar_density_facet_z_hor_CatCatNum.}
\title{Horizontal facet bar density by first numeric variable
Facet Coloured horizontal bar}
\arguments{
\item{x}{A category.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal facet bar density by first numeric variable
Facet Coloured horizontal bar
}
\section{ftypes}{
 Cat-Cat-Num
}

\examples{
gg_bar_density_facet_z_hor_CatCatNum.(sampleData("Cat-Cat-Num"))
}
