% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{gg_bar_density_facet_z_hor_CatCatNum}
\alias{gg_bar_density_facet_z_hor_CatCatNum}
\title{Horizontal facet bar density by first numeric variable
Facet Coloured horizontal bar}
\usage{
gg_bar_density_facet_z_hor_CatCatNum(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, fillLabel = NULL,
  reverse = FALSE, leg_pos = "right", angle_x = 0, aggregation = "sum",
  text = TRUE, color_text = "black", type = "count", ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal facet bar density by first numeric variable
Facet Coloured horizontal bar
}
\section{ftypes}{
 Cat-Cat-Num
}

\examples{
gg_bar_density_facet_z_hor_CatCatNum(sampleData("Cat-Cat-Num"))
}
