% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possible_viz.R
\name{gg_bar_density_y_ver_CatNum.}
\alias{gg_bar_density_y_ver_CatNum.}
\title{Vertical bar density by first numeric variable
vertical bar}
\usage{
gg_bar_density_y_ver_CatNum.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, fillLabel = NULL,
  reverse = FALSE, line_mean = FALSE, text = TRUE, type = "count",
  color_text = "black", leg_pos = "right", aggregation = "sum",
  angle_x = 0, ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical bar density by first numeric variable
vertical bar
}
\section{ftypes}{
 Cat-Num
}

\examples{
gg_bar_density_y_ver_CatNum.(sampleData("Cat-Num"))
}
