% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Da-Nu.R
\name{gg_bar_density_y_ver_DaNu.}
\alias{gg_bar_density_y_ver_DaNu.}
\title{Vertical bar density by first numeric variable
vertical coloured bar}
\usage{
gg_bar_density_y_ver_DaNu.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, text = TRUE, type = "count",
  color_text = "black", yLabel = NULL, fillLabel = NULL,
  leg_pos = "right", aggregation = "sum", angle_x = 0, reverse = FALSE,
  ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical bar density by first numeric variable
vertical coloured bar
}
\section{ftypes}{
 Da-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
