% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{gg_bar_dot_facet_hor_CatCat}
\alias{gg_bar_dot_facet_hor_CatCat}
\title{Horizontal dot bar facet
Facet Horizontal Dot Bar}
\usage{
gg_bar_dot_facet_hor_CatCat(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, leg_pos = "right",
  angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal dot bar facet
Facet Horizontal Dot Bar
}
\section{ftypes}{
 Cat-Cat
}

\examples{
gg_bar_dot_facet_hor_CatCat(sampleData("Cat-Cat"))
}
