% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possible_viz.R
\name{gg_bar_dot_facet_ver_CatCat.}
\alias{gg_bar_dot_facet_ver_CatCat.}
\title{Vertical dot bar facet
Facet Vertical Dot Bar}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical dot bar facet
Facet Vertical Dot Bar
}
\section{ftypes}{
 Cat-Cat
}

\examples{
gg_bar_dot_facet_ver_CatCat.(sampleData("Cat-Cat"))
}
