% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possible_viz.R
\name{gg_bar_dot_hor_Cat.}
\alias{gg_bar_dot_hor_Cat.}
\title{Horizontal dot bar
Horizontal Dot Bar}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal dot bar
Horizontal Dot Bar
}
\section{ftypes}{
 Cat
}

\examples{
gg_bar_dot_hor_Cat.(sampleData("Cat"))
}
