% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{gg_bar_dot_hor_Cat}
\alias{gg_bar_dot_hor_Cat}
\title{Horizontal dot bar
Horizontal Dot Bar}
\usage{
gg_bar_dot_hor_Cat(data, titleLabel = "", xLabel = NULL, yLabel = NULL,
  subtitle = "", caption = "", leg_pos = "right", angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal dot bar
Horizontal Dot Bar
}
\section{ftypes}{
 Cat
}

\examples{
gg_bar_dot_hor_Cat(sampleData("Cat"))
}
