% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{gg_bar_dot_ver_Cat}
\alias{gg_bar_dot_ver_Cat}
\title{Vertical dot bar
Vertical Dot Bar}
\usage{
gg_bar_dot_ver_Cat(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, leg_pos = "right", angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical dot bar
Vertical Dot Bar
}
\section{ftypes}{
 Cat
}

\examples{
gg_bar_dot_ver_Cat(sampleData("Cat"))
}
