% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possible_viz.R
\name{gg_bar_dot_ver_CatNum.}
\alias{gg_bar_dot_ver_CatNum.}
\title{Vertical dot bar
Vertical Dot Bar}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical dot bar
Vertical Dot Bar
}
\section{ftypes}{
 Cat-Num
}

\examples{
gg_bar_dot_ver_CatNum(sampleData("Cat-Num"))
}
