% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Nu.R
\name{gg_bar_facet_circular_CaCaNu.}
\alias{gg_bar_facet_circular_CaCaNu.}
\title{gg_bar_facet_circular_CaCaNu.
Circular Bar}
\usage{
gg_bar_facet_circular_CaCaNu.(data, titleLabel = "", subtitle = "",
  caption = "", leg_pos = "right", width = 0.85, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_bar_facet_circular_CaCaNu.
Circular Bar
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

