% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Nu.R
\name{gg_bar_facet_coloured_parameter_ver_hor_CaCaNu.}
\alias{gg_bar_facet_coloured_parameter_ver_hor_CaCaNu.}
\title{gg_bar_facet_coloured_parameter_ver_hor_CaCaNu.
Facet Horizontal coloured by parameter Bars}
\usage{
gg_bar_facet_coloured_parameter_ver_hor_CaCaNu.(data, titleLabel = "",
  xLabel = NULL, yLabel = NULL, parameter1 = NULL, parameter2 = NULL,
  leg_pos = "right", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_bar_facet_coloured_parameter_ver_hor_CaCaNu.
Facet Horizontal coloured by parameter Bars
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

