% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca.R
\name{gg_bar_facet_coloured_y_ver_CaCa.}
\alias{gg_bar_facet_coloured_y_ver_CaCa.}
\title{gg_bar_facet_coloured_y_ver_CaCa.
Facet Vertical coloured bars}
\usage{
gg_bar_facet_coloured_y_ver_CaCa.(data, titleLabel = "", xLabel = NULL,
  yLabel = "Count", leg_pos = "right", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_bar_facet_coloured_y_ver_CaCa.
Facet Vertical coloured bars
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

