% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Nu.R
\name{gg_bar_facet_coloured_y_ver_CaCaNu.}
\alias{gg_bar_facet_coloured_y_ver_CaCaNu.}
\title{gg_bar_facet_coloured_y_ver_CaCaNu.
vertical bar}
\usage{
gg_bar_facet_coloured_y_ver_CaCaNu.(data, titleLabel = "", xLabel = NULL,
  yLabel = NULL, leg_pos = "right", ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_bar_facet_coloured_y_ver_CaCaNu.
vertical bar
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

