% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possible_viz.R
\name{gg_bar_facet_hor_CatCatNum}
\alias{gg_bar_facet_hor_CatCatNum}
\title{Horizontal facet bar
Horizontal bar graph}
\usage{
gg_bar_facet_hor_CatCatNum(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, leg_pos = "right",
  angle_x = 0, aggregation = "sum", text = TRUE,
  color_text = "black", type = "count", ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A category.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal facet bar
Horizontal bar graph
}
\section{ftypes}{
 Cat-Cat-Num
}

\examples{
gg_bar_facet_hor_CatCatNum(sampleData("Cat-Cat-Num"))
}
