% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possible_viz.R
\name{gg_bar_facet_ver_CatCatNum}
\alias{gg_bar_facet_ver_CatCatNum}
\title{Vertical facet bar
Facet vertical bargraph}
\usage{
gg_bar_facet_ver_CatCatNum(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, leg_pos = "right",
  angle_x = 0, aggregation = "sum", text = TRUE, color_text = "black",
  type = "count", ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A category.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical facet bar
Facet vertical bargraph
}
\section{ftypes}{
 Cat-Cat-Num
}

\examples{
gg_bar_facet_ver_CatCatNum(sampleData("Cat-Cat-Num"))
}
