% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possible_viz.R
\name{gg_bar_facet_ver_CatYeaNum}
\alias{gg_bar_facet_ver_CatYeaNum}
\title{Vertical bar facet
Barras vertical}
\usage{
gg_bar_facet_ver_CatYeaNum(data, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical bar facet
Barras vertical
}
\section{ftypes}{
 Cat-Yea-Num
}

\examples{
gg_bar_facet_ver_CatYeaNum(sampleData("Cat-Yea-Num"))
}
