% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Nu-bars.R
\name{gg_bar_grouped2_hor_CaCaNu.}
\alias{gg_bar_grouped2_hor_CaCaNu.}
\title{Horizontal grouped bar by first variable
Barras grouped}
\usage{
gg_bar_grouped2_hor_CaCaNu.(data, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal grouped bar by first variable
Barras grouped
}
\section{ftypes}{
 Ca-Ca-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
