% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{gg_bar_grouped_CatCat}
\alias{gg_bar_grouped_CatCat}
\title{Grouped bar (categories, ordered categories)}
\usage{
gg_bar_grouped_CatCat(data, title = NULL, subtitle = NULL,
  caption = NULL, horLabel = NULL, verLabel = NULL, horLine = NULL,
  verLine = NULL, colors = c("#009EE3", "#F9B233"), dropNa = FALSE,
  format = c("", ""), labelRatio = 0.1, labelWrap = 12,
  marks = c(".", ","), nDigits = 2, order = NULL,
  orientation = "ver", percentage = FALSE, sort = "no",
  sliceN = NULL, showText = TRUE, theme = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Ggplot2 visualization
}
\description{
Compare quantities among two categories
}
\section{ctypes}{

Cat-Cat, Cat-Dat, Cat-Yea, Yea-Cat, Yea-Dat, Yea-Yea, Dat-Cat, Dat-Yea, Dat-Dat
}

\examples{
gg_bar_grouped_CatCat(sampleData("Cat-Cat", nrow = 10))
}
