% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{gg_bar_grouped_CatCatNum}
\alias{gg_bar_grouped_CatCatNum}
\title{Grouped bar (categories, ordered categories, numbers)}
\usage{
gg_bar_grouped_CatCatNum(data, title = NULL, subtitle = NULL,
  caption = NULL, horLabel = NULL, verLabel = NULL, horLine = NULL,
  verLine = NULL, agg = "sum", colors = c("#009EE3", "#F9B233"),
  colorText = "black", dropNa = c(FALSE, FALSE), format = c("", ""),
  labelRatio = 0.1, labelWrap = c(12, 12), leyendPosition = "right",
  marks = c(".", ","), nDigits = 2, order1 = NULL, order2 = NULL,
  orientation = "ver", percentage = FALSE, showText = TRUE,
  theme = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Ggplot2 visualization
}
\description{
Compare quantities among two categories
}
\section{ctypes}{

Cat-Cat-Num, Cat-Dat-Num, Cat-Yea-Num, Yea-Cat-Num, Yea-Dat-Num, Yea-Yea-Num, Dat-Cat-Num, Dat-Yea-Num, Dat-Dat-Num
}

\examples{
gg_bar_grouped_CatCatNum(sampleData("Cat-Cat-Num", nrow = 10))
}
