% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca.R
\name{gg_bar_grouped_coloured_ver_CaCa.}
\alias{gg_bar_grouped_coloured_ver_CaCa.}
\title{Vertical grouped bar by second variable
Unstacked Coloured vertical Bar}
\usage{
gg_bar_grouped_coloured_ver_CaCa.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, fillLabel = NULL,
  leg_pos = "right", text = TRUE, type = "count", color_text = "black",
  angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical grouped bar by second variable
Unstacked Coloured vertical Bar
}
\section{ftypes}{
 Ca-Ca
}

\examples{
add(1, 1)
add(10, 1)
}
