% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Nu-bars.R
\name{gg_bar_grouped_hor_CaCaNu.}
\alias{gg_bar_grouped_hor_CaCaNu.}
\title{Horizontal grouped bar by second variable
horizontal bar graph}
\usage{
gg_bar_grouped_hor_CaCaNu.(data, titleLabel = "", subtitle = "",
  caption = "", fillLabel = NULL, xLabel = NULL, yLabel = NULL,
  leg_pos = "right", aggregation = "sum", text = TRUE,
  color_text = "black", type = "count", angle_x = 0, ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A category.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal grouped bar by second variable
horizontal bar graph
}
\section{ftypes}{
 Ca-Ca-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
