% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possible_viz.R
\name{gg_bar_grouped_hor_CatCatNum.}
\alias{gg_bar_grouped_hor_CatCatNum.}
\title{Horizontal grouped bar by second variable
horizontal bar graph}
\usage{
gg_bar_grouped_hor_CatCatNum.(data, titleLabel = "", subtitle = "",
  caption = "", fillLabel = NULL, xLabel = NULL, yLabel = NULL,
  leg_pos = "right", aggregation = "sum", text = TRUE,
  color_text = "black", type = "count", angle_x = 0, ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A category.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal grouped bar by second variable
horizontal bar graph
}
\section{ftypes}{
 Cat-Cat-Num
}

\examples{
gg_bar_grouped_hor_CatCatNum.(sampleData("Cat-Cat-Num"))
}
