% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ye-Nu-bars.R
\name{gg_bar_grouped_hor_CatYeaNum.}
\alias{gg_bar_grouped_hor_CatYeaNum.}
\title{Horizontal grouped bar by first variable
Barras stacked}
\usage{
gg_bar_grouped_hor_CatYeaNum.(data, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal grouped bar by first variable
Barras stacked
}
\section{ftypes}{
 Cat-Yea-Num
}

\examples{
add(1, 1)
add(10, 1)
}
