% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ye-Nu.R
\name{gg_bar_grp_ver_CaYeNu.}
\alias{gg_bar_grp_ver_CaYeNu.}
\title{gg_bar_grp_ver_CaYeNu.: title.
Barras stacked
Tiene múltiples líneas}
\usage{
gg_bar_grp_ver_CaYeNu.(data, title = NULL, xlab = NULL, ylab = NULL,
  leg_pos = "right", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_bar_grp_ver_CaYeNu.: title.
Barras stacked
Tiene múltiples líneas
}
\section{ftypes}{
 Ca-Ye-Nu,Ca-Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

