% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca.R
\name{gg_bar_hor_Ca.}
\alias{gg_bar_hor_Ca.}
\title{gg_bar_hor_Ca.
Horizontal Bars}
\usage{
gg_bar_hor_Ca.(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, text = TRUE, type = "percent",
  text_size = 3, pos = 50, leg_pos = "right", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_bar_hor_Ca.
Horizontal Bars
}
\section{ftypes}{
 Ca
}
\examples{
add(1, 1)
add(10, 1)
}

