% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_bar_hor_CaNu.}
\alias{gg_bar_hor_CaNu.}
\title{gg_bar_hor_CaNu.
Horizontal Bars}
\usage{
gg_bar_hor_CaNu.(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, leg_pos = "right", text = TRUE,
  type = "count", text_size = 5, aggregation = "sum", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_bar_hor_CaNu.
Horizontal Bars
}
\section{ftypes}{
 Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

