% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{gg_bar_hor_CatNum.}
\alias{gg_bar_hor_CatNum.}
\title{Horizontal bar
Horizontal Bars}
\usage{
gg_bar_hor_CatNum.(data, titleLabel = "", subtitle = "", caption = "",
  topn = NULL, sort = "no", xLabel = NULL, yLabel = NULL,
  line_mean = FALSE, leg_pos = "right", text = FALSE, type = "count",
  color_text = "black", aggregation = "sum", angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal bar
Horizontal Bars
}
\section{ftypes}{
 Cat-Num
}

\examples{
add(1, 1)
add(10, 1)
}
