% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Da-Nu.R
\name{gg_bar_hor_DatNum.}
\alias{gg_bar_hor_DatNum.}
\title{Horizontal bar
bar}
\usage{
gg_bar_hor_DatNum.(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, angle_x = 0, ...)
}
\arguments{
\item{x}{A data.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal bar
bar
}
\section{ftypes}{
 Dat-Num
}
\examples{
add(1, 1)
add(10, 1)
}

