% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ye-Nu.R
\name{gg_bar_hor_YeNu.}
\alias{gg_bar_hor_YeNu.}
\title{gg_bar_hor_YeNu.
horizontal bar}
\usage{
gg_bar_hor_YeNu.(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = NULL, text = TRUE, type = "percent", size_text = 3,
  yLabel = NULL, leg_pos = "right", ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_bar_hor_YeNu.
horizontal bar
}
\section{ftypes}{
 Ye-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

