% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{gg_bar_ordered_hor_CatNum.}
\alias{gg_bar_ordered_hor_CatNum.}
\title{Ordered horizontal bar
Ordered horizontal Bars}
\usage{
gg_bar_ordered_hor_CatNum.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, leg_pos = "right",
  text = TRUE, type = "count", color_text = "black",
  aggregation = "sum", angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Ordered horizontal bar
Ordered horizontal Bars
}
\section{ftypes}{
 Cat-Num
}

\examples{
add(1, 1)
add(10, 1)
}
