% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca.R
\name{gg_bar_ordered_ver_Ca.}
\alias{gg_bar_ordered_ver_Ca.}
\title{gg_bar_ordered_ver_Ca.
Ordered vertical Bars}
\usage{
gg_bar_ordered_ver_Ca.(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, text = TRUE, type = "percent",
  size_text = 3, pos = 0.5, leg_pos = "right", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_bar_ordered_ver_Ca.
Ordered vertical Bars
}
\section{ftypes}{
 Ca
}
\examples{
add(1, 1)
add(10, 1)
}

