% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca.R
\name{gg_bar_ordered_ver_Cat.}
\alias{gg_bar_ordered_ver_Cat.}
\title{Ordered vertical bar
Ordered vertical Bars}
\usage{
gg_bar_ordered_ver_Cat.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, text = TRUE,
  type = "count", color_text = "black", leg_pos = "right", angle_x = 0,
  ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Ordered vertical bar
Ordered vertical Bars
}
\section{ftypes}{
 Cat
}
\examples{
add(1, 1)
add(10, 1)
}

