% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca.R
\name{gg_bar_polar_Ca.}
\alias{gg_bar_polar_Ca.}
\title{Polar bar
Polar Bar}
\usage{
gg_bar_polar_Ca.(data, width = 0.95, titleLabel = "", subtitle = "",
  caption = "", fillLabel = NULL, leg_pos = "right", text = TRUE,
  type = "count", color_text = "black", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Polar bar
Polar Bar
}
\section{ftypes}{
 Ca
}

\examples{
add(1, 1)
add(10, 1)
}
