% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_bar_polar_CaNu.}
\alias{gg_bar_polar_CaNu.}
\title{gg_bar_polar_CaNu.
Polar Bar}
\usage{
gg_bar_polar_CaNu.(data, width = 0.95, titleLabel = "", fillLabel = NULL,
  leg_pos = "right", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_bar_polar_CaNu.
Polar Bar
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

