% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_bar_polar_CatNum.}
\alias{gg_bar_polar_CatNum.}
\title{Polar bar
Polar Bar}
\usage{
gg_bar_polar_CatNum.(data, width = 0.95, titleLabel = "", subtitle = "",
  aggregation = "sum", caption = "", fillLabel = NULL,
  leg_pos = "right", text = TRUE, color_text = "black", type = "count",
  ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Polar bar
Polar Bar
}
\section{ftypes}{
 Cat-Num
}
\examples{
add(1, 1)
add(10, 1)
}

