% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca.R
\name{gg_bar_polar_stacked_100_CaCa.}
\alias{gg_bar_polar_stacked_100_CaCa.}
\title{gg_bar_polar_stacked_100_CaCa.
Stacked Polar  100pct}
\usage{
gg_bar_polar_stacked_100_CaCa.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, text = TRUE,
  type = "percent", size_text = 2, leg_pos = "right", width = 1,
  l_ncol = 1, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_bar_polar_stacked_100_CaCa.
Stacked Polar  100pct
}
\section{ftypes}{
 Ca,Ca-Nu, Ca-Ca
}
\examples{
add(1, 1)
add(10, 1)
}

