% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possible_viz.R
\name{gg_bar_polar_stacked_100_CatCat.}
\alias{gg_bar_polar_stacked_100_CatCat.}
\title{Polar 100% stacked bar
Stacked Polar  100pct}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Polar 100% stacked bar
Stacked Polar  100pct
}
\section{ftypes}{
 Cat-Cat
}

\examples{
add(1, 1)
add(10, 1)
}
