% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca.R
\name{gg_bar_polar_stacked_CaCa.}
\alias{gg_bar_polar_stacked_CaCa.}
\title{Polar stacked bar
Stacked Polar bar}
\usage{
gg_bar_polar_stacked_CaCa.(data, titleLabel = "", subtitle = "",
  caption = "", width = 0.95, fillLabel = NULL, text = TRUE,
  type = "count", color_text = "black", leg_pos = "right", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Polar stacked bar
Stacked Polar bar
}
\section{ftypes}{
 Ca-Ca
}

\examples{
add(1, 1)
add(10, 1)
}
