% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possible_viz.R
\name{gg_bar_polar_stacked_CatCat.}
\alias{gg_bar_polar_stacked_CatCat.}
\title{Polar stacked bar
Stacked Polar bar}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Polar stacked bar
Stacked Polar bar
}
\section{ftypes}{
 Cat-Cat
}

\examples{
add(1, 1)
add(10, 1)
}
