% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{gg_bar_single_stacked_hor_CatNum}
\alias{gg_bar_single_stacked_hor_CatNum}
\title{Horizontal stacked bar
Single Horizontal Stacked Bar}
\usage{
gg_bar_single_stacked_hor_CatNum(data, titleLabel = "", subtitle = "",
  caption = "", fillLabel = NULL, leg_pos = "right", width = 0.3,
  aggregation = "sum", text = TRUE, type = "count",
  color_text = "black", angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal stacked bar
Single Horizontal Stacked Bar
}
\section{ftypes}{
 Cat-Num
}

\examples{
gg_bar_single_stacked_hor_CatNum(sampleData("Cat-Num"))
}
