% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_bar_single_stacked_ver_CaNu.}
\alias{gg_bar_single_stacked_ver_CaNu.}
\title{Vertical stacked bar
Single Vertical Stacked Bar}
\usage{
gg_bar_single_stacked_ver_CaNu.(data, titleLabel = "", subtitle = "",
  caption = "", yLabel = "Count", leg_pos = "right", width = 0.3,
  aggregation = "sum", text = TRUE, type = "count",
  color_text = "black", angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical stacked bar
Single Vertical Stacked Bar
}
\section{ftypes}{
 Ca-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
