% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca.R
\name{gg_bar_single_stacked_ver_Cat.}
\alias{gg_bar_single_stacked_ver_Cat.}
\title{Vertical stacked bar
Single Vertical Stacked Bar}
\usage{
gg_bar_single_stacked_ver_Cat.(data, titleLabel = "", subtitle = "",
  caption = "", fillLabel = NULL, leg_pos = "right", width = 0.3,
  text = TRUE, type = "count", color_text = "black", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical stacked bar
Single Vertical Stacked Bar
}
\section{ftypes}{
 Cat
}
\examples{
add(1, 1)
add(10, 1)
}

