% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Nu-bars.R
\name{gg_bar_stacked2_100_hor_CaCaNu.}
\alias{gg_bar_stacked2_100_hor_CaCaNu.}
\title{gg_bar_stacked2_100_hor_CaCaNu.: title.
Barras grouped
Tiene múltiples líneas}
\usage{
gg_bar_stacked2_100_hor_CaCaNu.(data, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_bar_stacked2_100_hor_CaCaNu.: title.
Barras grouped
Tiene múltiples líneas
}
\section{ftypes}{
 Ca-Ye-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
