% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possible_viz.R
\name{gg_bar_stacked2_100_ver_CatCatNum.}
\alias{gg_bar_stacked2_100_ver_CatCatNum.}
\title{Vertical 100% stacked bar by first variable
Barras grouped}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical 100% stacked bar by first variable
Barras grouped
}
\section{ftypes}{
 Cat-Cat-Num
}

\examples{
gg_bar_stacked2_100_ver_CatCatNum.(sampleData("Cat-Cat-Num"))
}
