% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Nu-bars.R
\name{gg_bar_stacked2_hor_CatCatNum.}
\alias{gg_bar_stacked2_hor_CatCatNum.}
\title{Horizontal stacked bar by first variable
Barras grouped
Tiene múltiples líneas}
\usage{
gg_bar_stacked2_hor_CatCatNum.(data, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal stacked bar by first variable
Barras grouped
Tiene múltiples líneas
}
\section{ftypes}{
 Cat-Cat-Num
}
\examples{
add(1, 1)
add(10, 1)
}

