% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{gg_bar_stacked_100_CatCat}
\alias{gg_bar_stacked_100_CatCat}
\title{100% Stacked bar (categories, ordered categories)}
\usage{
gg_bar_stacked_100_CatCat(data, title = NULL, subtitle = NULL,
  caption = NULL, horLabel = NULL, verLabel = NULL, horLine = NULL,
  verLine = NULL, colors = c("#009EE3", "#F9B233"),
  colorText = "black", dropNa = c(FALSE, FALSE), format = c("", ""),
  labelRatio = 0.1, labelWrap = c(12, 12), leyendLayout = "right",
  marks = c(".", ","), nDigits = 2, order1 = NULL, order2 = NULL,
  orientation = "ver", percentage = FALSE, showText = TRUE,
  theme = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Ggplot2 visualization
}
\description{
Compare quantities among two categories
}
\section{ctypes}{

Cat-Cat, Cat-Dat, Cat-Yea, Yea-Cat, Yea-Dat, Yea-Yea, Dat-Cat, Dat-Yea, Dat-Dat
}

\examples{
gg_bar_stacked_100_CatCat(sampleData("Cat-Cat", nrow = 10))
}
