% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Ye-Nu.R
\name{gg_bar_stacked_100_hor_CaCaYeNu.}
\alias{gg_bar_stacked_100_hor_CaCaYeNu.}
\title{gg_bar_stacked_100_hor_CaCaYeNu.
Stacked}
\usage{
gg_bar_stacked_100_hor_CaCaYeNu.(data, titleLabel = "", subtitle = "",
  caption = "", fillLabel = NULL, xlab = NULL, ylab = NULL,
  leg_pos = "right", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_bar_stacked_100_hor_CaCaYeNu.
Stacked
}
\section{ftypes}{
 Ca-Ca-Ye-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
