% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possible_viz.R
\name{gg_bar_stacked_100_hor_CatCatNum.}
\alias{gg_bar_stacked_100_hor_CatCatNum.}
\title{Horizontal 100% stacked bar by second variable
100 horizontal stacked bar graph}
\arguments{
\item{x}{A category.}

\item{y}{A category.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal 100% stacked bar by second variable
100 horizontal stacked bar graph
}
\section{ftypes}{
 Cat-Cat-Num
}

\examples{
gg_bar_stacked_100_hor_CatCatNum.(sampleData("Cat-Cat-Num"))
}
