% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possible_viz.R
\name{gg_bar_stacked_100_hor_CatYeaNum.}
\alias{gg_bar_stacked_100_hor_CatYeaNum.}
\title{Horizontal 100% stacked bar by first variable
Barras horizontal stacked}
\usage{
gg_bar_stacked_100_hor_CatYeaNum.(data, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal 100% stacked bar by first variable
Barras horizontal stacked
}
\section{ftypes}{
 Cat-Yea-Num
}

\examples{
gg_bar_stacked_100_hor_CatYeaNum.(sampleData("Cat-Yea-Num"))
}
