% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca.R
\name{gg_bar_stacked_100_ver_CaCa.}
\alias{gg_bar_stacked_100_ver_CaCa.}
\title{Vertical 100% stacked bar
Stacked 100pct vertical Bar}
\usage{
gg_bar_stacked_100_ver_CaCa.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, leg_pos = "right",
  text = TRUE, type = "count", color_text = "black", angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical 100% stacked bar
Stacked 100pct vertical Bar
}
\section{ftypes}{
 Ca-Ca
}

\examples{
add(1, 1)
add(10, 1)
}
