% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Nu-bars.R
\name{gg_bar_stacked_100_ver_CaCaNu.}
\alias{gg_bar_stacked_100_ver_CaCaNu.}
\title{Vertical 100% stacked bar by first variable
100 vertical stacked bar graph}
\usage{
gg_bar_stacked_100_ver_CaCaNu.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, leg_pos = "right",
  aggregation = "mean", ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A category.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical 100% stacked bar by first variable
100 vertical stacked bar graph
}
\section{ftypes}{
 Ca-Ca-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
