% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Ca-Nu.R
\name{gg_bar_stacked_100_ver_CatCatCatNum.}
\alias{gg_bar_stacked_100_ver_CatCatCatNum.}
\title{Vertical 100% stacked bar by first variable + facet by third variable
Stacked}
\usage{
gg_bar_stacked_100_ver_CatCatCatNum.(data, titleLabel = "", subtitle = "",
  caption = "", fillLabel = NULL, xLabel = NULL, yLabel = NULL,
  leg_pos = "right", aggregation = "sum", text = TRUE, type = "count",
  color_text = "black", angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical 100% stacked bar by first variable + facet by third variable
Stacked
}
\section{ftypes}{
 Cat-Cat-Cat-Num
}
\examples{
add(1, 1)
add(10, 1)
}

