% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{gg_bar_stacked_100_ver_CatCatNum}
\alias{gg_bar_stacked_100_ver_CatCatNum}
\title{Vertical 100% stacked bar by second variable
100 vertical stacked bar graph}
\usage{
gg_bar_stacked_100_ver_CatCatNum(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, fillLabel = NULL,
  leg_pos = "right", aggregation = "sum", text = TRUE,
  color_text = "black", type = "count", angle_x = 0, ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A category.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical 100% stacked bar by second variable
100 vertical stacked bar graph
}
\section{ftypes}{
 Cat-Cat-Num
}

\examples{
gg_bar_stacked_100_ver_CatCatNum(sampleData("Cat-Cat-Num"))
}
