% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possible_viz.R
\name{gg_bar_stacked_100_ver_CatCatYeaNum}
\alias{gg_bar_stacked_100_ver_CatCatYeaNum}
\title{Vertical 100% stacked bar by first variable + facet by second variable
Stacked}
\usage{
gg_bar_stacked_100_ver_CatCatYeaNum(data, titleLabel = "",
  subtitle = "", caption = "", fillLabel = NULL, xLabel = NULL,
  yLabel = NULL, leg_pos = "right", aggregation = "sum",
  text = TRUE, type = "count", color_text = "black", angle_x = 0,
  ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical 100% stacked bar by first variable + facet by second variable
Stacked
}
\section{ftypes}{
 Cat-Cat-Yea-Num
}

\examples{
gg_bar_stacked_100_ver_CatCatYeaNum(sampleData("Cat-Cat-Yea-Num"))
}
