% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{gg_bar_stacked_100_ver_CatYeaNum}
\alias{gg_bar_stacked_100_ver_CatYeaNum}
\title{Vertical 100% stacked bar by first variable
100 vertical stacked bar graph}
\usage{
gg_bar_stacked_100_ver_CatYeaNum(data, ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A category.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical 100% stacked bar by first variable
100 vertical stacked bar graph
}
\section{ftypes}{
 Cat-Yea-Num
}

\examples{
gg_bar_stacked_100_ver_CatYeaNum(sampleData("Cat-Yea-Num"))
}
