% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possible_viz.R
\name{gg_bar_stacked_dot_hor_CatCat.}
\alias{gg_bar_stacked_dot_hor_CatCat.}
\title{Horizontal stacked dot bar
Stacked horizontal dot Bar}
\usage{
gg_bar_stacked_dot_hor_CatCat.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, fillLabel = NULL,
  leg_pos = "right", angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal stacked dot bar
Stacked horizontal dot Bar
}
\section{ftypes}{
 Cat-Cat
}

\examples{
gg_bar_stacked_dot_hor_CatCat.(sampleData("Cat-Cat"))
}
