% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{gg_bar_stacked_dot_ver_CatCat}
\alias{gg_bar_stacked_dot_ver_CatCat}
\title{Vertical stacked dot bar
Stacked Vertical dot Bar}
\usage{
gg_bar_stacked_dot_ver_CatCat(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, fillLabel = NULL,
  leg_pos = "right", angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical stacked dot bar
Stacked Vertical dot Bar
}
\section{ftypes}{
 Cat-Cat
}

\examples{
gg_bar_stacked_dot_ver_CatCat(sampleData("Cat-Cat"))
}
