% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Da-Nu.R
\name{gg_bar_stacked_hor_CatDatNum.}
\alias{gg_bar_stacked_hor_CatDatNum.}
\title{Horizontal stacked bar
horizontal stacked bar graph}
\usage{
gg_bar_stacked_hor_CatDatNum.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, fillLabel = NULL,
  leg_pos = "right", angle_x = 0, aggregation = "sum", hline = NULL,
  text = TRUE, type = "count", color_text = "black", ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A category.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal stacked bar
horizontal stacked bar graph
}
\section{ftypes}{
 Cat-Dat-Num
}
\examples{
add(1, 1)
add(10, 1)
}

