% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possible_viz.R
\name{gg_bar_stacked_ordered_ver_CatCat}
\alias{gg_bar_stacked_ordered_ver_CatCat}
\title{Ordered vertical stacked bar by second variable
Ordered Stacked vertical Bar}
\usage{
gg_bar_stacked_ordered_ver_CatCat(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, fillLabel = NULL,
  leg_pos = "right", text = TRUE, type = "count",
  color_text = "black", angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Ordered vertical stacked bar by second variable
Ordered Stacked vertical Bar
}
\section{ftypes}{
 Cat-Cat
}

\examples{
gg_bar_stacked_ordered_ver_CatCat(sampleData("Cat-Cat"))
}
