% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possible_viz.R
\name{gg_bar_stacked_polar_100_CatCatNum.}
\alias{gg_bar_stacked_polar_100_CatCatNum.}
\title{Stacked 100% polar bar
Stacked Polar Bar 100}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Stacked 100% polar bar
Stacked Polar Bar 100
}
\section{ftypes}{
 Cat-Cat-Num
}

\examples{
add(1, 1)
add(10, 1)
}
