% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Nu.R
\name{gg_bar_stacked_polar_CaCaNu.}
\alias{gg_bar_stacked_polar_CaCaNu.}
\title{gg_bar_stacked_polar_CaCaNu.
Stacked Polar Bar}
\usage{
gg_bar_stacked_polar_CaCaNu.(data, width = 0.95, titleLabel = "",
  leg_pos = "right", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_bar_stacked_polar_CaCaNu.
Stacked Polar Bar
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

