% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possible_viz.R
\name{gg_bar_stacked_polar_CatCatNum.}
\alias{gg_bar_stacked_polar_CatCatNum.}
\title{Stacked polar bar
Stacked Polar Bar}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Stacked polar bar
Stacked Polar Bar
}
\section{ftypes}{
 Cat-Cat-Num
}

\examples{
add(1, 1)
add(10, 1)
}
