% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Da-Nu.R
\name{gg_bar_stacked_ver_CaDaNu.}
\alias{gg_bar_stacked_ver_CaDaNu.}
\title{Vertical stacked bar
vertical stacked bar graph}
\usage{
gg_bar_stacked_ver_CaDaNu.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, leg_pos = "right",
  angle_x = 0, aggregation = "sum", hline = NULL, text = TRUE,
  type = "count", color_text = "black", ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A category.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical stacked bar
vertical stacked bar graph
}
\section{ftypes}{
 Ca-Da-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
