% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Nu-bars.R
\name{gg_bar_stacked_ver_CatCatNum.}
\alias{gg_bar_stacked_ver_CatCatNum.}
\title{Vertical stacked bar by second variable
vertical stacked bar graph}
\usage{
gg_bar_stacked_ver_CatCatNum.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, fillLabel = NULL,
  leg_pos = "right", aggregation = "sum", text = TRUE, type = "count",
  color_text = "black", angle_x = 0, ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A category.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical stacked bar by second variable
vertical stacked bar graph
}
\section{ftypes}{
 Cat-Cat-Num
}

\examples{
add(1, 1)
add(10, 1)
}
