% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{gg_bar_stacked_ver_CatCatYeaNum.}
\alias{gg_bar_stacked_ver_CatCatYeaNum.}
\title{Vertical stacked bar by first variable + facet by second variable
Stacked}
\usage{
gg_bar_stacked_ver_CatCatYeaNum.(data, titleLabel = "", subtitle = "",
  caption = "", fillLabel = NULL, xLabel = NULL, yLabel = NULL,
  leg_pos = "right", aggregation = "sum", text = TRUE, type = "count",
  color_text = "black", angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical stacked bar by first variable + facet by second variable
Stacked
}
\section{ftypes}{
 Cat-Cat-Yea-Num
}

\examples{
add(1, 1)
add(10, 1)
}
