% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_bar_ver_CatNum.}
\alias{gg_bar_ver_CatNum.}
\title{Vertical bar
Vertical Bars}
\usage{
gg_bar_ver_CatNum.(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, topn = NULL, sort = "no",
  line_mean = FALSE, leg_pos = "right", text = FALSE, type = "number",
  color_text = "black", aggregation = "sum", angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical bar
Vertical Bars
}
\section{ftypes}{
 Cat-Num
}
\examples{
add(1, 1)
add(10, 1)
}

