% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{gg_bar_ver_YeaNum}
\alias{gg_bar_ver_YeaNum}
\title{Vertical bar
vertical bar}
\usage{
gg_bar_ver_YeaNum(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = NULL, text = TRUE, type = "count", color_text = "black",
  yLabel = NULL, leg_pos = "right", angle_x = 0, aggregation = "sum",
  ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical bar
vertical bar
}
\section{ftypes}{
 Yea-Num
}

\examples{
gg_bar_ver_YeaNum(sampleData("Yea-Num"))
}
