% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{gg_bar_YeaNum}
\alias{gg_bar_YeaNum}
\title{Bar (years, numbers)}
\usage{
gg_bar_YeaNum(data, title = NULL, subtitle = NULL, caption = NULL,
  horLabel = NULL, verLabel = NULL, horLine = NULL, verLine = NULL,
  agg = "sum", colors = NULL, colorText = "black",
  colorScale = "no", dropNa = FALSE, format = c("", ""),
  highlightValue = NULL, highlightValueColor = NULL,
  labelRatio = 0.1, labelWrap = 12, marks = c(".", ","),
  nDigits = 0, order = NULL, orientation = "ver",
  percentage = FALSE, sort = "no", sliceN = NULL, showText = TRUE,
  theme = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Ggplot2 visualization
}
\description{
Compare quantities over years
}
\section{ctypes}{

Yea-Num
}

\examples{
gg_bar_YeaNum(sampleData("Yea-Num", nrow = 10))
}
