% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca.R
\name{gg_bars_ver.}
\alias{gg_bars_ver.}
\title{gg_bars_ver.}
\usage{
gg_bars_ver.(data, titleLabel = "Report", xLabel = "Frequency",
  yLabel = "Categories", fillLabel = "Types", leg_pos = "right")
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Hola esta es la descripción
Tiene múltiples líneas
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

