% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nu.R
\name{gg_box_flip_Nu.}
\alias{gg_box_flip_Nu.}
\title{Horizontal boxplot
Box plot flipped}
\usage{
gg_box_flip_Nu.(data, titleLabel = "", subtitle = "", caption = "",
  yLabel = NULL, xLabel = NULL, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal boxplot
Box plot flipped
}
\section{ftypes}{
 Nu
}

\examples{
add(1, 1)
add(10, 1)
}
