% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nu.R
\name{gg_box_Nu.}
\alias{gg_box_Nu.}
\title{Vertical boxplot
Box plot}
\usage{
gg_box_Nu.(data, title = "", subtitle = "", caption = "", yLabel = NULL,
  xLabel = NULL, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical boxplot
Box plot
}
\section{ftypes}{
 Nu
}

\examples{
add(1, 1)
add(10, 1)
}
