% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_boxplot_dot_CaNu.}
\alias{gg_boxplot_dot_CaNu.}
\title{Vertical boxplot + scatter
Boxplot + dot jitter}
\usage{
gg_boxplot_dot_CaNu.(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, leg_pos = "right", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical boxplot + scatter
Boxplot + dot jitter
}
\section{ftypes}{
 Ca-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
