% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{gg_boxplot_dot_CatNum.}
\alias{gg_boxplot_dot_CatNum.}
\title{Vertical boxplot + jitter
Boxplot + dot jitter}
\usage{
gg_boxplot_dot_CatNum.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, leg_pos = "right",
  angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical boxplot + jitter
Boxplot + dot jitter
}
\section{ftypes}{
 Cat-Num
}

\examples{
add(1, 1)
add(10, 1)
}
