% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_boxplot_flip_CaNu.}
\alias{gg_boxplot_flip_CaNu.}
\title{Vertical boxplot
Boxplot flipped}
\usage{
gg_boxplot_flip_CaNu.(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, leg_pos = "right", angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical boxplot
Boxplot flipped
}
\section{ftypes}{
 Ca-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
