% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca.R
\name{gg_bubble_Ca2.}
\alias{gg_bubble_Ca2.}
\title{gg_bubble_Ca2.
bubbles}
\usage{
gg_bubble_Ca2.(data, titleLabel = "", subtitle = "", caption = "",
  sep = 3, lim_inf = -40, lim_sup = 40, fillLabel = NULL, ...)
}
\arguments{
\item{x}{A category.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_bubble_Ca2.
bubbles
}
\section{ftypes}{
 Ca
}

\examples{
add(1, 1)
add(10, 1)
}
