% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca.R
\name{gg_bubble_CaCa.}
\alias{gg_bubble_CaCa.}
\title{Bubble
Bubble}
\usage{
gg_bubble_CaCa.(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, text = TRUE, color_text = "black",
  type = "count", angle_x = 0, shape_type = 19, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Bubble
Bubble
}
\section{ftypes}{
 Ca-Ca
}

\examples{
add(1, 1)
add(10, 1)
}
