% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Nu.R
\name{gg_bubble_CaCaNu.}
\alias{gg_bubble_CaCaNu.}
\title{gg_bubble_CaCaNu.
Bubble}
\usage{
gg_bubble_CaCaNu.(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_bubble_CaCaNu.
Bubble
}
\section{ftypes}{
 Ca-Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

