% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_bubble_CaNu.}
\alias{gg_bubble_CaNu.}
\title{Bubble
Bubble}
\usage{
gg_bubble_CaNu.(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = NULL, text = TRUE, type = "count", color_text = "black",
  leg_pos = "right", aggregation = "sum", angle_x = 0, shape_type = 19,
  ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Bubble
Bubble
}
\section{ftypes}{
 Ca-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
