% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R, R/Ca.R
\name{gg_bubble_CaNu2.}
\alias{gg_bubble_CaNu2.}
\title{gg_bubble_CaNu2.
Bubbles}
\usage{
gg_bubble_CaNu2.(data, titleLabel = "", sep = 3, lim_inf = -40,
  lim_sup = 40, fillLabel = NULL, ...)

gg_bubble_CaNu2.(data, titleLabel = "", sep = 3, lim_inf = -40,
  lim_sup = 40, fillLabel = NULL, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}

\item{x}{A category.}
}
\value{
The sum of \code{x} and \code{y}.

The sum of \code{x} and \code{y}.
}
\description{
gg_bubble_CaNu2.
Bubbles

gg_bubble_CaNu2.
bubbles
}
\section{ftypes}{
 Ca,Ca-Nu

 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
add(1, 1)
add(10, 1)
}

