% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca.R
\name{gg_bubble_Cat.}
\alias{gg_bubble_Cat.}
\title{Bubble
Bubble}
\usage{
gg_bubble_Cat.(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = NULL, shape_type = 19, angle_x = 0, text = TRUE,
  type = "count", color_text = "black", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Bubble
Bubble
}
\section{ftypes}{
 Cat
}
\examples{
add(1, 1)
add(10, 1)
}

