% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ge-Nu.R
\name{gg_bubble_co_CaGe.}
\alias{gg_bubble_co_CaGe.}
\title{gg_bubble_co_CaGe.
Points inside Colombia's deptos map}
\usage{
gg_bubble_co_CaGe.(data, titleLabel = "", fillLabel = NULL,
  leg_pos = "right", color_map = "gray", color_frontier = "white",
  scale_point = 0.25, alpha = 0.5, ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_bubble_co_CaGe.
Points inside Colombia's deptos map
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

