% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ge-Nu.R
\name{gg_bubble_co_CaGnm.}
\alias{gg_bubble_co_CaGnm.}
\title{Points inside Colombia's deptos map
Points inside Colombia's deptos map}
\usage{
gg_bubble_co_CaGnm.(data, titleLabel = "", subtitle = "", caption = "",
  fillLabel = NULL, leg_pos = "right", color_map = "gray",
  color_frontier = "white", scale_point = 0.25, alpha = 0.5, ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Points inside Colombia's deptos map
Points inside Colombia's deptos map
}
\section{ftypes}{
 Ca-Gnm
}

\examples{
add(1, 1)
add(10, 1)
}
