% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ge-Nu.R
\name{gg_bubble_co_Ge.}
\alias{gg_bubble_co_Ge.}
\title{gg_bubble_co_Ge.
Points inside Colombia's deptos map}
\usage{
gg_bubble_co_Ge.(data, titleLabel = "", subtitle = "", caption = "",
  fillLabel = NULL, prop_text = 0.5, color_point = "red",
  leg_pos = "right", text = FALSE, text_size = 2, color_map = "gray",
  color_frontier = "white", scale_point = 0.25, alpha = 0.5, ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_bubble_co_Ge.
Points inside Colombia's deptos map
}
\section{ftypes}{
 Ca-Ge, Ge, Ge-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

