% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca.R
\name{gg_bubble_coloured_Ca.}
\alias{gg_bubble_coloured_Ca.}
\title{Bubble coloured
Coloured Bubble}
\usage{
gg_bubble_coloured_Ca.(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = NULL, fillLabel = NULL, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Bubble coloured
Coloured Bubble
}
\section{ftypes}{
 Ca
}

\examples{
add(1, 1)
add(10, 1)
}
