% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca.R
\name{gg_bubble_coloured_CaCa.}
\alias{gg_bubble_coloured_CaCa.}
\title{gg_bubble_coloured_CaCa.
Coloured Bubble}
\usage{
gg_bubble_coloured_CaCa.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_bubble_coloured_CaCa.
Coloured Bubble
}
\section{ftypes}{
 Ca-Ca
}
\examples{
add(1, 1)
add(10, 1)
}

