% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_bubble_coloured_CaNu.}
\alias{gg_bubble_coloured_CaNu.}
\title{gg_bubble_coloured_CaNu.
Coloured Bubble}
\usage{
gg_bubble_coloured_CaNu.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_bubble_coloured_CaNu.
Coloured Bubble
}
\section{ftypes}{
 Ca-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
