% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Nu.R
\name{gg_bubble_coloured_x_CaCaNu.}
\alias{gg_bubble_coloured_x_CaCaNu.}
\title{Bubble coloured by first variable
Coloured Bubble first Ca}
\usage{
gg_bubble_coloured_x_CaCaNu.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, aggregation = "sum",
  angle_x = 0, text = TRUE, color_text = "black", type = "count", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Bubble coloured by first variable
Coloured Bubble first Ca
}
\section{ftypes}{
 Ca-Ca-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
