% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble.R
\name{gg_bubble_coloured_x_CatCat.}
\alias{gg_bubble_coloured_x_CatCat.}
\title{Bubble coloured by first variable
Coloured Bubble}
\usage{
gg_bubble_coloured_x_CatCat.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, text = TRUE,
  color_text = "black", type = "count", angle_x = 0, shape_type = 19,
  ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Bubble coloured by first variable
Coloured Bubble
}
\section{ftypes}{
 Cat-Cat
}

\examples{
add(1, 1)
add(10, 1)
}
