% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Nu.R
\name{gg_bubble_coloured_y_CatCatNum.}
\alias{gg_bubble_coloured_y_CatCatNum.}
\title{Bubble coloured by second variable
Coloured Bubble second Cat}
\usage{
gg_bubble_coloured_y_CatCatNum.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, aggregation = "sum",
  angle_x = 0, text = TRUE, color_text = "black", type = "count",
  shape_type = 19, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Bubble coloured by second variable
Coloured Bubble second Cat
}
\section{ftypes}{
 Cat-Cat-Num
}
\examples{
add(1, 1)
add(10, 1)
}

