% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble.R
\name{gg_bubble_density_y_CatNum.}
\alias{gg_bubble_density_y_CatNum.}
\title{Bubble coloured by first numeric variable
Coloured Bubble}
\usage{
gg_bubble_density_y_CatNum.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, fillLabel = NULL, text = TRUE,
  type = "count", color_text = "black", leg_pos = "right",
  aggregation = "sum", angle_x = 0, shape_type = 19, reverse = FALSE,
  ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Bubble coloured by first numeric variable
Coloured Bubble
}
\section{ftypes}{
 Cat-Num
}

\examples{
add(1, 1)
add(10, 1)
}
