% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ge-Nu.R
\name{gg_bubble_depto_CaGe.}
\alias{gg_bubble_depto_CaGe.}
\title{gg_bubble_depto_CaGe.
Points inside Colombia's mpios map}
\usage{
gg_bubble_depto_CaGe.(data, titleLabel = "Report", depto_ = "05",
  fillLabel = NULL, color_point = "red", color_map = "gray",
  color_frontier = "white", scale_point = 0.25)
}
\arguments{
\item{x}{A category.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_bubble_depto_CaGe.
Points inside Colombia's mpios map
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

