% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ge-Nu.R
\name{gg_bubble_latam_Ge.}
\alias{gg_bubble_latam_Ge.}
\title{gg_bubble_latam_Ge.
Points inside Latam map}
\usage{
gg_bubble_latam_Ge.(data, titleLabel = "", fillLabel = NULL,
  color_point = "red", leg_pos = "right", color_map = "gray",
  color_frontier = "white", scale_point = 0.25, alpha = 0.5, ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_bubble_latam_Ge.
Points inside Latam map
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

