% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ge-Nu.R
\name{gg_bubble_latam_GeNu.}
\alias{gg_bubble_latam_GeNu.}
\title{gg_bubble_latam_GeNu.
Points inside Latam map}
\usage{
gg_bubble_latam_GeNu.(data, titleLabel = "Report", fillLabel = NULL,
  color_point = "red", color_map = "gray", color_frontier = "white",
  scale_point = 0.25)
}
\arguments{
\item{x}{A category.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_bubble_latam_GeNu.
Points inside Latam map
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

