% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Da-Nu.R
\name{gg_bubbles_DaNu.}
\alias{gg_bubbles_DaNu.}
\title{Bubbles
bubbles}
\usage{
gg_bubbles_DaNu.(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, angle_x = 0, shape_type = 19, ...)
}
\arguments{
\item{x}{A data.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Bubbles
bubbles
}
\section{ftypes}{
 Da-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
