% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Da-Nu.R
\name{gg_bubbles_density_x_DatNum.}
\alias{gg_bubbles_density_x_DatNum.}
\title{Bubbles density by first numeric variable
bubbles}
\usage{
gg_bubbles_density_x_DatNum.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, angle_x = 0,
  shape_type = 19, leg_pos = "right", reverse = FALSE, ...)
}
\arguments{
\item{x}{A data.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Bubbles density by first numeric variable
bubbles
}
\section{ftypes}{
 Dat-Num
}
\examples{
add(1, 1)
add(10, 1)
}

