% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca.R
\name{gg_bullseye_Ca.}
\alias{gg_bullseye_Ca.}
\title{Bullseye
Bullseye}
\usage{
gg_bullseye_Ca.(data, titleLabel = "", subtitle = "", caption = "",
  fillLabel = NULL, leg_pos = "right", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Bullseye
Bullseye
}
\section{ftypes}{
 Ca
}

\examples{
add(1, 1)
add(10, 1)
}
