% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Nu.R
\name{gg_bullseye_CaCaNu.}
\alias{gg_bullseye_CaCaNu.}
\title{gg_bullseye_CaCaNu.
Facet Bullseye}
\usage{
gg_bullseye_CaCaNu.(data, titleLabel = "Report", fillLabel = NULL,
  leg_pos = "right")
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_bullseye_CaCaNu.
Facet Bullseye
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

