% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_bullseye_CaNu.}
\alias{gg_bullseye_CaNu.}
\title{gg_bullseye_CaNu.
Bullseye}
\usage{
gg_bullseye_CaNu.(data, titleLabel = "", subtitle = "", caption = "",
  leg_pos = "right", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_bullseye_CaNu.
Bullseye
}
\section{ftypes}{
 Ca, Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

