% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Nu.R
\name{gg_bullseye_facet_CatCatNum.}
\alias{gg_bullseye_facet_CatCatNum.}
\title{Bullseye facet
Facet Bullseye}
\usage{
gg_bullseye_facet_CatCatNum.(data, titleLabel = "", subtitle = "",
  caption = "", fillLabel = NULL, leg_pos = "right", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Bullseye facet
Facet Bullseye
}
\section{ftypes}{
 Cat-Cat-Num
}
\examples{
add(1, 1)
add(10, 1)
}

