% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ge-Nu.R
\name{gg_choropleth_co_GeNu.}
\alias{gg_choropleth_co_GeNu.}
\title{gg_choropleth_co_GeNu.
Choropleth of Colombia's deptos}
\usage{
gg_choropleth_co_GeNu.(data, titleLabel = "", subtitle = "", caption = "",
  reverse = FALSE, text_size = 2, fillLabel = NULL, leg_pos = "right",
  text = FALSE, prop_text = 0.5, color_map = "gray",
  color_frontier = "white", ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_choropleth_co_GeNu.
Choropleth of Colombia's deptos
}
\section{ftypes}{
 Ge-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

