% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ge-Nu.R
\name{gg_choropleth_depto_GeNu.}
\alias{gg_choropleth_depto_GeNu.}
\title{gg_choropleth_depto_GeNu.
Choropleth by filtering deptos}
\usage{
gg_choropleth_depto_GeNu.(data, titleLabel = "Report", depto_ = "05",
  fillLabel = NULL, leg_pos = "right", color_map = "gray",
  color_frontier = "white", color_low = "lightblue",
  color_high = "darkblue")
}
\arguments{
\item{x}{A category.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_choropleth_depto_GeNu.
Choropleth by filtering deptos
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

