% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ge-Nu.R
\name{gg_choropleth_latam_GeNu.}
\alias{gg_choropleth_latam_GeNu.}
\title{Choropleth of Latam
Choropleth of Latam}
\usage{
gg_choropleth_latam_GeNu.(data, titleLabel = "", subtitle = "",
  caption = "", reverse = FALSE, fillLabel = NULL, leg_pos = "right",
  color_map = "gray", color_frontier = "white", ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Choropleth of Latam
Choropleth of Latam
}
\section{ftypes}{
 Ge-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
