% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ye-Nu.R
\name{gg_circle_CatYeaNum.}
\alias{gg_circle_CatYeaNum.}
\title{Scatter
ciculos}
\usage{
gg_circle_CatYeaNum.(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, leg_pos = "right", shape_type = 19,
  angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Scatter
ciculos
}
\section{ftypes}{
 Cat-Yea-Num
}

\examples{
add(1, 1)
add(10, 1)
}
