% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca.R
\name{gg_circular_bar_Ca.}
\alias{gg_circular_bar_Ca.}
\title{gg_circular_bar_Ca.
Circular Bar}
\usage{
gg_circular_bar_Ca.(data, titleLabel = "Report", fillLabel = NULL,
  leg_pos = "right", width = 0.85)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_circular_bar_Ca.
Circular Bar
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

