% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca.R
\name{gg_coloured_bar_ver_Ca.}
\alias{gg_coloured_bar_hor_Ca.}
\alias{gg_coloured_bar_ver_Ca.}
\title{gg_coloured_bar_ver_Ca.
Vertical coloured bars}
\usage{
gg_coloured_bar_ver_Ca.(data, titleLabel = "Report", xLabel = NULL,
  yLabel = "Count", fillLabel = NULL, leg_pos = "right")

gg_coloured_bar_hor_Ca.(data, titleLabel = "Report", xLabel = NULL,
  yLabel = "Count", fillLabel = NULL, leg_pos = "right")
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}

\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.

The sum of \code{x} and \code{y}.
}
\description{
gg_coloured_bar_ver_Ca.
Vertical coloured bars

gg_coloured_bar_hor_Ca.
Horizontal coloured Bars
}
\section{ftypes}{
 Ca,Ca-Nu

 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
add(1, 1)
add(10, 1)
}

