% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convert-Da_funs.R
\name{gg_conv_Da2Ca_years.}
\alias{gg_conv_Da2Ca_years.}
\title{gg_conv_Da2Ca_years. : title.
Convert Data type to Category years}
\usage{
gg_conv_Da2Ca_years.(data)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_conv_Da2Ca_years. : title.
Convert Data type to Category years
}
\section{ftypes}{
 Da-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

