% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nu.R
\name{gg_cumm_dist_Nu.}
\alias{gg_cumm_dist_Nu.}
\title{gg_cumm_dist_Nu.
Cumulative distribution function}
\usage{
gg_cumm_dist_Nu.(data, title = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_cumm_dist_Nu.
Cumulative distribution function
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

