% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nu-Nu.R
\name{gg_dens_flip_NuNu.}
\alias{gg_dens_flip_NuNu.}
\title{gg_dens_flip_NuNu.
Density 2D flipped}
\usage{
gg_dens_flip_NuNu.(data, title = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, reverse = FALSE, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_dens_flip_NuNu.
Density 2D flipped
}
\section{ftypes}{
 Nu-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

