% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nu-Nu.R
\name{gg_dens_NuNu.}
\alias{gg_dens_NuNu.}
\title{Horizontal density 2D bin
Density in 2D}
\usage{
gg_dens_NuNu.(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, reverse = FALSE, angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal density 2D bin
Density in 2D
}
\section{ftypes}{
 Nu-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
