% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densities.R
\name{gg_density_multi_dist_coloured_CatNum.}
\alias{gg_density_multi_dist_coloured_CatNum.}
\title{Density distribution
Coloured Density Distribution}
\usage{
gg_density_multi_dist_coloured_CatNum.(data, titleLabel = "",
  subtitle = "", caption = "", xLabel = NULL, yLabel = NULL,
  fillLabel = NULL, leg_pos = "right", angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Density distribution
Coloured Density Distribution
}
\section{ftypes}{
 Cat-Num
}

\examples{
add(1, 1)
add(10, 1)
}
