% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_dist_hist_hor_facet_CatNum.}
\alias{gg_dist_hist_hor_facet_CatNum.}
\title{Horizontal histogram + distribution facet
Facet Horizontal Histogram + Dist}
\usage{
gg_dist_hist_hor_facet_CatNum.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, leg_pos = "right",
  angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal histogram + distribution facet
Facet Horizontal Histogram + Dist
}
\section{ftypes}{
 Cat-Num
}

\examples{
add(1, 1)
add(10, 1)
}
