% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histograms.R
\name{gg_dist_hist_ver_facet_CatNum.}
\alias{gg_dist_hist_ver_facet_CatNum.}
\alias{gg_dist_hist_ver_facet_CatNum.}
\title{Vertical histogram + distribution facet
Facet Vertical Histogram + Dist}
\usage{
gg_dist_hist_ver_facet_CatNum.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, leg_pos = "right",
  angle_x = 0, ...)

gg_dist_hist_ver_facet_CatNum.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, leg_pos = "right",
  angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}

\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.

The sum of \code{x} and \code{y}.
}
\description{
Vertical histogram + distribution facet
Facet Vertical Histogram + Dist

Vertical histogram + distribution facet
Facet Vertical Histogram + Dist
}
\section{ftypes}{
 Cat-Num

 Cat-Num
}

\examples{
add(1, 1)
add(10, 1)
add(1, 1)
add(10, 1)
}
