% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_dist_hist_ver_mean_facet_CaNu.}
\alias{gg_dist_hist_ver_mean_facet_CaNu.}
\title{gg_dist_hist_ver_mean_facet_CaNu.
Facet Vertical Histogram + Dist + Mean}
\usage{
gg_dist_hist_ver_mean_facet_CaNu.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = "Count", leg_pos = "right", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_dist_hist_ver_mean_facet_CaNu.
Facet Vertical Histogram + Dist + Mean
}
\section{ftypes}{
 Ca, Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

