% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_dist_hor_facet_CaNu.}
\alias{gg_dist_hor_facet_CaNu.}
\title{Horizontal distribution facet
Facet Horizontal Dist}
\usage{
gg_dist_hor_facet_CaNu.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, leg_pos = "right",
  angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal distribution facet
Facet Horizontal Dist
}
\section{ftypes}{
 Ca-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
