% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Da-Nu_funs.R
\name{gg_div_DaNu.}
\alias{gg_div_DaNu.}
\title{gg_div_DaNu.
div}
\usage{
gg_div_DaNu.(data, title = "", subtitle = "", caption = "", xlab = NULL,
  ylab = NULL, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_div_DaNu.
div
}
\section{ftypes}{
 Da-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

