% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca.R
\name{gg_donut_Ca.}
\alias{gg_donut_Ca.}
\title{gg_donut_Ca.
Donut}
\usage{
gg_donut_Ca.(data, titleLabel = "", subtitle = "", caption = "",
  fillLabel = NULL, width = 0.3, text = TRUE, type = "percent",
  text_size = 4, leg_pos = "right", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_donut_Ca.
Donut
}
\section{ftypes}{
 Ca
}

\examples{
add(1, 1)
add(10, 1)
}
