% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_donut_CaNu.}
\alias{gg_donut_CaNu.}
\title{gg_donut_CaNu.
dount}
\usage{
gg_donut_CaNu.(data, titleLabel = "", subtitle = "", caption = "",
  width = 0.3, leg_pos = "right", aggregation = "sum", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_donut_CaNu.
dount
}
\section{ftypes}{
 Ca-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
