% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donuts.R
\name{gg_donut_Cat.}
\alias{gg_donut_Cat.}
\title{Donut
Donut}
\usage{
gg_donut_Cat.(data, titleLabel = "", subtitle = "", caption = "",
  fillLabel = NULL, width = 0.3, text = TRUE, type = "count",
  color_text = "black", leg_pos = "right", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Donut
Donut
}
\section{ftypes}{
 Cat
}

\examples{
add(1, 1)
add(10, 1)
}
