% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportions.R
\name{gg_donut_Cat}
\alias{gg_donut_Cat}
\title{Donut (categories)}
\usage{
gg_donut_Cat(data, title = NULL, subtitle = NULL, caption = NULL,
  colors = NULL, colorText = "#5A6B72", colorScale = "discrete",
  dropNa = FALSE, prefix = NULL, suffix = NULL, labelRatio = 0.1,
  labelWrap = 12, legendPosition = "bottom", legendTitle = NULL,
  marks = c(".", ","), nDigits = 0, order = NULL,
  orientation = "ver", percentage = FALSE, sort = "no",
  sliceN = NULL, showText = TRUE, theme = NULL, ...)

gg_donut_Cat(data, title = NULL, subtitle = NULL, caption = NULL,
  colors = NULL, colorText = "#5A6B72", colorScale = "discrete",
  dropNa = FALSE, prefix = NULL, suffix = NULL, labelRatio = 0.1,
  labelWrap = 12, legendPosition = "bottom", legendTitle = NULL,
  marks = c(".", ","), nDigits = 0, order = NULL,
  orientation = "ver", percentage = FALSE, sort = "no",
  sliceN = NULL, showText = TRUE, theme = NULL, ...)
}
\arguments{
\item{data}{A data.frame}

\item{data}{A data.frame}
}
\value{
Ggplot2 visualization

Ggplot2 visualization
}
\description{
Compare category's levels

Compare category's levels
}
\section{ctypes}{

Cat, Yea, Dat


Cat, Yea, Dat
}

\examples{
gg_donut_Cat(sampleData("Cat", nrow = 10))
gg_donut_Cat(sampleData("Cat", nrow = 10))
}
