% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_donut_CatNum.}
\alias{gg_donut_CatNum.}
\title{Donut
dount}
\usage{
gg_donut_CatNum.(data, titleLabel = "", subtitle = "", caption = "",
  fillLabel = NULL, width = 0.3, leg_pos = "right", aggregation = "sum",
  text = TRUE, color_text = "black", type = "count", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Donut
dount
}
\section{ftypes}{
 Cat-Num
}
\examples{
add(1, 1)
add(10, 1)
}

