% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_pie_CatNum.R
\name{gg_donut_CatNum}
\alias{gg_donut_CatNum}
\title{Donut (categories, numbers)}
\usage{
gg_donut_CatNum(data = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Ggplot2 visualization
}
\description{
Compare aggregations among category's levels
}
\section{ctypes}{

Cat-Num, Dat-Num, Yea-Num
}

\examples{
gg_donut_CatNum(sampleData("Cat-Num", nrow = 10))
}
