% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca.R
\name{gg_donut_facet_CaCa.}
\alias{gg_donut_facet_CaCa.}
\title{Donut facet
Facet Donut}
\usage{
gg_donut_facet_CaCa.(data, titleLabel = "", subtitle = "", caption = "",
  width = 0.3, leg_pos = "right", text = TRUE, type = "count",
  color_text = "black", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Donut facet
Facet Donut
}
\section{ftypes}{
 Ca-Ca
}

\examples{
add(1, 1)
add(10, 1)
}
