% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Nu.R
\name{gg_donut_facet_CaCaNu.}
\alias{gg_donut_facet_CaCaNu.}
\title{Donut facet
Facet Donut}
\usage{
gg_donut_facet_CaCaNu.(data, titleLabel = "", subtitle = "", caption = "",
  width = 0.3, leg_pos = "right", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Donut facet
Facet Donut
}
\section{ftypes}{
 Ca-Ca-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
