% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca.R
\name{gg_donut_facet_CatCat.}
\alias{gg_donut_facet_CatCat.}
\title{Donut facet
Facet Donut}
\usage{
gg_donut_facet_CatCat.(data, titleLabel = "", subtitle = "", caption = "",
  fillLabel = NULL, width = 0.3, leg_pos = "right", text = TRUE,
  type = "count", color_text = "black", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Donut facet
Facet Donut
}
\section{ftypes}{
 Cat-Cat
}

\examples{
add(1, 1)
add(10, 1)
}
