% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histograms.R
\name{gg_dot_bar_flip_Num.}
\alias{gg_dot_bar_flip_Num.}
\title{Horizontal histogram dot bar
Dot bar flipped}
\usage{
gg_dot_bar_flip_Num.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal histogram dot bar
Dot bar flipped
}
\section{ftypes}{
 Num
}

\examples{
add(1, 1)
add(10, 1)
}
