% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca.R
\name{gg_dot_bar_hor_Ca.}
\alias{gg_dot_bar_hor_Ca.}
\title{gg_dot_bar_hor_Ca.
Horizontal Dot Bar}
\usage{
gg_dot_bar_hor_Ca.(data, titleLabel = "", xLabel = NULL, yLabel = "Count",
  subtitle = "", caption = "", fillLabel = NULL, leg_pos = "right", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_dot_bar_hor_Ca.
Horizontal Dot Bar
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

