% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_dot_bar_hor_CaNu.}
\alias{gg_dot_bar_hor_CaNu.}
\title{gg_dot_bar_hor_CaNu.
horizontal bar}
\usage{
gg_dot_bar_hor_CaNu.(data, titleLabel = "Report", xLabel = NULL,
  yLabel = NULL)
}
\arguments{
\item{x}{A category.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_dot_bar_hor_CaNu.
horizontal bar
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

