% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_dot_bar_hor_CatNum.}
\alias{gg_dot_bar_hor_CatNum.}
\title{Horizontal dot bar
Horizontal Dot Bar}
\usage{
gg_dot_bar_hor_CatNum.(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, leg_pos = "right", angle_x = 0,
  aggregation = "sum", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal dot bar
Horizontal Dot Bar
}
\section{ftypes}{
 Cat-Num
}

\examples{
add(1, 1)
add(10, 1)
}
