% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca.R
\name{gg_dot_bar_hor_facet_CaCa.}
\alias{gg_dot_bar_hor_facet_CaCa.}
\title{Horizontal dot bar facet
Facet Horizontal Dot Bar}
\usage{
gg_dot_bar_hor_facet_CaCa.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, fillLabel = NULL,
  leg_pos = "right", angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal dot bar facet
Facet Horizontal Dot Bar
}
\section{ftypes}{
 Ca-Ca
}

\examples{
add(1, 1)
add(10, 1)
}
