% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{gg_dot_bar_hor_facet_CatCat.}
\alias{gg_dot_bar_hor_facet_CatCat.}
\title{Horizontal dot bar facet
Facet Horizontal Dot Bar}
\usage{
gg_dot_bar_hor_facet_CatCat.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, leg_pos = "right",
  angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal dot bar facet
Facet Horizontal Dot Bar
}
\section{ftypes}{
 Cat-Cat
}

\examples{
add(1, 1)
add(10, 1)
}
