% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nu.R
\name{gg_dot_bar_Nu.}
\alias{gg_dot_bar_Nu.}
\title{gg_dot_bar_Nu.
Dot bar}
\usage{
gg_dot_bar_Nu.(data, title = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_dot_bar_Nu.
Dot bar
}
\section{ftypes}{
 Ca-Nu, Nu, Nu-Nu, Da-Nu, Ye-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

