% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_dot_bar_ver_CaNu.}
\alias{gg_donut_CaNu.}
\alias{gg_dot_bar_ver_CaNu.}
\title{gg_dot_bar_ver_CaNu.
Vertical Dot Bar}
\usage{
gg_donut_CaNu.(data, titleLabel = "", subtitle = "", caption = "",
  fillLabel = NULL, width = 0.3, leg_pos = "right", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_dot_bar_ver_CaNu.
Vertical Dot Bar
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

