% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca.R
\name{gg_dot_bar_ver_facet_CaCa.}
\alias{gg_dot_bar_ver_facet_CaCa.}
\title{gg_dot_bar_ver_facet_CaCa.
Facet Vertical Dot Bar}
\usage{
gg_dot_bar_ver_facet_CaCa.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, ncol = 1,
  fillLabel = NULL, leg_pos = "right", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_dot_bar_ver_facet_CaCa.
Facet Vertical Dot Bar
}
\section{ftypes}{
 Ca,Ca-Nu, Ca-Ca
}
\examples{
add(1, 1)
add(10, 1)
}

