% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_dot_dist_hor_facet_CaNu.}
\alias{gg_dot_dist_hor_facet_CaNu.}
\title{gg_dot_dist_hor_facet_CaNu.
Facet Horizontal Dot Dist}
\usage{
gg_dot_dist_hor_facet_CaNu.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = "Count", leg_pos = "right",
  size = 3, alpha = 0.3, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_dot_dist_hor_facet_CaNu.
Facet Horizontal Dot Dist
}
\section{ftypes}{
 Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

