% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densities.R, R/histograms.R
\name{gg_dot_dist_ver_facet_CatNum.}
\alias{gg_dot_dist_ver_facet_CatNum.}
\alias{gg_dot_dist_ver_facet_CatNum.}
\title{Vertical density dot + distribution facet
Facet Vertical Dot Dist}
\usage{
gg_dot_dist_ver_facet_CatNum.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, leg_pos = "right",
  alpha = 0.3, angle_x = 0, ...)

gg_dot_dist_ver_facet_CatNum.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, leg_pos = "right",
  alpha = 0.3, angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}

\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.

The sum of \code{x} and \code{y}.
}
\description{
Vertical density dot + distribution facet
Facet Vertical Dot Dist

Vertical density dot + distribution facet
Facet Vertical Dot Dist
}
\section{ftypes}{
 Cat-Num

 Cat-Num
}

\examples{
add(1, 1)
add(10, 1)
add(1, 1)
add(10, 1)
}
