% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histograms.R
\name{gg_dot_hist_hor_mean_facet_CatNum.}
\alias{gg_dot_hist_hor_mean_facet_CatNum.}
\title{Horizontal density dot + histogram + mean facet
Facet Horizontal Histogram + Mean + Dot}
\usage{
gg_dot_hist_hor_mean_facet_CatNum.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, leg_pos = "right",
  alpha = 0.3, angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal density dot + histogram + mean facet
Facet Horizontal Histogram + Mean + Dot
}
\section{ftypes}{
 Cat-Num
}

\examples{
add(1, 1)
add(10, 1)
}
