% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histograms.R
\name{gg_dot_hist_ver_facet_CatNum.}
\alias{gg_dot_hist_ver_facet_CatNum.}
\title{Vertical density dot + histogram facet
Facet Vertical Dot Histogram}
\usage{
gg_dot_hist_ver_facet_CatNum.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, leg_pos = "right",
  alpha = 0.3, angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical density dot + histogram facet
Facet Vertical Dot Histogram
}
\section{ftypes}{
 Cat-Num
}

\examples{
add(1, 1)
add(10, 1)
}
