% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_facet_area_hor_CaNu.}
\alias{gg_facet_area_hor_CaNu.}
\title{gg_facet_area_hor_CaNu.
Facet Horizontal Area}
\usage{
gg_facet_area_hor_CaNu.(data, titleLabel = "", xLabel = "Index",
  yLabel = NULL, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_facet_area_hor_CaNu.
Facet Horizontal Area
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

