% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca.R
\name{gg_facet_bullseye_CaCa.}
\alias{gg_facet_bullseye_CaCa.}
\title{gg_facet_bullseye_CaCa.
Facet Bullseye}
\usage{
gg_facet_bullseye_CaCa.(data, titleLabel = "", leg_pos = "right", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_facet_bullseye_CaCa.
Facet Bullseye
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

