% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Nu.R
\name{gg_facet_circular_bar_CaCaNu.}
\alias{gg_facet_circular_bar_CaCaNu.}
\title{gg_facet_circular_bar_CaCaNu.
Circular Bar}
\usage{
gg_facet_circular_bar_CaCaNu.(data, titleLabel = "Report", fillLabel = NULL,
  leg_pos = "right", width = 0.85)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_facet_circular_bar_CaCaNu.
Circular Bar
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

