% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Nu.R
\name{gg_facet_coloured_bar_z_ver_CaCaNu.}
\alias{gg_facet_coloured_bar_z_ver_CaCaNu.}
\title{gg_facet_coloured_bar_z_ver_CaCaNu.
Facet coloured vertical bar}
\usage{
gg_facet_coloured_bar_z_ver_CaCaNu.(data, titleLabel = "Report",
  xLabel = NULL, yLabel = NULL, fillLabel = NULL, leg_pos = "right")
}
\arguments{
\item{x}{A category.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_facet_coloured_bar_z_ver_CaCaNu.
Facet coloured vertical bar
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

