% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Nu.R
\name{gg_facet_coloured_parameter_bar_ver_CaCaNu.}
\alias{gg_facet_coloured_parameter_bar_ver_CaCaNu.}
\title{gg_facet_coloured_parameter_bar_ver_CaCaNu.
Facet Vertical coloured by parameter bars}
\usage{
gg_facet_coloured_parameter_bar_ver_CaCaNu.(data, titleLabel = "Report",
  xLabel = NULL, yLabel = "Count", parameter1 = NULL, parameter2 = NULL,
  leg_pos = "right")
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_facet_coloured_parameter_bar_ver_CaCaNu.
Facet Vertical coloured by parameter bars
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

