% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_facet_dist_hist_hor_CaNu.}
\alias{gg_facet_dist_hist_hor_CaNu.}
\alias{gg_facet_dist_hist_ver_CaNu.}
\title{gg_facet_dist_hist_hor_CaNu.
Facet Vertical Histogram + Dist}
\usage{
gg_facet_dist_hist_ver_CaNu.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = "Count", fillLabel = NULL,
  leg_pos = "right", ...)

gg_facet_dist_hist_hor_CaNu.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = "Count", fillLabel = NULL,
  leg_pos = "right", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}

\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.

The sum of \code{x} and \code{y}.
}
\description{
gg_facet_dist_hist_hor_CaNu.
Facet Vertical Histogram + Dist

gg_facet_dist_hist_hor_CaNu.
Facet Horizontal Histogram + Dist
}
\section{ftypes}{
 Ca,Ca-Nu

 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
add(1, 1)
add(10, 1)
}

