% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca.R
\name{gg_facet_donut_CaCa.}
\alias{gg_facet_donut_CaCa.}
\title{gg_facet_donut_CaCa.
Facet Donut}
\usage{
gg_facet_donut_CaCa.(data, titleLabel = "", fillLabel = NULL, width = 0.3,
  leg_pos = "right")
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_facet_donut_CaCa.
Facet Donut
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

