% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca.R
\name{gg_facet_dot_bar_ver_CaCa.}
\alias{gg_facet_dot_bar_ver_CaCa.}
\title{gg_facet_dot_bar_ver_CaCa.
Facet Vertical Dot Bar}
\usage{
gg_facet_dot_bar_ver_CaCa.(data, titleLabel = "", xLabel = NULL,
  yLabel = NULL, fillLabel = NULL, leg_pos = "right", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_facet_dot_bar_ver_CaCa.
Facet Vertical Dot Bar
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

