% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_facet_dot_dist_hist_mean_hor_CaNu.}
\alias{gg_facet_dot_dist_hist_mean_hor_CaNu.}
\title{gg_facet_dot_dist_hist_mean_hor_CaNu.
Facet Horizontal Histogram + Dist + Mean + Dot}
\usage{
gg_facet_dot_dist_hist_mean_hor_CaNu.(data, titleLabel = "Report",
  xLabel = NULL, yLabel = "Count", fillLabel = NULL, leg_pos = "right",
  ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_facet_dot_dist_hist_mean_hor_CaNu.
Facet Horizontal Histogram + Dist + Mean + Dot
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

