% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_facet_hist_ver_CaNu.}
\alias{gg_facet_hist_mean_ver_CaNu.}
\alias{gg_facet_hist_ver_CaNu.}
\title{gg_facet_hist_ver_CaNu.
Facet Vertical Histogram + Mean}
\usage{
gg_facet_hist_mean_ver_CaNu.(data, titleLabel = "Report", xLabel = NULL,
  yLabel = "Count", fillLabel = NULL, leg_pos = "right")

gg_facet_hist_ver_CaNu.(data, titleLabel = "Report", xLabel = NULL,
  yLabel = "Count", fillLabel = NULL, leg_pos = "right")
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}

\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.

The sum of \code{x} and \code{y}.
}
\description{
gg_facet_hist_ver_CaNu.
Facet Vertical Histogram + Mean

gg_facet_hist_ver_CaNu.
Facet Vertical Histogram
}
\section{ftypes}{
 Ca,Ca-Nu

 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
add(1, 1)
add(10, 1)
}

