% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_facet_line_CaNu.}
\alias{gg_facet_line_CaNu.}
\title{gg_facet_line_CaNu.
Facet Line}
\usage{
gg_facet_line_CaNu.(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = "Index", yLabel = NULL, fillLabel = NULL, type = 1, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_facet_line_CaNu.
Facet Line
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

