% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca.R
\name{gg_facet_line_point_ver_CaCa.}
\alias{gg_facet_line_point_ver_CaCa.}
\title{gg_facet_line_point_ver_CaCa.
Vertical Line Point}
\usage{
gg_facet_line_point_ver_CaCa.(data, titleLabel = "", xLabel = NULL,
  yLabel = "Count", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_facet_line_point_ver_CaCa.
Vertical Line Point
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

