% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca.R
\name{gg_facet_pie_CaCa.}
\alias{gg_facet_pie_CaCa.}
\title{gg_facet_pie_CaCa.
Facet Pie}
\usage{
gg_facet_pie_CaCa.(data, titleLabel = "", fillLabel = NULL,
  leg_pos = "right", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_facet_pie_CaCa.
Facet Pie
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

