% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_facet_point_CaNu.}
\alias{gg_facet_point_CaNu.}
\title{gg_facet_point_CaNu.
Facet Point}
\usage{
gg_facet_point_CaNu.(data, titleLabel = "Report", xLabel = "Index",
  yLabel = NULL, fillLabel = NULL, type = 1)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_facet_point_CaNu.
Facet Point
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

