% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nu.R
\name{gg_flip_box_Nu.}
\alias{gg_flip_box_Nu.}
\title{gg_flip_box_Nu.
Box plot}
\usage{
gg_flip_box_Nu.(data, title = "", subtitle = "", caption = "",
  yLabel = NULL, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_flip_box_Nu.
Box plot
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

