% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nu-Nu.R
\name{gg_flip_hist2D_NuNu.}
\alias{gg_flip_hist2D_NuNu.}
\title{gg_flip_hist2D_NuNu.
Waterfall}
\usage{
gg_flip_hist2D_NuNu.(data, title = "", subtitle = "", caption = "",
  xLabel = "", yLabel = "", labelText = "", reverse = FALSE, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_flip_hist2D_NuNu.
Waterfall
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

