% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca.R
\name{gg_gauge_Ca.}
\alias{gg_gauge_Ca.}
\title{Gauge
Gauge}
\usage{
gg_gauge_Ca.(data, titleLabel = "", subtitle = "", caption = "",
  ncol = 2, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Gauge
Gauge
}
\section{ftypes}{
 Ca
}

\examples{
add(1, 1)
add(10, 1)
}
