% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nu.R
\name{gg_hist_dens_Nu.}
\alias{gg_hist_dens_Nu.}
\title{Histogram + density
Histograms with density}
\usage{
gg_hist_dens_Nu.(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, alfa = 0.5, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Histogram + density
Histograms with density
}
\section{ftypes}{
 Nu
}

\examples{
add(1, 1)
add(10, 1)
}
