% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nu-Nu.R
\name{gg_hist_flip_NumNum.}
\alias{gg_hist_flip_NumNum.}
\title{Horizontal histogram
Histogram 2d flipped}
\usage{
gg_hist_flip_NumNum.(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, reverse = FALSE, angle_x = 0,
  leg_pos = "right", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal histogram
Histogram 2d flipped
}
\section{ftypes}{
 Num-Num
}
\examples{
add(1, 1)
add(10, 1)
}

