% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_hist_hor_facet_CaNu.}
\alias{gg_hist_hor_facet_CaNu.}
\title{Horizontal histogram facet
Facet Horizontal Histogram}
\usage{
gg_hist_hor_facet_CaNu.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = "Count", leg_pos = "right", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal histogram facet
Facet Horizontal Histogram
}
\section{ftypes}{
 Ca-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
