% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nu.R
\name{gg_hist_Nu.}
\alias{gg_hist_Nu.}
\title{gg_hist_Nu.
Histogram}
\usage{
gg_hist_Nu.(data, title = "", subtitle = "", caption = "", xLabel = "",
  yLabel = NULL, size = 1, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_hist_Nu.
Histogram
}
\section{ftypes}{
 Nu
}
\examples{
add(1, 1)
add(10, 1)
}

