% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nu.R
\name{gg_hist_Num.}
\alias{gg_hist_Num.}
\title{Histogram
Histogram}
\usage{
gg_hist_Num.(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Histogram
Histogram
}
\section{ftypes}{
 Num
}
\examples{
add(1, 1)
add(10, 1)
}

