% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histograms.R
\name{gg_hist_stacked_ver_CatNum.}
\alias{gg_hist_stacked_ver_CatNum.}
\title{Vertical stacked histogram
Stacked Vertical Histogram}
\usage{
gg_hist_stacked_ver_CatNum.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, fillLabel = NULL,
  leg_pos = "right", angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical stacked histogram
Stacked Vertical Histogram
}
\section{ftypes}{
 Cat-Num
}

\examples{
add(1, 1)
add(10, 1)
}
