% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histograms.R
\name{gg_histogram_CatDat.}
\alias{gg_histogram_CatDat.}
\title{Histogram
histogram}
\usage{
gg_histogram_CatDat.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, fillLabel = NULL,
  angle_x = 0, leg_pos = "right", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Histogram
histogram
}
\section{ftypes}{
 Cat-Dat
}

\examples{
add(1, 1)
add(10, 1)
}
