% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Da-Nu.R
\name{gg_horizon_DaNu.}
\alias{gg_horizon_DaNu.}
\title{gg_horizon_DaNu.
Horizon}
\usage{
gg_horizon_DaNu.(data, title = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, leg_pos = "right", reverse = FALSE,
  angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_horizon_DaNu.
Horizon
}
\section{ftypes}{
 Da-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
