% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nu.R
\name{gg_horizon_Nu.}
\alias{gg_horizon_Nu.}
\title{gg_horizon_Nu.
Horizon}
\usage{
gg_horizon_Nu.(data, titleLabel = "Report", xLabel = "Index",
  yLabel = NULL, leg_pos = "right", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_horizon_Nu.
Horizon
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

