% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{gg_horizon_NumNum.}
\alias{gg_horizon_NumNum.}
\title{Horizon
Horizon}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizon
Horizon
}
\section{ftypes}{
 Num-Num
}

\examples{
add(1, 1)
add(10, 1)
}
