% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{gg_horizon_YeaNum.}
\alias{gg_horizon_YeaNum.}
\title{Horizon
Horizon}
\usage{
gg_horizon_YeaNum.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, leg_pos = "right",
  reverse = FALSE, angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizon
Horizon
}
\section{ftypes}{
 Yea-Num
}

\examples{
add(1, 1)
add(10, 1)
}
