% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ye-Nu.R
\name{gg_horizon_YeNu.}
\alias{gg_horizon_YeNu.}
\title{gg_horizon_YeNu.
Horizon}
\usage{
gg_horizon_YeNu.(data, title = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, leg_pos = "right", reverse = FALSE, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_horizon_YeNu.
Horizon
}
\section{ftypes}{
 Ye-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
