% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possible_viz.R
\name{gg_line_Cat.}
\alias{gg_line_Cat.}
\title{Horizontal line
Horizontal Line}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal line
Horizontal Line
}
\section{ftypes}{
 Cat
}

\examples{
gg_line_Cat.(sampleData("Cat"))
}
