% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{gg_line_CatCat}
\alias{gg_line_CatCat}
\title{Lines (categories, ordered categories)}
\usage{
gg_line_CatCat(data, title = NULL, subtitle = NULL, caption = NULL,
  horLabel = NULL, verLabel = NULL, horLine = NULL, verLine = NULL,
  colors = NULL, colorText = "#5A6B72", colorScale = "discrete",
  dropNaV = c(FALSE, FALSE), prefix = NULL, suffix = NULL,
  labelRatio = 1, labelWrapV = c(12, 12), legendPosition = "bottom",
  legendTitle = NULL, marks = c(".", ","), nDigits = 0,
  order = NULL, orientation = "ver", percentage = FALSE,
  shapeType = 19, showText = TRUE, spline = FALSE,
  startAtZero = TRUE, theme = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Ggplot2 visualization
}
\description{
Compare quantities among two categories
}
\section{ctypes}{

Cat-Cat, Cat-Dat, Cat-Yea, Yea-Cat, Yea-Dat, Yea-Yea, Dat-Cat, Dat-Yea, Dat-Dat
}

\examples{
gg_line_CatCat(sampleData("Cat-Cat", nrow = 10))
}
