% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{gg_line_CatCatNum}
\alias{gg_line_CatCatNum}
\title{Lines (categories, ordered categories, numbers)}
\usage{
gg_line_CatCatNum(data = NULL, opts = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Ggplot2 visualization
}
\description{
Compare quantities among two categories
}
\section{ctypes}{

Cat-Cat-Num, Cat-Dat-Num, Cat-Yea-Num, Yea-Cat-Num, Yea-Dat-Num, Yea-Yea-Num, Dat-Cat-Num, Dat-Yea-Num, Dat-Dat-Num
}

\examples{
gg_line_CatCatNum(sampleData("Cat-Cat-Num", nrow = 10))
}
