% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu-Nu.R
\name{gg_line_CatNumNum.}
\alias{gg_line_CatNumNum.}
\title{Line
lines}
\usage{
gg_line_CatNumNum.(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, fillLabel = NULL, leg_pos = "right",
  angle_x = 0, ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Line
lines
}
\section{ftypes}{
 Cat-Num-Num
}

\examples{
add(1, 1)
add(10, 1)
}
