% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_line_CatYeaNum.R
\name{gg_line_CatYea}
\alias{gg_line_CatYea}
\title{line Chart Cat Yea}
\usage{
gg_line_CatYea(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
This chart does not allow for chaning orientation
}
\examples{
gg_line_CatYea(sample_data("Cat-Yea", nrow = 10))
}
