% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{gg_line_CatYeaNum}
\alias{gg_line_CatYeaNum}
\title{Lines (catrgories, years, numbers)}
\usage{
gg_line_CatYeaNum(data, title = NULL, subtitle = NULL,
  caption = NULL, horLabel = NULL, verLabel = NULL, horLine = NULL,
  verLine = NULL, agg = "sum", colors = NULL, colorText = "black",
  colorScale = "discrete", dropNa = c(FALSE, FALSE), format = c("",
  ""), graphType = "grouped", labelRatio = 1, labelWrap = c(12, 12),
  legendPosition = "right", legendTitle = NULL, marks = c(".", ","),
  nDigits = 2, order = NULL, orientation = "ver",
  percentage = FALSE, shapeType = 19, showText = TRUE,
  spline = FALSE, startAtZero = TRUE, theme = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Ggplot2 visualization
}
\description{
Compare quantities among categories over years
}
\section{ctypes}{

Cat-Yea-Num
}

\examples{
gg_line_CatYeaNum(sampleData("Cat-Yea-Num", nrow = 10))
}
