% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Da-Nu.R
\name{gg_line_DaNu.}
\alias{gg_line_DaNu.}
\title{Line
Lines}
\usage{
gg_line_DaNu.(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Line
Lines
}
\section{ftypes}{
 Da-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
