% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possible_viz.R
\name{gg_line_facet_CatCatNum.}
\alias{gg_line_facet_CatCatNum.}
\title{Horizontal line facet
horizontal linegraph}
\arguments{
\item{x}{A category.}

\item{y}{A category.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal line facet
horizontal linegraph
}
\section{ftypes}{
 Cat-Cat-Num
}

\examples{
gg_line_facet_CatCatNum.(sampleData("Cat-Cat-Num"))
}
