% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possible_viz.R
\name{gg_line_facet_ver_CatCatNum.}
\alias{gg_line_facet_ver_CatCatNum.}
\title{Vertical line facet
vertical linegraph}
\arguments{
\item{x}{A category.}

\item{y}{A category.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical line facet
vertical linegraph
}
\section{ftypes}{
 Cat-Cat-Num
}

\examples{
add(1, 1)
add(10, 1)
}
