% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca.R
\name{gg_line_hor_Ca.}
\alias{gg_line_hor_Ca.}
\title{Horizontal line
Horizontal Line}
\usage{
gg_line_hor_Ca.(data, titleLabel = "", xLabel = NULL, subtitle = "",
  caption = "", yLabel = NULL, leg_pos = "right", angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal line
Horizontal Line
}
\section{ftypes}{
 Ca
}

\examples{
add(1, 1)
add(10, 1)
}
