% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ye-Nu.R
\name{gg_line_hor_CaYeNu.}
\alias{gg_line_hor_CaYeNu.}
\title{Horizontal line
Tiene múltiples líneas}
\usage{
gg_line_hor_CaYeNu.(data, title = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, leg_pos = "right", angle = 0,
  nbreaks = NULL, symbol = NULL, aggregation = "mean", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal line
Tiene múltiples líneas
}
\section{ftypes}{
 Ca-Ye-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
