% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Nu.R
\name{gg_line_hor_facet_CatCatNum.}
\alias{gg_line_hor_facet_CatCatNum.}
\title{Horizontal line facet
horizontal linegraph}
\usage{
gg_line_hor_facet_CatCatNum.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, aggregation = "sum",
  angle_x = 0, shape_type = 19, ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A category.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal line facet
horizontal linegraph
}
\section{ftypes}{
 Cat-Cat-Num
}

\examples{
add(1, 1)
add(10, 1)
}
