% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ye-Nu.R
\name{gg_line_hor_YeNu.}
\alias{gg_line_hor_YeNu.}
\title{Horizontal line
Horizontal Lines}
\usage{
gg_line_hor_YeNu.(data, title = "", subtitle = "", caption = "",
  xlab = NULL, ylab = NULL, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal line
Horizontal Lines
}
\section{ftypes}{
 Ye-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
