% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possible_viz.R
\name{gg_line_multi_CatCatNum.}
\alias{gg_line_multi_CatCatNum.}
\title{Grouped line by first variable
Grouped Line Coloured}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Grouped line by first variable
Grouped Line Coloured
}
\section{ftypes}{
 Cat-Cat-Num
}

\examples{
gg_line_multi_CatCatNum.(sampleData("Cat-Cat-Num"))
}
