% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{gg_line_multi_CatCatNum}
\alias{gg_line_multi_CatCatNum}
\title{Grouped line by first variable
Grouped Line Coloured}
\usage{
gg_line_multi_CatCatNum(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, fillLabel = NULL,
  aggregation = "sum", leg_pos = "right", angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Grouped line by first variable
Grouped Line Coloured
}
\section{ftypes}{
 Cat-Cat-Num
}

\examples{
gg_line_multi_CatCatNum(sampleData("Cat-Cat-Num"))
}
