% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possible_viz.R
\name{gg_line_multi_CatDatNum.}
\alias{gg_line_multi_CatDatNum.}
\title{Grouped line
Grouped Line Coloured}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Grouped line
Grouped Line Coloured
}
\section{ftypes}{
 Cat-Dat-Num
}

\examples{
gg_line_multi_CatDatNum.(sampleData("Cat-Dat-Num"))
}
