% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possible_viz.R
\name{gg_line_multi_CatNumNum.}
\alias{gg_line_multi_CatNumNum.}
\title{Line
lines}
\arguments{
\item{x}{A category.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Line
lines
}
\section{ftypes}{
 Cat-Num-Num
}

\examples{
gg_line_multi_CatNumNum.(sampleData("Cat-Num-Num"))
}
