% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{gg_line_multi_CatYeaNum}
\alias{gg_line_multi_CatYeaNum}
\title{Horizontal line
Tiene múltiples líneas}
\usage{
gg_line_multi_CatYeaNum(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, fillLabel = NULL,
  leg_pos = "right", angle_x = 0, nbreaks = NULL, shape_type = 19,
  aggregation = "sum", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal line
Tiene múltiples líneas
}
\section{ftypes}{
 Cat-Yea-Num
}

\examples{
gg_line_multi_CatYeaNum(sampleData("Cat-Yea-Num"))
}
