% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possible_viz.R
\name{gg_line_multi_NumNum.}
\alias{gg_line_multi_NumNum.}
\title{Line for each numeric variable
Line for each Num}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Line for each numeric variable
Line for each Num
}
\section{ftypes}{
 Num-Num
}

\examples{
gg_line_multi_NumNum.(sampleData("Num-Num"))
}
