% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nu-Nu.R
\name{gg_line_multi_NuNu.}
\alias{gg_line_multi_NuNu.}
\title{Line for each numeric variable
Line for each Nu}
\usage{
gg_line_multi_NuNu.(data, title = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Line for each numeric variable
Line for each Nu
}
\section{ftypes}{
 Nu-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
