% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possible_viz.R
\name{gg_line_Num.}
\alias{gg_line_Num.}
\title{Vertical line
Line}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical line
Line
}
\section{ftypes}{
 Num
}

\examples{
gg_line_Num.(sampleData("Num"))
}
