% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{gg_line_point_Cat}
\alias{gg_line_point_Cat}
\title{Horizontal line + point
Horizontal Line Point}
\usage{
gg_line_point_Cat(data, titleLabel = "", xLabel = NULL,
  subtitle = "", caption = "", yLabel = NULL, leg_pos = "right",
  shape_type = 19, angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal line + point
Horizontal Line Point
}
\section{ftypes}{
 Cat
}

\examples{
gg_line_point_Cat(sampleData("Cat"))
}
