% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{gg_line_point_DatNum}
\alias{gg_line_point_DatNum}
\title{Line + point
Lines}
\usage{
gg_line_point_DatNum(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, shape_type = 19, hline = NULL,
  angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Line + point
Lines
}
\section{ftypes}{
 Dat-Num
}

\examples{
gg_line_point_DatNum(sampleData("Dat-Num"))
}
