% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{gg_line_point_facet_CatNum}
\alias{gg_line_point_facet_CatNum}
\title{Line + point facet
Facet Line Point}
\usage{
gg_line_point_facet_CatNum(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, shape_type = 19,
  angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Line + point facet
Facet Line Point
}
\section{ftypes}{
 Cat-Num
}

\examples{
gg_line_point_facet_CatNum(sampleData("Cat-Num"))
}
