% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possible_viz.R
\name{gg_line_point_facet_DatNum}
\alias{gg_line_point_facet_DatNum}
\title{Line + point facet by years
Line Points facet by years}
\usage{
gg_line_point_facet_DatNum(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, angle_x = 0,
  shape_type = 19, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Line + point facet by years
Line Points facet by years
}
\section{ftypes}{
 Dat-Num
}

\examples{
gg_line_point_facet_DatNum(sampleData("Dat-Num"))
}
