% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{gg_line_point_facet_hor_CatCat}
\alias{gg_line_point_facet_hor_CatCat}
\title{Horizontal line + point facet
Horizontal Line Point}
\usage{
gg_line_point_facet_hor_CatCat(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, angle_x = 0,
  shape_type = 19, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal line + point facet
Horizontal Line Point
}
\section{ftypes}{
 Cat-Cat
}

\examples{
gg_line_point_facet_hor_CatCat(sampleData("Cat-Cat"))
}
