% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possible_viz.R
\name{gg_line_point_facet_ver_CatCat.}
\alias{gg_line_point_facet_ver_CatCat.}
\title{Vertical line + point facet
Vertical Line Point}
\usage{
gg_line_point_facet_ver_CatCat.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, angle_x = 0,
  shape_point = 19, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical line + point facet
Vertical Line Point
}
\section{ftypes}{
 Cat-Cat
}

\examples{
add(1, 1)
add(10, 1)
}
