% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nu.R
\name{gg_line_point_flip_Nu.}
\alias{gg_line_point_flip_Nu.}
\title{gg_line_point_flip_Nu.
Line with point plot}
\usage{
gg_line_point_flip_Nu.(data, title = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_line_point_flip_Nu.
Line with point plot
}
\section{ftypes}{
 Nu, Nu-Nu, Ye-Nu, Da-Nu, Ca-Nu, Ca
}
\examples{
add(1, 1)
add(10, 1)
}

