% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possible_viz.R
\name{gg_line_point_multi_CatCatNum.}
\alias{gg_line_point_multi_CatCatNum.}
\title{Grouped line + points by first variable
Grouped Line Color Point}
\usage{
gg_line_point_multi_CatCatNum.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, fillLabel = NULL,
  leg_pos = "right", aggregation = "sum", shape_type = 19,
  angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Grouped line + points by first variable
Grouped Line Color Point
}
\section{ftypes}{
 Cat-Cat-Num
}

\examples{
gg_line_point_multi_CatCatNum.(sampleData("Cat-Cat-Num"))
}
