% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possible_viz.R
\name{gg_line_point_multi_CatNum.}
\alias{gg_line_point_multi_CatNum.}
\title{Grouped line + point
Grouped Line Color Point}
\usage{
gg_line_point_multi_CatNum.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, fillLabel = NULL,
  leg_pos = "right", shape_type = 19, angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Grouped line + point
Grouped Line Color Point
}
\section{ftypes}{
 Cat-Num
}

\examples{
gg_line_point_multi_CatNum.(sampleData("Cat-Num"))
}
