% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possible_viz.R
\name{gg_line_point_multi_CatNumNum}
\alias{gg_line_point_multi_CatNumNum}
\title{Line + point
Point Lines}
\usage{
gg_line_point_multi_CatNumNum(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, fillLabel = NULL,
  leg_pos = "right", shape_type = 19, angle_x = 0, ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Line + point
Point Lines
}
\section{ftypes}{
 Cat-Num-Num
}

\examples{
gg_line_point_multi_CatNumNum(sampleData("Cat-Num-Num"))
}
