% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possible_viz.R
\name{gg_line_point_multi_CatYeaNum.}
\alias{gg_line_point_multi_CatYeaNum.}
\title{Horizontal line + points
Tiene múltiples líneas con puntos}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal line + points
Tiene múltiples líneas con puntos
}
\section{ftypes}{
 Cat-Yea-Num
}

\examples{
gg_line_point_multi_CatYeaNum.(sampleData("Cat-Yea-Num"))
}
