% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nu.R
\name{gg_line_point_Nu.}
\alias{gg_line_point_Nu.}
\title{Vertical line + point
Line wiht point plot}
\usage{
gg_line_point_Nu.(data, title = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical line + point
Line wiht point plot
}
\section{ftypes}{
 Nu
}

\examples{
add(1, 1)
add(10, 1)
}
