% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{gg_line_point_Num}
\alias{gg_line_point_Num}
\title{Vertical line + point
Line wiht point plot}
\usage{
gg_line_point_Num(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, shape_type = 19, angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical line + point
Line wiht point plot
}
\section{ftypes}{
 Num
}

\examples{
gg_line_point_Num(sampleData("Num"))
}
