% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nu-Nu.R
\name{gg_line_point_NuNu.}
\alias{gg_line_point_NuNu.}
\title{gg_line_point_NuNu.
Line point plot}
\usage{
gg_line_point_NuNu.(data, title = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, shape = 1, size = 3, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_line_point_NuNu.
Line point plot
}
\section{ftypes}{
 Ca-Ye-Nu, Ye-Nu, Nu-Nu, Da-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

