% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possible_viz.R
\name{gg_line_point_ver_Cat.}
\alias{gg_line_point_ver_Cat.}
\title{Vertical line + point
Vertical Line Point}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical line + point
Vertical Line Point
}
\section{ftypes}{
 Cat
}

\examples{
add(1, 1)
add(10, 1)
}
