% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca.R
\name{gg_line_point_ver_facet_CaCa.}
\alias{gg_line_point_ver_facet_CaCa.}
\title{Vertical line + point facet
Vertical Line Point}
\usage{
gg_line_point_ver_facet_CaCa.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, angle_x = 0,
  shape_point = 19, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical line + point facet
Vertical Line Point
}
\section{ftypes}{
 Ca-Ca
}

\examples{
add(1, 1)
add(10, 1)
}
