% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possible_viz.R
\name{gg_line_point_YeaNum.}
\alias{gg_line_point_YeaNum.}
\title{Line + point
Lines}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Line + point
Lines
}
\section{ftypes}{
 Yea-Num
}

\examples{
gg_line_point_YeaNum.(sampleData("Yea-Num"))
}
