% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Da-Nu.R
\name{gg_line_points_facet_DaNu.}
\alias{gg_line_points_facet_DaNu.}
\title{Line + point facet by years
Line Points facet by years}
\usage{
gg_line_points_facet_DaNu.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, angle_x = 0,
  shape_type = 19, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Line + point facet by years
Line Points facet by years
}
\section{ftypes}{
 Da-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
