% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca.R
\name{gg_line_ver_Ca.}
\alias{gg_line_ver_Ca.}
\title{gg_line_ver_Ca.
Vertical Line}
\usage{
gg_line_ver_Ca.(data, titleLabel = "Report", xLabel = NULL,
  yLabel = "Count", leg_pos = "right")
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_line_ver_Ca.
Vertical Line
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

