% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca.R
\name{gg_line_ver_facet_CaCa.}
\alias{gg_line_ver_facet_CaCa.}
\title{gg_line_ver_facet_CaCa.
Vertical Line}
\usage{
gg_line_ver_facet_CaCa.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = "Count", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_line_ver_facet_CaCa.
Vertical Line
}
\section{ftypes}{
 Ca-Ca
}
\examples{
add(1, 1)
add(10, 1)
}

