% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Nu.R
\name{gg_line_ver_facet_CaCaNu.}
\alias{gg_line_ver_facet_CaCaNu.}
\title{Vertical line facet
vertical linegraph}
\usage{
gg_line_ver_facet_CaCaNu.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = "Types", yLabel = "Frequency", ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A category.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical line facet
vertical linegraph
}
\section{ftypes}{
 Ca-Ca-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
