% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R, R/possible_viz.R
\name{gg_line_YeaNum}
\alias{gg_line_YeaNum}
\title{Lines (years, numbers)}
\usage{
gg_line_YeaNum(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, angle_x = 0, ...)

gg_line_YeaNum(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, angle_x = 0, ...)
}
\arguments{
\item{data}{A data.frame}

\item{x}{A number.}

\item{y}{A number.}
}
\value{
Ggplot2 visualization

The sum of \code{x} and \code{y}.
}
\description{
Compare quantities over years
}
\section{ctypes}{

Cat-Num, Dat-Num, Yea-Num
}

\section{ftypes}{
 Yea-Num
}

\examples{
gg_line_YeaNum(sampleData("Yea-Num", nrow = 10))
gg_line_YeaNum(sampleData("Yea-Num"))
}
