% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ye-Nu.R
\name{gg_lines_hor_YeNu.}
\alias{gg_lines_hor_YeNu.}
\title{gg_lines_hor_YeNu.: title.
Horizontal Lines
Tiene múltiples líneas}
\usage{
gg_lines_hor_YeNu.(data, title = "", subtitle = "", caption = "",
  xlab = NULL, ylab = NULL, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_lines_hor_YeNu.: title.
Horizontal Lines
Tiene múltiples líneas
}
\section{ftypes}{
 Ye-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

