% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Da-Nu_funs.R
\name{gg_lines_points_DaNu.}
\alias{gg_lines_points_DaNu.}
\title{gg_lines_points_DaNu.
Lines}
\usage{
gg_lines_points_DaNu.(data, title = "", subtitle = "", caption = "",
  xlab = NULL, ylab = NULL, type = 1, hline = NULL, angle_x = 45,
  size_text = 8, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_lines_points_DaNu.
Lines
}
\section{ftypes}{
 Da-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

