% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Da-Nu.R
\name{gg_lollipop_DaNu.}
\alias{gg_lollipop_DaNu.}
\title{gg_lollipop_DaNu.
lollipop}
\usage{
gg_lollipop_DaNu.(data, title = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, angle_x = 0, ...)
}
\arguments{
\item{x}{A data.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_lollipop_DaNu.
lollipop
}
\section{ftypes}{
 Ca-Ye-Nu, Ye-Nu, Nu-Nu, Da-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

