% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ye-Nu.R
\name{gg_lollipop_YeaNum.}
\alias{gg_lollipop_YeaNum.}
\title{Lollipop
lollipop. segment and point
gg_lollipop_YeaNum.}
\usage{
gg_lollipop_YeaNum.(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, angle_x = 0, shape_type = 19, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Lollipop
lollipop. segment and point
gg_lollipop_YeaNum.
}
\section{ftypes}{
 Yea-Num
}
\examples{
add(1, 1)
add(10, 1)
}

