% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ye-Nu.R
\name{gg_lollipop_YeNu.}
\alias{gg_lollipop_YeNu.}
\title{Lollipop
lollipop. segment and point
gg_lollipop_YeNu.}
\usage{
gg_lollipop_YeNu.(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, angle_x = 0, shape_type = 19, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Lollipop
lollipop. segment and point
gg_lollipop_YeNu.
}
\section{ftypes}{
 Ye-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
