% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca.R
\name{gg_marimekko_hor_CatCat.}
\alias{gg_marimekko_hor_CatCat.}
\title{Horizontal marimekko
Horizontal Marimekko}
\usage{
gg_marimekko_hor_CatCat.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, fillLabel = NULL,
  leg_pos = "right", angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal marimekko
Horizontal Marimekko
}
\section{ftypes}{
 Cat-Cat
}
\examples{
add(1, 1)
add(10, 1)
}

