% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca.R
\name{gg_marimekko_ver_CaCa.}
\alias{gg_marimekko_ver_CaCa.}
\title{gg_marimekko_ver_CaCa.
Vertical Marimekko}
\usage{
gg_marimekko_ver_CaCa.(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, leg_pos = "right", l_ncol = 1, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_marimekko_ver_CaCa.
Vertical Marimekko
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

