% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Nu.R
\name{gg_multi_line_CaCaNu.}
\alias{gg_multi_line_CaCaNu.}
\title{gg_multi_line_CaCaNu.
Grouped Line Coloured}
\usage{
gg_multi_line_CaCaNu.(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, fillLabel = NULL, leg_pos = "right",
  type = 1, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_multi_line_CaCaNu.
Grouped Line Coloured
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

