% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Da-Nu.R
\name{gg_multi_line_CaDaNu.}
\alias{gg_multi_line_CaDaNu.}
\title{gg_multi_line_CaDaNu.
Grouped Line Coloured}
\usage{
gg_multi_line_CaDaNu.(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, fillLabel = NULL, leg_pos = "right",
  type = 1, text_size = 8, angle_x = 45, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_multi_line_CaDaNu.
Grouped Line Coloured
}
\section{ftypes}{
 Ca-Da-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
