% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Nu.R
\name{gg_multi_line_CatCatNum.}
\alias{gg_multi_line_CatCatNum.}
\title{Grouped line by first variable
Grouped Line Coloured}
\usage{
gg_multi_line_CatCatNum.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, fillLabel = NULL,
  aggregation = "sum", leg_pos = "right", angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Grouped line by first variable
Grouped Line Coloured
}
\section{ftypes}{
 Cat-Cat-Num
}

\examples{
add(1, 1)
add(10, 1)
}
