% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{gg_multi_line_CatDatNum.}
\alias{gg_multi_line_CatDatNum.}
\title{Grouped line
Grouped Line Coloured}
\usage{
gg_multi_line_CatDatNum.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, fillLabel = NULL,
  leg_pos = "right", shape_type = 19, angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Grouped line
Grouped Line Coloured
}
\section{ftypes}{
 Cat-Dat-Num
}

\examples{
add(1, 1)
add(10, 1)
}
