% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Nu.R
\name{gg_multi_line_point_CaCaNu.}
\alias{gg_multi_line_point_CaCaNu.}
\title{Grouped line + points by first variable
Grouped Line Color Point}
\usage{
gg_multi_line_point_CaCaNu.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, fillLabel = NULL,
  leg_pos = "right", shape_type = 19, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Grouped line + points by first variable
Grouped Line Color Point
}
\section{ftypes}{
 Ca-Ca-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
