% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_multi_line_point_CaNu.}
\alias{gg_multi_line_point_CaNu.}
\title{gg_multi_line_point_CaNu.
Grouped Line Color Point}
\usage{
gg_multi_line_point_CaNu.(data, titleLabel = "Report", xLabel = "Index",
  yLabel = NULL, fillLabel = NULL, leg_pos = "right", type = 1, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_multi_line_point_CaNu.
Grouped Line Color Point
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

