% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_ordered_bar_ver_CaNu.}
\alias{gg_ordered_bar_ver_CaNu.}
\title{gg_ordered_bar_ver_CaNu.
Ordered vertical Bars}
\usage{
gg_ordered_bar_ver_CaNu.(data, titleLabel = "Report", xLabel = NULL,
  yLabel = NULL, leg_pos = "right")
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_ordered_bar_ver_CaNu.
Ordered vertical Bars
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

