% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca.R
\name{gg_ordered_stacked_bar_hor_CaCa}
\alias{gg_ordered_stacked_bar_hor_CaCa}
\title{gg_ordered_stacked_bar_hor_CaCa
Ordered Stacked horizontal Bar}
\usage{
gg_ordered_stacked_bar_hor_CaCa(data, titleLabel = "Report", xLabel = NULL,
  yLabel = "Count", fillLabel = NULL, leg_pos = "right")
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_ordered_stacked_bar_hor_CaCa
Ordered Stacked horizontal Bar
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

