% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_pie_CaNu.}
\alias{gg_pie_CaNu.}
\title{Pie
Pie}
\usage{
gg_pie_CaNu.(data, titleLabel = "", subtitle = "", caption = "",
  fillLabel = NULL, text = TRUE, type = "count", color_text = "black",
  leg_pos = "right", aggregation = "sum", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Pie
Pie
}
\section{ftypes}{
 Ca-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
