% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_pie_CatNum.R
\name{gg_pie_CatNum}
\alias{gg_pie_CatNum}
\title{Pie (categories, numbers)}
\usage{
gg_pie_CatNum(data = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Ggplot2 visualization
}
\description{
Compare aggregations among category's levels
}
\details{
Color by, does not work, defaults to Cat var
}
\section{ctypes}{

Cat-Num, Dat-Num, Yea-Num
}

\examples{
gg_pie_CatNum(sample_data("Cat-Num", nrow = 10))
}
