% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca.R
\name{gg_pie_facet_CaCa.}
\alias{gg_pie_facet_CaCa.}
\title{gg_pie_facet_CaCa.
Facet Pie}
\usage{
gg_pie_facet_CaCa.(data, titleLabel = "", subtitle = "", caption = "",
  fillLabel = NULL, leg_pos = "right", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_pie_facet_CaCa.
Facet Pie
}
\section{ftypes}{
 Ca-Ca
}

\examples{
add(1, 1)
add(10, 1)
}
