% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Nu.R
\name{gg_pie_facet_CaCaNu.}
\alias{gg_pie_facet_CaCaNu.}
\title{Pie facet by second variable
Facet Pie}
\usage{
gg_pie_facet_CaCaNu.(data, titleLabel = "", subtitle = "", caption = "",
  leg_pos = "right", aggregation = "sum", text = TRUE, type = "count",
  color_text = "black", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Pie facet by second variable
Facet Pie
}
\section{ftypes}{
 Ca-Ca-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
