% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu-Nu.R
\name{gg_point_CaNuNu.}
\alias{gg_point_CaNuNu.}
\title{gg_point_CaNuNu.
point}
\usage{
gg_point_CaNuNu.(data, titleLabel = "", xLabel = NULL, yLabel = NULL,
  leg_pos = "right", shape_point = 1, ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_point_CaNuNu.
point
}
\section{ftypes}{
 Ca-Nu-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

