% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ge-Nu.R
\name{gg_point_co_GeNu.}
\alias{gg_point_co_GeNu.}
\title{gg_point_co_GeNu.
Points inside Colombia's deptos map}
\usage{
gg_point_co_GeNu.(data, titleLabel = "Report", fillLabel = NULL,
  leg_pos = "right")
}
\arguments{
\item{x}{A category.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_point_co_GeNu.
Points inside Colombia's deptos map
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

