% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Da-Nu.R
\name{gg_point_DatNum.}
\alias{gg_point_DatNum.}
\title{Scatter
Point scatter plot}
\usage{
gg_point_DatNum.(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, angle_x = 0, shape_type = 19, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Scatter
Point scatter plot
}
\section{ftypes}{
 Dat-Num
}
\examples{
add(1, 1)
add(10, 1)
}

