% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_point_facet_CaNu.}
\alias{gg_point_facet_CaNu.}
\title{Scatter facet
Facet Point}
\usage{
gg_point_facet_CaNu.(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = "Index", yLabel = NULL, shape_type = 19, angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Scatter facet
Facet Point
}
\section{ftypes}{
 Ca-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
