% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{gg_point_flip_Num.}
\alias{gg_point_flip_Num.}
\title{Horizontal scatter
Line plot}
\usage{
gg_point_flip_Num.(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, shape_type = 19, angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal scatter
Line plot
}
\section{ftypes}{
 Num
}

\examples{
add(1, 1)
add(10, 1)
}
