% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{gg_point_grouped_CatNum.}
\alias{gg_point_grouped_CatNum.}
\title{Grouped scatter
Grouped Color Point}
\usage{
gg_point_grouped_CatNum.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, fillLabel = NULL,
  leg_pos = "right", shape_type = 19, angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Grouped scatter
Grouped Color Point
}
\section{ftypes}{
 Cat-Num
}

\examples{
add(1, 1)
add(10, 1)
}
