% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nu.R
\name{gg_point_Num.}
\alias{gg_point_Num.}
\title{Vertical scatter
Scatter plot}
\usage{
gg_point_Num.(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, shape_type = 19, angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical scatter
Scatter plot
}
\section{ftypes}{
 Num
}
\examples{
add(1, 1)
add(10, 1)
}

