% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_point_trend_line_facet_CatNum.}
\alias{gg_point_trend_line_facet_CatNum.}
\title{Trend line facet
Facet Trend Line}
\usage{
gg_point_trend_line_facet_CatNum.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, shape_type = 19,
  alpha = 0.3, se = FALSE, angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Trend line facet
Facet Trend Line
}
\section{ftypes}{
 Cat-Num
}
\examples{
add(1, 1)
add(10, 1)
}

