% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Da.R
\name{gg_pointline_hor_CaDa.}
\alias{gg_pointline_hor_CaDa.}
\title{gg_pointline_hor_CaDa.: title.
pointlines}
\usage{
gg_pointline_hor_CaDa.(data, title = "", subtitle = "", caption = "",
  xlab = NULL, clab = NULL, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_pointline_hor_CaDa.: title.
pointlines
}
\section{ftypes}{
 Ca-Da
}
\examples{
add(1, 1)
add(10, 1)
}

