% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possible_viz.R
\name{gg_pointline_hor_CatDat.}
\alias{gg_pointline_hor_CatDat.}
\title{Horizontal line + point
pointlines}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal line + point
pointlines
}
\section{ftypes}{
 Cat-Dat
}

\examples{
add(1, 1)
add(10, 1)
}
