% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Da-Nu.R
\name{gg_points_facet_DaNu.}
\alias{gg_points_facet_DaNu.}
\title{gg_points_facet_DaNu.
Points facet by years}
\usage{
gg_points_facet_DaNu.(data, title = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_points_facet_DaNu.
Points facet by years
}
\section{ftypes}{
 Da-Nu, Ye-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

