% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{gg_points_facet_DatNum.}
\alias{gg_points_facet_DatNum.}
\title{Points facet by years
Points facet by years}
\usage{
gg_points_facet_DatNum.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, angle_x = 0,
  shape_type = 19, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Points facet by years
Points facet by years
}
\section{ftypes}{
 Dat-Num
}

\examples{
add(1, 1)
add(10, 1)
}
