% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca.R
\name{gg_polar_bar_Ca.}
\alias{gg_polar_bar_Ca.}
\title{gg_polar_bar_Ca.
Polar Bar}
\usage{
gg_polar_bar_Ca.(data, width = 0.95, titleLabel = "Report",
  fillLabel = NULL, leg_pos = "right")
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_polar_bar_Ca.
Polar Bar
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

