% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-NuP.R
\name{gg_radar_CaNuP.}
\alias{gg_radar_CaNuP.}
\title{gg_radar_CaNuP.
point}
\usage{
gg_radar_CaNuP.(data, titleLabel = "", subtitle = "", caption = "",
  leg_pos = "right", rescale = FALSE, aggregation = "mean", ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_radar_CaNuP.
point
}
\section{ftypes}{
 Ca-Nu-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

