% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu-Nu.R
\name{gg_scatter_agg_trend_CatNumNum.}
\alias{gg_scatter_agg_trend_CatNumNum.}
\title{Scatter + Aggregation on both numeric variables + trend
pointlines}
\usage{
gg_scatter_agg_trend_CatNumNum.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, fillLabel = NULL,
  angle_x = 0, aggregation = "sum", se = FALSE, shape_type = 19,
  leg_pos = "right", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Scatter + Aggregation on both numeric variables + trend
pointlines
}
\section{ftypes}{
 Cat-Num-Num
}

\examples{
add(1, 1)
add(10, 1)
}
