% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{gg_scatter_CatNumNum.}
\alias{gg_scatter_CatNumNum.}
\title{Scatter
Point}
\usage{
gg_scatter_CatNumNum.(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, fillLabel = NULL, leg_pos = "right",
  shape_type = 19, angle_x = 0, ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Scatter
Point
}
\section{ftypes}{
 Cat-Num-Num
}

\examples{
add(1, 1)
add(10, 1)
}
