% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_scatterCatNumNumNum.R
\name{gg_scatter_CatNumNumNum}
\alias{gg_scatter_CatNumNumNum}
\title{Scatter Numeric Numeric}
\usage{
gg_scatter_CatNumNumNum(data, ...)
}
\arguments{
\item{data}{A data.frame}

\item{...}{
  Arguments passed on to \code{\link[dsvizopts:chart_viz_options]{dsvizopts::dsviz_default_opts}}
  \describe{
    \item{\code{}}{}
  }}

\item{orientation}{Doesn't do anything for this type of chart.}

\item{order}{doesn't do anything}
}
\description{
This scatter
}
\section{ctypes}{

Cat-Num-Num-Num
}

\examples{
gg_scatter_CatNumNumNum(sample_data("Cat-Num-Num-Num", nrow = 10))
}
