% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Da-Nu.R
\name{gg_scatter_hor_CaDaNu.}
\alias{gg_scatter_hor_CaDaNu.}
\title{gg_scatter_hor_CaDaNu.
pointlines}
\usage{
gg_scatter_hor_CaDaNu.(data, title = "", subtitle = "", caption = "",
  xlab = NULL, ylab = NULL, clab = NULL, angle = 45, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_scatter_hor_CaDaNu.
pointlines
}
\section{ftypes}{
 Ca-Da-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
