% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Da-Nu.R
\name{gg_scatter_hor_CatDatNum.}
\alias{gg_scatter_hor_CatDatNum.}
\title{Horizontal scatter
pointlines}
\usage{
gg_scatter_hor_CatDatNum.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, fillLabel = NULL,
  angle_x = 0, shape_type = 19, leg_pos = "right", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal scatter
pointlines
}
\section{ftypes}{
 Cat-Dat-Num
}

\examples{
add(1, 1)
add(10, 1)
}
