% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu-Nu.R
\name{gg_scatter_trend_CaNuNu.}
\alias{gg_scatter_trend_CaNuNu.}
\title{gg_scatter_trend_CaNuNu.: title.
pointlines}
\usage{
gg_scatter_trend_CaNuNu.(data, title = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, cLabel = NULL, angle = 45,
  aggregation = "mean", pointLabels = FALSE, vjust = 0, hjust = 0,
  se = FALSE, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_scatter_trend_CaNuNu.: title.
pointlines
}
\section{ftypes}{
 Ca-Da-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
