% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca.R
\name{gg_single_stacked_bar_hor_Ca.}
\alias{gg_single_stacked_bar_hor_Ca.}
\title{gg_single_stacked_bar_hor_Ca.
Single Horizontal Stacked Bar}
\usage{
gg_single_stacked_bar_hor_Ca.(data, titleLabel = "Report", yLabel = "Count",
  fillLabel = "Types", leg_pos = "right", width = 0.3)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_single_stacked_bar_hor_Ca.
Single Horizontal Stacked Bar
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

