% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca.R
\name{gg_single_stacked_bar_ver_Ca.}
\alias{gg_single_stacked_bar_ver_Ca.}
\title{gg_single_stacked_bar_ver_Ca.
Single Vertical Stacked Bar}
\usage{
gg_single_stacked_bar_ver_Ca.(data, titleLabel = "Report",
  yLabel = "Frequency", fillLabel = "Types", leg_pos = "right",
  width = 0.3)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_single_stacked_bar_ver_Ca.
Single Vertical Stacked Bar
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

