% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ge-Nu.R
\name{gg_sketchmap_depto_GeNu.}
\alias{gg_sketchmap_depto_GeNu.}
\title{gg_sketchmap_depto_GeNu.
Sketch map departments}
\usage{
gg_sketchmap_depto_GeNu.(depto_ = depto_, color_map = "gold",
  color_frontier = "black")
}
\arguments{
\item{x}{A category.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_sketchmap_depto_GeNu.
Sketch map departments
}
\section{ftypes}{
 Ge-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
