% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_slope_CaNu.}
\alias{gg_slope_CaNu.}
\title{gg_slope_CaNu.
Slope}
\usage{
gg_slope_CaNu.(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, leg_pos = "right", overlap = TRUE,
  size_text = 6, size_point = 3, size_line = 1, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_slope_CaNu.
Slope
}
\section{ftypes}{
 Ca-Ye-Nu,Ca-Nu-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

