% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possible_viz.R
\name{gg_slope_CatNum.}
\alias{gg_slope_CatNum.}
\title{Slope
Slope}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Slope
Slope
}
\section{ftypes}{
 Cat-Num
}

\examples{
add(1, 1)
add(10, 1)
}
