% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ye-Nu.R
\name{gg_slope_CaYeNu.}
\alias{gg_slope_CaYeNu.}
\title{Slope
Slope}
\usage{
gg_slope_CaYeNu.(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, leg_pos = "right", text_size = 6,
  size_vjust = 1.5, overlap = TRUE, size_hjust = 0.5, size_point = 3,
  size_line = 1, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Slope
Slope
}
\section{ftypes}{
 Ca-Ye-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
