% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possible_viz.R
\name{gg_smooth_DatNum.}
\alias{gg_smooth_DatNum.}
\title{Smooth
smooth}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Smooth
smooth
}
\section{ftypes}{
 Dat-Num
}

\examples{
add(1, 1)
add(10, 1)
}
