% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_stacked_area_100_ver_CaNu.}
\alias{gg_stacked_area_100_ver_CaNu.}
\title{gg_stacked_area_100_ver_CaNu.
Stacked Vertical Area 100%}
\usage{
gg_stacked_area_100_ver_CaNu.(data, titleLabel = "Report", xLabel = "Index",
  yLabel = NULL, fillLabel = NULL, leg_pos = "top")
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_stacked_area_100_ver_CaNu.
Stacked Vertical Area 100%
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

