% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Nu.R
\name{gg_stacked_bar_100_ver_CaCaNu.}
\alias{gg_stacked_bar_100_ver_CaCaNu.}
\title{gg_stacked_bar_100_ver_CaCaNu.
100% vertical stacked bar graph}
\usage{
gg_stacked_bar_100_ver_CaCaNu.(data, titleLabel = "Report", xLabel = NULL,
  yLabel = NULL, fillLabel = NULL, leg_pos = "top")
}
\arguments{
\item{x}{A category.}

\item{y}{A category.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_stacked_bar_100_ver_CaCaNu.
100% vertical stacked bar graph
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

