% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca.R
\name{gg_stacked_dot_bar_hor_CaCa.}
\alias{gg_stacked_dot_bar_hor_CaCa.}
\title{Horizontal stacked dot bar
Stacked horizontal dot Bar}
\usage{
gg_stacked_dot_bar_hor_CaCa.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = "Count", leg_pos = "right",
  angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal stacked dot bar
Stacked horizontal dot Bar
}
\section{ftypes}{
 Ca-Ca
}

\examples{
add(1, 1)
add(10, 1)
}
