% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca.R
\name{gg_stacked_polar_bar_100_CaCa.}
\alias{gg_stacked_polar_bar_100_CaCa.}
\title{gg_stacked_polar_bar_100_CaCa.
Stacked Polar  100%}
\usage{
gg_stacked_polar_bar_100_CaCa.(data, titleLabel = "Report", xLabel = NULL,
  yLabel = "Count", fillLabel = NULL, leg_pos = "top", width = 1)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_stacked_polar_bar_100_CaCa.
Stacked Polar  100%
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

