% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Nu.R
\name{gg_stacked_polar_bar_100_CaCaNu.}
\alias{gg_stacked_polar_bar_100_CaCaNu.}
\title{gg_stacked_polar_bar_100_CaCaNu.
Stacked Polar Bar 100%}
\usage{
gg_stacked_polar_bar_100_CaCaNu.(data, width = 0.95, titleLabel = "Report",
  fillLabel = NULL, leg_pos = "right")
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_stacked_polar_bar_100_CaCaNu.
Stacked Polar Bar 100%
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

