% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Da-Nu.R
\name{gg_steam_CaDaNu.}
\alias{gg_steam_CaDaNu.}
\title{Steam
Steam}
\usage{
gg_steam_CaDaNu.(data, title = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, leg_pos = "right", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Steam
Steam
}
\section{ftypes}{
 Ca-Da-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
