% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steam.R
\name{gg_steam_CatNum.}
\alias{gg_steam_CatNum.}
\title{Steam
Steam}
\usage{
gg_steam_CatNum.(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, fillLabel = NULL,
  leg_pos = "right", angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Steam
Steam
}
\section{ftypes}{
 Cat-Num
}

\examples{
add(1, 1)
add(10, 1)
}
