% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steam.R
\name{gg_steam_CatNumNum.}
\alias{gg_steam_CatNumNum.}
\title{Steam
Steamgraph}
\usage{
gg_steam_CatNumNum.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, fillLabel = NULL,
  leg_pos = "right", angle_x = 0, aggregation = "sum", ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Steam
Steamgraph
}
\section{ftypes}{
 Cat-Num-Num
}

\examples{
add(1, 1)
add(10, 1)
}
