% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu-Nu.R
\name{gg_steamgraph_CaNuNu.}
\alias{gg_steamgraph_CaNuNu.}
\title{gg_steamgraph_CaNuNu.
steamgraph}
\usage{
gg_steamgraph_CaNuNu.(data, titleLabel = "", fillLabel = NULL,
  leg_pos = "right", ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_steamgraph_CaNuNu.
steamgraph
}
\section{ftypes}{
 Ca-Nu-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

