% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Da-Nu.R
\name{gg_stepped_stacked_area_DaNu.}
\alias{gg_area_stepped_stacked_DaNu.}
\alias{gg_stepped_stacked_area_DaNu.}
\title{gg_stepped_stacked_area_DaNu.
stepped stacked area.}
\usage{
gg_area_stepped_stacked_DaNu.(data, title = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, angle_x = 0, ...)
}
\arguments{
\item{x}{A data.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_stepped_stacked_area_DaNu.
stepped stacked area.
}
\section{ftypes}{
 Da-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

