% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Da-Nu.R
\name{gg_stream_CaDaNu.}
\alias{gg_stream_CaDaNu.}
\title{gg_stream_CaDaNu.
Stream}
\usage{
gg_stream_CaDaNu.(data, titleLabel = "Report", xLabel = NULL,
  yLabel = NULL, fillLabel = NULL, leg_pos = "right")
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_stream_CaDaNu.
Stream
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

