% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_stream_CaNu.}
\alias{gg_stream_CaNu.}
\title{gg_stream_CaNu.
Stream}
\usage{
gg_stream_CaNu.(data, titleLabel = "", xLabel = "Index", yLabel = NULL,
  fillLabel = NULL, leg_pos = "right", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_stream_CaNu.
Stream
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

