% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu-Nu.R
\name{gg_stream_CaNuNu.}
\alias{gg_stream_CaNuNu.}
\title{gg_stream_CaNuNu.
Streamgraph}
\usage{
gg_stream_CaNuNu.(data, titleLabel = "", fillLabel = NULL,
  leg_pos = "right", ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_stream_CaNuNu.
Streamgraph
}
\section{ftypes}{
 Ca-Nu-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

