% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ye-Nu.R
\name{gg_stream_CaYeNu.}
\alias{gg_stream_CaYeNu.}
\title{gg_stream_CaYeNu.
Stream}
\usage{
gg_stream_CaYeNu.(data, titleLabel = "", clab = NULL, leg_pos = "right",
  ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_stream_CaYeNu.
Stream
}
\section{ftypes}{
 Ca-Ye-Nu,Ca-Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

