% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Ca-Nu.R
\name{gg_sunburst_CatCatCatNum.}
\alias{gg_sunburst_CatCatCatNum.}
\title{Sunburst
sunburst}
\usage{
gg_sunburst_CatCatCatNum.(data, titleLabel = "", subtitle = "",
  caption = "", aggregation = "sum", fillLabel = NULL, xLabel = NULL,
  yLabel = NULL, leg_pos = "right", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Sunburst
sunburst
}
\section{ftypes}{
 Cat-Cat-Cat-Num
}

\examples{
add(1, 1)
add(10, 1)
}
