% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunburst.R
\name{gg_sunburst_CatCatNum.}
\alias{gg_sunburst_CatCatNum.}
\title{Sunburst
sunburst}
\usage{
gg_sunburst_CatCatNum.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, fillLabel = NULL,
  aggregation = "sum", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Sunburst
sunburst
}
\section{ftypes}{
 Cat-Cat-Num
}

\examples{
add(1, 1)
add(10, 1)
}
