% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca.R
\name{gg_treemap_Ca.}
\alias{gg_treemap_Ca.}
\title{Treemap coloured by first variable
Treemap fill by first Ca}
\usage{
gg_treemap_Ca.(data, titleLabel = "", subtitle = "", caption = "",
  text = TRUE, color_text = "black", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Treemap coloured by first variable
Treemap fill by first Ca
}
\section{ftypes}{
 Ca
}

\examples{
add(1, 1)
add(10, 1)
}
