% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap.R
\name{gg_treemap_Cat}
\alias{gg_treemap_Cat}
\title{Treemap (categories)}
\usage{
gg_treemap_Cat(
  data = NULL,
  agg_text = NULL,
  caption = NULL,
  colors = NULL,
  color_scale = "discrete",
  drop_na = FALSE,
  highlight_value = NULL,
  highlight_value_color = "#F9B233",
  label_wrap = 12,
  marks = c(".", ","),
  n_digits = NULL,
  order = NULL,
  percentage = FALSE,
  prefix = NULL,
  slice_n = NULL,
  sort = "no",
  subtitle = NULL,
  suffix = NULL,
  text_color = "#5A6B72",
  text_show = TRUE,
  text_size = 3,
  theme = NULL,
  title = NULL,
  opts = NULL,
  ...
)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare aggregations among category's levels
}
\section{ctypes}{

Cat, Yea, Dat
}

\examples{
gg_treemap_Cat(sampleData("Cat", nrow = 10))
}
