% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_treemap_CatCatNum.R
\name{gg_treemap_CatCat}
\alias{gg_treemap_CatCat}
\title{treemap Chart Cat Cat}
\usage{
gg_treemap_CatCat(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
This chart does not allow for chaning orientation
}
\examples{
gg_treemap_CatCat(sample_data("Cat-Cat", nrow = 10))
}
