% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap.R
\name{gg_treemap_CatCatNum}
\alias{gg_treemap_CatCatNum}
\title{Treemap (categories, categories, numbers)}
\usage{
gg_treemap_CatCatNum(data, title = NULL, subtitle = NULL,
  caption = NULL, agg = "sum", colors = NULL,
  colorScale = "discrete", dropNaV = c(FALSE, FALSE), format = c("",
  ""), labelWrapV = c(12, 12), marks = c(".", ","), nDigits = NULL,
  percentage = FALSE, showText = TRUE, showLegend = TRUE,
  legendPosition = c("right", "bottom"), theme = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Ggplot visualization
}
\description{
Compare aggregations among category's levels
}
\section{ctypes}{

Cat-Cat-Num, Cat-Yea-Num, Cat-Dat-Num,
}

\examples{
gg_treemap_CatNum(sampleData("Cat-Cat-Num", nrow = 10))
}
