% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap.R
\name{gg_treemap_CatNumP}
\alias{gg_treemap_CatNumP}
\title{Treemap (categories, categories)}
\usage{
gg_treemap_CatNumP(data = NULL, opts = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare aggregations among category's levels
}
\section{ctypes}{

Cat-NumP, Yea-NumP, Dat-NumP
}

\examples{
gg_treemap_CatNumP(sampleData("Cat-NumP", nrow = 10))
}
