% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_treemap_CatYeaNum.R
\name{gg_treemap_CatYea}
\alias{gg_treemap_CatYea}
\title{treemap Chart Cat Yea}
\usage{
gg_treemap_CatYea(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
This chart does not allow for chaning orientation
}
\examples{
gg_treemap_CatYea(sample_data("Cat-Yea", nrow = 10))
}
