% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_treemap_CatYeaNum.R
\name{gg_treemap_CatYeaNum}
\alias{gg_treemap_CatYeaNum}
\title{treemap Chart Categories Yea Numeric}
\usage{
gg_treemap_CatYeaNum(data, ...)
}
\arguments{
\item{data}{A data.frame}

\item{...}{
  Arguments passed on to \code{\link[dsvizopts:chart_viz_options]{dsvizopts::dsviz_default_opts}}
  \describe{
    \item{\code{}}{}
  }}

\item{orientation}{Doesn't do anything for this type of chart.}

\item{order}{doesn't do anything}
}
\description{
This chart does not allow for chaning orientation
}
\section{ctypes}{

Cat-Yea-Num, Cat-Yea
}

\examples{
gg_treemap_CatYeaNum(sample_data("Cat-Yea-Num", nrow = 10))
}
