% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Nu-Nu.R
\name{gg_treemap_density_x_CatCatNumNum.}
\alias{gg_treemap_density_x_CatCatNumNum.}
\title{Treemap density first numeric variable
Treemap Fill by second Num}
\usage{
gg_treemap_density_x_CatCatNumNum.(data, titleLabel = "", subtitle = "",
  caption = "", fillLabel = NULL, reverse = FALSE, text = TRUE,
  color_text = "black", aggregation = "sum", leg_pos = "right", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Treemap density first numeric variable
Treemap Fill by second Num
}
\section{ftypes}{
 Cat-Cat-Num-Num
}
\examples{
add(1, 1)
add(10, 1)
}

