% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Nu-Nu.R
\name{gg_treemap_density_y_CaCaNuNu.}
\alias{gg_treemap_density_y_CaCaNuNu.}
\title{Treemap density second numeric variable
Treemap Density by second Nu}
\usage{
gg_treemap_density_y_CaCaNuNu.(data, titleLabel = "", subtitle = "",
  caption = "", fillLabel = NULL, reverse = FALSE, text = TRUE,
  color_text = "black", aggregation = "sum", leg_pos = "right", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Treemap density second numeric variable
Treemap Density by second Nu
}
\section{ftypes}{
 Ca-Ca-Nu-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
