% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_treemap_density_y_CatNum.}
\alias{gg_treemap_density_y_CatNum.}
\title{Treemap density by first numeric variable
Treemap Density by Num}
\usage{
gg_treemap_density_y_CatNum.(data, titleLabel = "", subtitle = "",
  caption = "", fillLabel = NULL, reverse = FALSE, text = TRUE,
  color_text = "black", aggregation = "sum", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Treemap density by first numeric variable
Treemap Density by Num
}
\section{ftypes}{
 Cat-Num
}
\examples{
add(1, 1)
add(10, 1)
}

