% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap.R
\name{gg_treemap_density_z_CatCatNum.}
\alias{gg_treemap_density_z_CatCatNum.}
\title{Treemap density by numeric variable
Treemap Density by Num}
\usage{
gg_treemap_density_z_CatCatNum.(data, titleLabel = "", subtitle = "",
  caption = "", reverse = FALSE, fillLabel = NULL, text = TRUE,
  color_text = "black", aggregation = "sum", leg_pos = "right", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Treemap density by numeric variable
Treemap Density by Num
}
\section{ftypes}{
 Cat-Cat-Num
}

\examples{
add(1, 1)
add(10, 1)
}
