% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Ca.R
\name{gg_treemap_x_CaCaCa.}
\alias{gg_treemap_x_CaCaCa.}
\title{gg_treemap_x_CaCaCa.
Treemap fill first Ca}
\usage{
gg_treemap_x_CaCaCa.(data, titleLabel = "Report", fillLabel = NULL, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_treemap_x_CaCaCa.
Treemap fill first Ca
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

