% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Ca-Nu.R
\name{gg_treemap_x_CaCaCaNu.}
\alias{gg_treemap_x_CaCaCaNu.}
\title{gg_treemap_x_CaCaCaNu.
Treemap fill first Ca}
\usage{
gg_treemap_x_CaCaCaNu.(data, titleLabel = "", subtitle = "", caption = "",
  fillLabel = NULL, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_treemap_x_CaCaCaNu.
Treemap fill first Ca
}
\section{ftypes}{
 Ca-Ca-Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

