% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap.R
\name{gg_treemap_x_CatCatCatNum.}
\alias{gg_treemap_x_CatCatCatNum.}
\title{Treemap coloured by first variable
Treemap fill first Cat}
\usage{
gg_treemap_x_CatCatCatNum.(data, titleLabel = "", subtitle = "",
  caption = "", fillLabel = NULL, aggregation = "sum",
  text = "TRUE", color_text = "black", leg_pos = "right", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Treemap coloured by first variable
Treemap fill first Cat
}
\section{ftypes}{
 Cat-Cat-Cat-Num
}

\examples{
add(1, 1)
add(10, 1)
}
