% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_treemap_x_CatNum.}
\alias{gg_treemap_x_CatNum.}
\title{Treemap coloured by first variable
Treemap fill by first Cat}
\usage{
gg_treemap_x_CatNum.(data, titleLabel = "", subtitle = "", caption = "",
  text = TRUE, color_text = "black", aggregation = "sum", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Treemap coloured by first variable
Treemap fill by first Cat
}
\section{ftypes}{
 Cat-Num
}

\examples{
add(1, 1)
add(10, 1)
}
