% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Ca.R
\name{gg_treemap_z_CaCaCa.}
\alias{gg_treemap_z_CaCaCa.}
\title{gg_treemap_z_CaCaCa.
Treemap fill third Ca}
\usage{
gg_treemap_z_CaCaCa.(data, titleLabel = "", subtitle = "", caption = "",
  fillLabel = NULL, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_treemap_z_CaCaCa.
Treemap fill third Ca
}
\section{ftypes}{
 Ca-Ca-Ca
}
\examples{
add(1, 1)
add(10, 1)
}

