% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Ca.R
\name{gg_treemap_z_CatCatCat.}
\alias{gg_treemap_z_CatCatCat.}
\title{Treemap coloured by third variable
Treemap fill third Cat}
\usage{
gg_treemap_z_CatCatCat.(data, titleLabel = "", subtitle = "",
  caption = "", fillLabel = NULL, text = "TRUE", color_text = "black",
  leg_pos = "right", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Treemap coloured by third variable
Treemap fill third Cat
}
\section{ftypes}{
 Cat-Cat-Cat
}

\examples{
add(1, 1)
add(10, 1)
}
