% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_trend_ribbon_facet_CatNum.}
\alias{gg_trend_ribbon_facet_CatNum.}
\title{Trend ribbon facet
Facet Trend ribbon}
\usage{
gg_trend_ribbon_facet_CatNum.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, shape_type = 19,
  alpha = 0.3, angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Trend ribbon facet
Facet Trend ribbon
}
\section{ftypes}{
 Cat-Num
}

\examples{
add(1, 1)
add(10, 1)
}
