% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Da-Nu.R
\name{gg_violin_DaNu.}
\alias{gg_violin_DaNu.}
\title{gg_violin_DaNu.
Violin}
\usage{
gg_violin_DaNu.(data, title = "", subtitle = "", caption = "",
  xlab = NULL, ylab = NULL, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_violin_DaNu.
Violin
}
\section{ftypes}{
 Da-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

