% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_violin_dot_mult_flip_CatNum.}
\alias{gg_violin_dot_mult_flip_CatNum.}
\title{Horizontal violin + jitter
Violin + dot jitter flipped}
\usage{
gg_violin_dot_mult_flip_CatNum.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, leg_pos = "right",
  angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal violin + jitter
Violin + dot jitter flipped
}
\section{ftypes}{
 Cat-Num
}
\examples{
add(1, 1)
add(10, 1)
}

