% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nu.R
\name{gg_violin_flip_Num.}
\alias{gg_violin_flip_Num.}
\title{Vertical violin
Violin flipped}
\usage{
gg_violin_flip_Num.(data, titleLabel = "", subtitle = "", caption = "",
  yLabel = NULL, xLabel = NULL, angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical violin
Violin flipped
}
\section{ftypes}{
 Num
}
\examples{
add(1, 1)
add(10, 1)
}

