% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violin.R
\name{gg_violin_mult_CatNum.}
\alias{gg_violin_mult_CatNum.}
\title{Vertical violin
Violin}
\usage{
gg_violin_mult_CatNum.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, leg_pos = "right",
  angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical violin
Violin
}
\section{ftypes}{
 Cat-Num
}

\examples{
add(1, 1)
add(10, 1)
}
