% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nu.R
\name{gg_violin_Nu.}
\alias{gg_violin_Nu.}
\title{Horizontal violin
Violin}
\usage{
gg_violin_Nu.(data, title = "", subtitle = "", caption = "",
  yLabel = NULL, xLabel = NULL, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal violin
Violin
}
\section{ftypes}{
 Nu
}

\examples{
add(1, 1)
add(10, 1)
}
