% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca_funs.R
\name{gg_waffle.}
\alias{gg_waffle.}
\title{gg_waffle.: title.
Hola esta es la descripción
Tiene múltiples líneas}
\usage{
gg_waffle.(data, filas = 5, titulo = "Report")
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_waffle.: title.
Hola esta es la descripción
Tiene múltiples líneas
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

