% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca.R
\name{gg_waffle_Ca.}
\alias{gg_waffle_Ca.}
\title{gg_waffle_Ca.: title.
Waffle}
\usage{
gg_waffle_Ca.(data, filas = 5, titulo = "Report")
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_waffle_Ca.: title.
Waffle
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

