% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nu.R
\name{gg_waterfall_Nu.}
\alias{gg_waterfall_Nu.}
\title{gg_waterfall_Nu.
Waterfall}
\usage{
gg_waterfall_Nu.(data, titleLabel = "Report", xLabel = "Index",
  yLabel = NULL, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_waterfall_Nu.
Waterfall
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

