% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ye-Nu.R
\name{gg_waterfall_YeNu.}
\alias{gg_waterfall_YeNu.}
\title{Waterfall
Waterfall}
\usage{
gg_waterfall_YeNu.(data, title = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Waterfall
Waterfall
}
\section{ftypes}{
 Ye-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
