% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmagic-package.R
\docType{package}
\name{ggmagic-package}
\alias{ggmagic}
\alias{ggmagic-package}
\title{ggmagic: Wrapper for Multiple Plots using ggplot.}
\description{
Plot recommendation with tidy data fringes.
}
\keyword{internal}
