% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Nu.R
\name{horizontal_dotgraphCCN}
\alias{horizontal_dotgraphCCN}
\title{horizontal_dotgraphCCN
horizontal dot graph}
\usage{
horizontal_dotgraphCCN(data, titleLabel = "Report", xLabel = "Categories",
  yLabel = "Frequency", fillLabel = "Types", leg_pos = "top", ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A category.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
horizontal_dotgraphCCN
horizontal dot graph
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

