% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Nu.R
\name{horizontal_linegraphCCN}
\alias{horizontal_linegraphCCN}
\title{horizontal_linegraphCCN
horizontal linegraph}
\usage{
horizontal_linegraphCCN(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = "Types", yLabel = "Frequency", ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A category.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
horizontal_linegraphCCN
horizontal linegraph
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

