% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Nu.R
\name{horizontal_stacked_bargraphCCN}
\alias{horizontal_stacked_bargraphCCN}
\title{horizontal_stacked_bargraphCCN
horizontal stacked bar graph}
\usage{
horizontal_stacked_bargraphCCN(data, titleLabel = "Report",
  xLabel = "Category", yLabel = "Frequency", fillLabel = "Types",
  leg_pos = "top")
}
\arguments{
\item{x}{A category.}

\item{y}{A category.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
horizontal_stacked_bargraphCCN
horizontal stacked bar graph
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

