% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Nu.R
\name{horizontal_unstacked_bargraphCCN}
\alias{horizontal_unstacked_bargraphCCN}
\title{horizontal_unstacked_bargraphCCN
horizontal unstacked bargraph}
\usage{
horizontal_unstacked_bargraphCCN(data, titleLabel = "", xLabel = "Category",
  yLabel = "Frequency", fillLabel = "Types", leg_pos = "right", ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A category.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
horizontal_unstacked_bargraphCCN
horizontal unstacked bargraph
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

