% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Nu.R
\name{ordered_horizontal_bargraphCCN}
\alias{ordered_horizontal_bargraphCCN}
\title{ordered_horizontal_bargraphCCN
ordered horizontal bar graph}
\usage{
ordered_horizontal_bargraphCCN(data, titleLabel = "", xLabel = "Frequency",
  yLabel = "Categories", fillLabel = "Types", leg_pos = "right", ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A category.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
ordered_horizontal_bargraphCCN
ordered horizontal bar graph
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

