% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Nu.R
\name{ordered_vertical_bargraphCCN}
\alias{ordered_vertical_bargraphCCN}
\title{ordered_vertical_bargraphCCN
ordered vertical bar graph}
\usage{
ordered_vertical_bargraphCCN(data, titleLabel = "Report",
  xLabel = "Frequency", yLabel = "Categories", fillLabel = "Types",
  leg_pos = "right", ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A category.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
ordered_vertical_bargraphCCN
ordered vertical bar graph
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

