% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Nu.R
\name{vertical_dotgraphCCN}
\alias{vertical_dotgraphCCN}
\title{vertical_dotgraphCCN
vertical dot graph}
\usage{
vertical_dotgraphCCN(data, titleLabel = "Report", xLabel = "Categories",
  yLabel = "Frequency", fillLabel = "Types", leg_pos = "right", ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A category.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
vertical_dotgraphCCN
vertical dot graph
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

