% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Nu.R
\name{vertical_linegraphCCN}
\alias{vertical_linegraphCCN}
\title{vertical_linegraphCCN
vertical linegraph}
\usage{
vertical_linegraphCCN(data, titleLabel = "Report", xLabel = "Types",
  yLabel = "Frequency", ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A category.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
vertical_linegraphCCN
vertical linegraph
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

