% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Nu.R
\name{vertical_stacked_bargraphCCN}
\alias{vertical_stacked_bargraphCCN}
\title{vertical_stacked_bargraphCCN
vertical stacked bar graph}
\usage{
vertical_stacked_bargraphCCN(data, titleLabel = "Report",
  xLabel = "Category", yLabel = "Frequency", fillLabel = "Types",
  leg_pos = "top")
}
\arguments{
\item{x}{A category.}

\item{y}{A category.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
vertical_stacked_bargraphCCN
vertical stacked bar graph
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

