% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Nu.R
\name{vertical_unstacked_bargraphCCN}
\alias{vertical_unstacked_bargraphCCN}
\title{vertical_unstacked_bargraphCCN
vertical unstacked bargraph}
\usage{
vertical_unstacked_bargraphCCN(data, titleLabel = "", xLabel = "Category",
  yLabel = "Frequency", fillLabel = "Types", leg_pos = "top", ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A category.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
vertical_unstacked_bargraphCCN
vertical unstacked bargraph
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

