% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{gg_area_Cat}
\alias{gg_area_Cat}
\title{Area (categories)}
\usage{
gg_area_Cat(data, title = NULL, subtitle = NULL, caption = NULL,
  horLabel = NULL, verLabel = NULL, horLine = NULL, verLine = NULL,
  colors = NULL, colorText = "#5A6B72", colorOpacity = 0.6,
  colorScale = "no", dropNa = FALSE, prefix = NULL, suffix = NULL,
  labelRatio = 1, labelWrap = 12, marks = c(".", ","), nDigits = 0,
  order = NULL, orientation = "ver", percentage = FALSE,
  shapeType = 19, sort = "no", sliceN = NULL, showText = TRUE,
  spline = FALSE, startAtZero = TRUE, theme = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Ggplot2 visualization
}
\description{
Compare category's levels
}
\section{ctypes}{

Cat
}

\examples{
gg_area_Cat(sampleData("Cat", nrow = 10))
}
