% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{gg_area_CatCat}
\alias{gg_area_CatCat}
\title{Area (categories, ordered categories)}
\usage{
gg_area_CatCat(data, title = NULL, subtitle = NULL, caption = NULL,
  horLabel = NULL, verLabel = NULL, horLine = NULL, verLine = NULL,
  colors = NULL, colorText = "black", colorOpacity = 0.6,
  colorScale = "discrete", dropNaV = c(FALSE, FALSE), format = c("",
  ""), labelRatio = 1.1, labelWrapV = c(12, 12),
  legendPosition = "right", legendTitle = NULL, marks = c(".", ","),
  nDigits = 0, order1 = NULL, order2 = NULL, orientation = "ver",
  percentage = FALSE, shapeType = 19, showText = TRUE,
  spline = FALSE, startAtZero = TRUE, theme = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Ggplot2 visualization
}
\description{
Compare quantities among two categories
}
\section{ctypes}{

Cat-Cat, Cat-Dat, Cat-Yea, Yea-Cat, Yea-Dat, Yea-Yea, Dat-Cat, Dat-Yea, Dat-Dat
}

\examples{
gg_area_CatCat(sampleData("Cat-Cat", nrow = 10))
}
