% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{gg_area_CatCat}
\alias{gg_area_CatCat}
\title{Area (categories, ordered categories)}
\usage{
gg_area_CatCat(
  data = NULL,
  agg_text = NULL,
  caption = NULL,
  colors = NULL,
  fill_opacity = "0.7",
  color_scale = "discrete",
  drop_na = FALSE,
  drop_na_legend = FALSE,
  graph_type = "grouped",
  highlight_value = NULL,
  highlight_value_color = "#F9B233",
  hor_label = NULL,
  hor_line = NULL,
  hor_line_label = NULL,
  label_ratio = 1,
  label_wrap = 12,
  label_wrap_legend = 12,
  legend_position = "bottom",
  legend_show = TRUE,
  legend_title = NULL,
  marks = c(".", ","),
  n_digits = NULL,
  order1 = NULL,
  order2 = NULL,
  orientation = "ver",
  percentage = FALSE,
  prefix = NULL,
  shape_type = 19,
  slice_n = NULL,
  sort = "no",
  spline = FALSE,
  start_zero = TRUE,
  subtitle = NULL,
  suffix = NULL,
  text_color = "#212428",
  text_show = TRUE,
  text_size = 3,
  theme = NULL,
  title = NULL,
  ver_label = NULL,
  ver_line = NULL,
  opts = NULL,
  ...
)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Ggplot2 visualization
}
\description{
Compare quantities among two categories
}
\section{ctypes}{

Cat-Cat, Cat-Dat, Cat-Yea, Yea-Cat, Yea-Dat, Yea-Yea, Dat-Cat, Dat-Yea, Dat-Dat
}

\examples{
gg_area_CatCat(sampleData("Cat-Cat", nrow = 10))
}
