% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{gg_area_CatCatNum}
\alias{gg_area_CatCatNum}
\title{Area (categories, ordered categories, numbers)}
\usage{
gg_area_CatCatNum(data, title = NULL, subtitle = NULL,
  caption = NULL, horLabel = NULL, verLabel = NULL, horLine = NULL,
  verLine = NULL, agg = "sum", agg_text = NULL, colors = NULL,
  colorText = "#5A6B72", colorOpacity = 0.7, colorScale = "discrete",
  dropNaV = c(FALSE, FALSE), prefix = NULL, suffix = NULL,
  graphType = "grouped", labelRatio = 1.1, labelWrapV = c(12, 12),
  legendPosition = "right", legendTitle = NULL, marks = c(".", ","),
  nDigits = 0, order1 = NULL, order2 = NULL, orientation = "ver",
  percentage = FALSE, shapeType = 19, showText = TRUE,
  spline = FALSE, startAtZero = TRUE, theme = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Ggplot2 visualization
}
\description{
Compare quantities among two categories
}
\section{ctypes}{

Cat-Cat-Num, Cat-Dat-Num, Cat-Yea-Num, Yea-Cat-Num, Yea-Dat-Num, Yea-Yea-Num, Dat-Cat-Num, Dat-Yea-Num, Dat-Dat-Num
}

\examples{
gg_area_CatCatNum(sampleData("Cat-Cat-Num", nrow = 10))
}
