% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{gg_area_CatDatNum}
\alias{gg_area_CatDatNum}
\title{Area (dates, numbers)}
\usage{
gg_area_CatDatNum(data, title = NULL, subtitle = NULL,
  caption = NULL, horLabel = NULL, verLabel = NULL, horLine = NULL,
  verLine = NULL, agg = "sum", colors = NULL, colorText = "black",
  colorOpacity = 0.6, colorScale = "discrete", dropNaV = c(FALSE,
  FALSE), format = c("", ""), labelRatio = 1.1, labelWrapV = c(12,
  12), legendPosition = "right", legendTitle = NULL, marks = c(".",
  ","), nDigits = 0, order1 = NULL, order2 = NULL,
  orientation = "ver", percentage = FALSE, shapeType = 19,
  showText = TRUE, spline = FALSE, startAtZero = TRUE,
  theme = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Ggplot2 visualization
}
\description{
Compare quantities among categories over dates
}
\section{ctypes}{

Cat-Dat-Num
}

\examples{
gg_area_CatDatNum(sampleData("Cat-Dat-Num", nrow = 10))
}
