% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{gg_area_CatNum}
\alias{gg_area_CatNum}
\title{Area (categories, numbers)}
\usage{
gg_area_CatNum(
  data = NULL,
  agg = "sum",
  agg_text = NULL,
  background = "transparent",
  caption = NULL,
  colors = NULL,
  fill_opacity = 0.7,
  drop_na = FALSE,
  hor_label = NULL,
  hor_line = NULL,
  label_ratio = 1,
  label_wrap = 12,
  marks = c(".", ","),
  n_digits = NULL,
  order = NULL,
  orientation = "ver",
  percentage = FALSE,
  prefix = NULL,
  shape_size = 3,
  shape_type = 19,
  slice_n = NULL,
  sort = "no",
  start_zero = TRUE,
  subtitle = NULL,
  suffix = NULL,
  text_color = "#5A6B72",
  text_show = TRUE,
  text_size = 3,
  theme = NULL,
  title = NULL,
  ver_label = NULL,
  ver_line = NULL,
  opts = NULL,
  ...
)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Ggplot visualization
}
\description{
Compare aggregations among category's levels
}
\section{ctypes}{

Cat-Num
}

\examples{
gg_area_CatNum(sampleData("Cat-Num", nrow = 10))
}
