% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{gg_area_DatNum}
\alias{gg_area_DatNum}
\title{Vertical area (dates, numbers)}
\usage{
gg_area_DatNum(data, title = NULL, subtitle = NULL, caption = NULL,
  horLabel = NULL, verLabel = NULL, yLine = NULL, yLineLabel = NULL,
  agg = "sum", dropNa = FALSE, order = NULL, percentage = FALSE,
  theme = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Ggplot visualization
}
\description{
Compare quantities over dates
}
\section{ctypes}{

Dat-Num
}

\examples{
gg_area_DatNum(sampleData("Dat-Num", nrow = 10))
}
