% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_area_DatNum.R
\name{gg_area_DatNum}
\alias{gg_area_DatNum}
\title{Area Chart Datr Numeric}
\usage{
gg_area_DatNum(data, ...)
}
\arguments{
\item{data}{A data.frame}

\item{...}{
  Arguments passed on to \code{\link[dsvizopts:chart_viz_options]{dsvizopts::dsviz_default_opts}}
  \describe{
    \item{\code{}}{}
  }}

\item{orientation}{Doesn't do anything for this type of chart.}

\item{order}{doesn't do anything}
}
\description{
This chart does not allow for chaning orientation
}
\section{ctypes}{

Dat-Num, Dat-Num
}

\examples{
gg_area_DatNum(sample_data("Cat-Num", nrow = 10))
}
