% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Nu.R
\name{gg_area_stacked_100_ver_CaCaNu.}
\alias{gg_area_stacked_100_ver_CaCaNu.}
\title{gg_area_stacked_100_ver_CaCaNu.
Stacked area 100}
\usage{
gg_area_stacked_100_ver_CaCaNu.(data, titleLabel = "", xLabel = NULL,
  yLabel = NULL, leg_pos = "right", ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A category.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_area_stacked_100_ver_CaCaNu.
Stacked area 100
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

