% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Da-Nu.R
\name{gg_area_stacked_100_ver_CaDaNu.}
\alias{gg_area_stacked_100_ver_CaDaNu.}
\title{gg_area_stacked_100_ver_CaDaNu.
Stacked Vertical Area 100}
\usage{
gg_area_stacked_100_ver_CaDaNu.(data, titleLabel = "Report", xLabel = NULL,
  yLabel = NULL, fillLabel = NULL, leg_pos = "top", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_area_stacked_100_ver_CaDaNu.
Stacked Vertical Area 100
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

