% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{gg_area_YeaNum}
\alias{gg_area_YeaNum}
\title{Vertical area
Area}
\usage{
gg_area_YeaNum(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical area
Area
}
\section{ftypes}{
 Yea-Num
}

\examples{
gg_area_YeaNum(sampleData("Yea-Num"))
}
