% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{gg_area_YeaNum}
\alias{gg_area_YeaNum}
\title{Vertical area (years, numbers)}
\usage{
gg_area_YeaNum(data, title = NULL, subtitle = NULL, caption = NULL,
  horLabel = NULL, verLabel = NULL, yLine = NULL,
  yLineLabel = NULL, agg = "sum", dropNa = FALSE, order = NULL,
  percentage = FALSE, theme = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Ggplot visualization
}
\description{
Compare quantities over years
}
\section{ctypes}{

Yea-Num
}

\examples{
gg_area_YeaNum(sampleData("Yea-Num", nrow = 10))
}
