% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_bar_CatNum.R
\name{gg_bar_Cat}
\alias{gg_bar_Cat}
\title{Bar Chart Cat}
\usage{
gg_bar_Cat(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
This chart does not allow for chaning orientation
}
\examples{
gg_bar_Cat(sample_data("Cat", nrow = 10))
}
