% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{gg_bar_Cat}
\alias{gg_bar_Cat}
\title{Bar (categories)}
\usage{
gg_bar_Cat(data, title = NULL, subtitle = NULL, caption = NULL,
  horLabel = NULL, verLabel = NULL, horLine = NULL, verLine = NULL,
  colors = c("#009EE3", "#F9B233"), dropNa = FALSE, format = c("", ""),
  highlightValue = NULL, labelRatio = 0.1, labelWrap = 12,
  marks = c(".", ","), nDigits = 2, order = NULL, orientation = "ver",
  percentage = FALSE, sort = "no", sliceN = NULL, showText = TRUE,
  theme = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Ggplot2 visualization
}
\description{
Compare category's levels
}
\section{ctypes}{

Cat, Yea, Dat
}

\examples{
gg_bar_Cat(sampleData("Cat", nrow = 10))
}
