% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_bar_CatCatNum.R
\name{gg_bar_CatCat}
\alias{gg_bar_CatCat}
\title{Bar Chart Cat Cat}
\usage{
gg_bar_CatCat(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
This chart does not allow for chaning orientation
}
\examples{
gg_bar_CatCat(sample_data("Cat-Cat", nrow = 10))
}
