% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{gg_bar_CatCat}
\alias{gg_bar_CatCat}
\title{Bar (categories, ordered categories)}
\usage{
gg_bar_CatCat(data = NULL, agg_text = NULL, caption = NULL,
  colors = NULL, color_opacity = "0.7", color_scale = "discrete",
  drop_na_v = c(FALSE, FALSE), graph_type = "grouped",
  group_color = "transparent", highlight_value = NULL,
  highlight_value_color = "#F9B233", hor_label = NULL,
  hor_line = NULL, hor_line_label = NULL, label_ratio = 1,
  label_wrap_v = c(12, 12), lang = "es", legend_position = "bottom",
  legend_show = TRUE, legend_title = NULL, marks = c(".", ","),
  n_digits = NULL, order1 = NULL, order2 = NULL,
  orientation = "ver", percentage = FALSE, prefix = NULL,
  shape_type = 19, slice_n = NULL, sort = "no", subtitle = NULL,
  suffix = NULL, text_color = "#5A6B72", text_show = TRUE,
  text_size = 3, theme = NULL, title = NULL, ver_label = NULL,
  ver_line = NULL, ver_line_label = NULL, opts = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Ggplot2 visualization
}
\description{
Compare quantities among two categories
}
\section{ctypes}{

Cat-Cat, Cat-Dat, Cat-Yea, Yea-Cat, Yea-Dat, Yea-Yea, Dat-Cat, Dat-Yea, Dat-Dat
}

\examples{
gg_bar_CatCat(sampleData("Cat-Cat", nrow = 10))
}
