% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{gg_bar_CatCatNum}
\alias{gg_bar_CatCatNum}
\title{Bar (categories, ordered categories, numbers)}
\usage{
gg_bar_CatCatNum(data, title = NULL, subtitle = NULL, caption = NULL,
  horLabel = NULL, verLabel = NULL, horLine = NULL, verLine = NULL,
  agg = "sum", agg_text = NULL, colors = NULL,
  colorText = "#5A6B72", colorScale = "discrete", dropNaV = c(FALSE,
  FALSE), prefix = NULL, suffix = NULL, graphType = "grouped",
  labelRatio = 0.5, labelWrapV = c(12, 12),
  legendPosition = "bottom", legendTitle = NULL, marks = c(".", ","),
  nDigits = 0, order1 = NULL, order2 = NULL, orientation = "ver",
  percentage = FALSE, showText = TRUE, sizeText = 3, theme = NULL,
  ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Ggplot2 visualization
}
\description{
Compare quantities among two categories
}
\section{ctypes}{

Cat-Cat-Num, Cat-Dat-Num, Cat-Yea-Num, Yea-Cat-Num, Yea-Dat-Num, Yea-Yea-Num, Dat-Cat-Num, Dat-Yea-Num, Dat-Dat-Num
}

\examples{
gg_bar_CatCatNum(sampleData("Cat-Cat-Num", nrow = 10))
}
