% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{gg_bar_CatNum}
\alias{gg_bar_CatNum}
\title{Bar (categories, numbers)}
\usage{
gg_bar_CatNum(data, title = NULL, subtitle = NULL, caption = NULL,
  horLabel = NULL, verLabel = NULL, horLine = NULL, verLine = NULL,
  agg = "sum", agg_text = NULL, colors = NULL,
  colorText = "#5A6B72", colorScale = "no", dropNa = FALSE,
  prefix = NULL, suffix = NULL, highlightValue = NULL,
  highlightValueColor = NULL, labelRatio = 0.1, labelWrap = 12,
  marks = c(".", ","), nDigits = NULL, order = NULL,
  orientation = "ver", percentage = FALSE, sort = "no",
  sliceN = NULL, showText = TRUE, sizeText = 3, theme = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Ggplot2 visualization
}
\description{
Compare aggregations among category's levels
}
\section{ctypes}{

Cat-Num, Dat-Num, Yea-Num
}

\examples{
gg_bar_CatNum(sampleData("Cat-Num", nrow = 10))
}
