% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{gg_bar_CatNumP}
\alias{gg_bar_CatNumP}
\title{Bar (ordered category, n numbers)}
\usage{
gg_bar_CatNumP(data, title = NULL, subtitle = NULL, caption = NULL,
  horLabel = NULL, verLabel = NULL, horLine = NULL, verLine = NULL,
  agg = "sum", colors = NULL, colorText = "black",
  colorScale = "discrete", dropNa = c(FALSE, FALSE), format = c("",
  ""), graphType = "grouped", labelRatio = 0.5, labelWrap = c(12,
  12), legendPosition = "right", legendTitle = NULL, marks = c(".",
  ","), nDigits = 0, order1 = NULL, order2 = NULL,
  orientation = "ver", percentage = FALSE, showText = TRUE,
  theme = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Ggplot2 visualization
}
\description{
Compare n quantities among category's levels
}
\section{ctypes}{

Cat-NumP
}

\examples{
gg_bar_CatNumP(sampleData("Cat-NumP", nrow = 10))
}
