% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca.R
\name{gg_bar_circular_Cat.}
\alias{gg_bar_circular_Cat.}
\title{Circular bar
Circular Bar}
\usage{
gg_bar_circular_Cat.(data, titleLabel = "", subtitle = "", caption = "",
  fillLabel = NULL, leg_pos = "right", width = 0.85, text = TRUE,
  color_text = "black", type = "count", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Circular bar
Circular Bar
}
\section{ftypes}{
 Cat
}

\examples{
add(1, 1)
add(10, 1)
}
