% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca.R
\name{gg_bar_circular_facet_CaCa.}
\alias{gg_bar_circular_facet_CaCa.}
\title{Circular bar facet
Circular Bar}
\usage{
gg_bar_circular_facet_CaCa.(data, titleLabel = "", subtitle = "",
  caption = "", leg_pos = "right", width = 0.85, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Circular bar facet
Circular Bar
}
\section{ftypes}{
 Ca-Ca
}

\examples{
add(1, 1)
add(10, 1)
}
