% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Nu-bars.R
\name{gg_bar_coloured_hor_x_facet_CaCaNu.}
\alias{gg_bar_coloured_hor_x_facet_CaCaNu.}
\title{gg_bar_coloured_hor_x_facet_CaCaNu.
horizontal bar}
\usage{
gg_bar_coloured_hor_x_facet_CaCaNu.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, leg_pos = "right", ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_bar_coloured_hor_x_facet_CaCaNu.
horizontal bar
}
\section{ftypes}{
 Ca-Ca-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
