% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_bar_coloured_parameter_ver_CaNu.}
\alias{gg_bar_coloured_parameter_ver_CaNu.}
\title{gg_bar_coloured_parameter_ver_CaNu.
Vertical coloured by parameter bars}
\usage{
gg_bar_coloured_parameter_ver_CaNu.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, order = FALSE,
  parameter = NULL, line_mean = FALSE, text = TRUE, type = "count",
  text_size = 3, leg_pos = "right", aggregation = "mean", angle = 0,
  ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_bar_coloured_parameter_ver_CaNu.
Vertical coloured by parameter bars
}
\section{ftypes}{
 Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

