% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_bar_coloured_parameter_ver_CaNu.}
\alias{gg_bar_coloured_parameter_ver_CaNu.}
\title{gg_bar_coloured_parameter_ver_CaNu.
Vertical coloured by parameter bars}
\usage{
gg_bar_coloured_parameter_ver_CaNu.(data, titleLabel = "", xLabel = NULL,
  yLabel = NULL, parameter = NULL, leg_pos = "right", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_bar_coloured_parameter_ver_CaNu.
Vertical coloured by parameter bars
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

