% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ye-Nu.R
\name{gg_bar_coloured_x_ver_YeNu.}
\alias{gg_bar_coloured_x_ver_YeNu.}
\title{Vertical bar coloured by first variable
vertical coloured bar}
\usage{
gg_bar_coloured_x_ver_YeNu.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, text = TRUE, type = "percent",
  color_text = "black", yLabel = NULL, leg_pos = "right",
  aggregation = "sum", angle_x = 0, ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical bar coloured by first variable
vertical coloured bar
}
\section{ftypes}{
 Ye-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
