% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_bar_coloured_y_ver_CaNu.}
\alias{gg_bar_coloured_y_ver_CaNu.}
\title{gg_bar_coloured_y_ver_CaNu.
vertical bar}
\usage{
gg_bar_coloured_y_ver_CaNu.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, reverse = FALSE, text = TRUE,
  type = "percent", size_text = 3, yLabel = NULL, leg_pos = "right",
  ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_bar_coloured_y_ver_CaNu.
vertical bar
}
\section{ftypes}{
 Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

