% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Da-Nu_funs.R
\name{gg_bar_DaNu.}
\alias{gg_bar_DaNu.}
\alias{gg_bubbles_DaNu.}
\alias{gg_lollipop_DaNu.}
\alias{gg_stepped_stacked_area_DaNu.}
\title{gg_bar_DaNu. : title.
bar}
\usage{
gg_bar_DaNu.(data, title = "", xlab = NULL, ylab = NULL)

gg_bubbles_DaNu.(data, title = "", xlab = NULL, ylab = NULL)

gg_lollipop_DaNu.(data, title = "", xlab = NULL, ylab = NULL)

gg_stepped_stacked_area_DaNu.(data, title = "", xlab = NULL, ylab = NULL)
}
\arguments{
\item{x}{A data.}

\item{y}{A number.}

\item{x}{A data.}

\item{y}{A number.}

\item{x}{A data.}

\item{y}{A number.}

\item{x}{A data.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.

The sum of \code{x} and \code{y}.

The sum of \code{x} and \code{y}.

The sum of \code{x} and \code{y}.
}
\description{
gg_bar_DaNu. : title.
bar

gg_bubbles_DaNu. : title.
bubbles

gg_lollipop_DaNu. : title.
lollipop

gg_stepped_stacked_area_DaNu. : title.
stepped stacked area.
}
\section{ftypes}{
 Da-Nu

 Da-Nu

 Da-Nu

 Da-Nu
}
\examples{
add(1, 1)
add(10, 1)
add(1, 1)
add(10, 1)
add(1, 1)
add(10, 1)
add(1, 1)
add(10, 1)
}

