% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Nu-bars.R
\name{gg_bar_facet_ver_CaCaNu.}
\alias{gg_bar_facet_ver_CaCaNu.}
\title{gg_bar_facet_ver_CaCaNu.
Facet vertical bargraph}
\usage{
gg_bar_facet_ver_CaCaNu.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, leg_pos = "right",
  angle_x = 45, aggregation = "mean", fillMissing = TRUE,
  fillMissingValue = 0, groupLabelPosition = "top", ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A category.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_bar_facet_ver_CaCaNu.
Facet vertical bargraph
}
\section{ftypes}{
 Ca-Ca-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
