% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ye-Nu-bars.R
\name{gg_bar_grouped_ver_CaYeNu.}
\alias{gg_bar_grouped_ver_CaYeNu.}
\title{gg_bar_grouped_ver_CaYeNu.
Barras stacked}
\usage{
gg_bar_grouped_ver_CaYeNu.(data, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_bar_grouped_ver_CaYeNu.
Barras stacked
}
\section{ftypes}{
 Ca-Ye-Nu
}

\examples{
add(1, 1)
add(10, 1)

}
