% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca.R
\name{gg_bar_polar_stacked_100_CaCa.}
\alias{gg_bar_polar_stacked_100_CaCa.}
\title{Polar 100% stacked bar
Stacked Polar  100pct}
\usage{
gg_bar_polar_stacked_100_CaCa.(data, titleLabel = "", subtitle = "",
  caption = "", fillLabel = NULL, leg_pos = "right", width = 0.95,
  text = TRUE, color_text = "black", type = "count", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Polar 100% stacked bar
Stacked Polar  100pct
}
\section{ftypes}{
 Ca-Ca
}

\examples{
add(1, 1)
add(10, 1)
}
