% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca.R
\name{gg_bar_single_stacked_ver_Ca.}
\alias{gg_bar_single_stacked_ver_Ca.}
\title{gg_bar_single_stacked_ver_Ca.
Single Vertical Stacked Bar}
\usage{
gg_bar_single_stacked_ver_Ca.(data, titleLabel = "", subtitle = "",
  caption = "", fillLabel = NULL, leg_pos = "right", width = 0.3, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_bar_single_stacked_ver_Ca.
Single Vertical Stacked Bar
}
\section{ftypes}{
 Ca
}
\examples{
add(1, 1)
add(10, 1)
}

