% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Nu.R
\name{gg_bar_stacked_100_ver_CaCaNu.}
\alias{gg_bar_stacked_100_ver_CaCaNu.}
\title{gg_bar_stacked_100_ver_CaCaNu.
100 vertical stacked bar graph}
\usage{
gg_bar_stacked_100_ver_CaCaNu.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, leg_pos = "right", ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A category.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_bar_stacked_100_ver_CaCaNu.
100 vertical stacked bar graph
}
\section{ftypes}{
 Ca-Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

