% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ye-Nu-bars.R
\name{gg_bar_stacked_100_ver_CaYeNu.}
\alias{gg_bar_stacked_100_ver_CaYeNu.}
\title{Horizontal 100% stacked bar by second variable
100 horizontal stacked bar graph}
\usage{
gg_bar_stacked_100_ver_CaYeNu.(data, ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A category.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal 100% stacked bar by second variable
100 horizontal stacked bar graph
}
\section{ftypes}{
 Ca-Ye-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
