% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca.R
\name{gg_bar_stacked_polar_CaCa.}
\alias{gg_bar_stacked_polar_CaCa.}
\title{gg_bar_stacked_polar_CaCa.
Stacked Polar bar}
\usage{
gg_bar_stacked_polar_CaCa.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, text = TRUE,
  type = "percent", size_text = 3, leg_pos = "right", l_ncol = 1, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_bar_stacked_polar_CaCa.
Stacked Polar bar
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

