% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Da-Nu.R
\name{gg_bar_stacked_ver_CaDaNu.}
\alias{gg_bar_stacked_ver_CaDaNu.}
\title{gg_bar_stacked_ver_CaDaNu.
vertical stacked bar graph}
\usage{
gg_bar_stacked_ver_CaDaNu.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, leg_pos = "right",
  angle_x = 45, text_size = 10, hline = NULL, ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A category.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_bar_stacked_ver_CaDaNu.
vertical stacked bar graph
}
\section{ftypes}{
 Ca-Da-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

