% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{gg_bar_YeaNum}
\alias{gg_bar_YeaNum}
\title{Bar (years, numbers)}
\usage{
gg_bar_YeaNum(data, title = NULL, subtitle = NULL, caption = NULL,
  horLabel = NULL, verLabel = NULL, horLine = NULL, verLine = NULL,
  agg = "sum", angleX = 0, colors = c("#009EE3", "#F9B233"),
  dropNa = FALSE, format = c("", ""), highlightValue = NULL,
  order = NULL, orientation = "ver", percentage = FALSE, sort = "no",
  sliceN = NULL, showText = TRUE, theme = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Ggplot2 visualization
}
\description{
Compare quantities over years
}
\section{ctypes}{

Yea-Num
}

\examples{
gg_bar_YeaNum(sampleData("Yea-Num", nrow = 10))
}
