% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{gg_box_DatNum.}
\alias{gg_box_DatNum.}
\title{Boxplot}
\usage{
gg_box_DatNum.(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Boxplot
}
\section{ftypes}{
 Dat-Num
}

\examples{
add(1, 1)
add(10, 1)
}
