% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_bubble_CaNu.}
\alias{gg_bubble_CaNu.}
\title{gg_bubble_CaNu.
Bubble}
\usage{
gg_bubble_CaNu.(data, titleLabel = "", xLabel = NULL, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_bubble_CaNu.
Bubble
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

