% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_bubble_CaNu2.}
\alias{gg_bubble_CaNu2.}
\title{gg_bubble_CaNu2.
Bubbles}
\usage{
gg_bubble_CaNu2.(data, titleLabel = "", subtitle = "", caption = "",
  sep = 3, lim_inf = -80, lim_sup = 80, xLabel = NULL, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_bubble_CaNu2.
Bubbles
}
\section{ftypes}{
 Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

