% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ge-Nu.R
\name{gg_bubble_depto_CaGe.}
\alias{gg_bubble_depto_CaGe.}
\title{gg_bubble_depto_CaGe.
Points inside Colombia's mpios map}
\usage{
gg_bubble_depto_CaGe.(data, titleLabel = "", subtitle = "", caption = "",
  depto_ = "05", leg_pos = "right", fillLabel = NULL, text = FALSE,
  text_size = 2.5, prop_text = 0.1, color_map = "gray",
  color_frontier = "white", scale_point = 0.25, alpha = 0.5, ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_bubble_depto_CaGe.
Points inside Colombia's mpios map
}
\section{ftypes}{
 Ca-Ge
}
\examples{
add(1, 1)
add(10, 1)
}

