% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Da-Nu.R
\name{gg_bubbles_DaNu.}
\alias{gg_bubbles_DaNu.}
\title{gg_bubbles_DaNu.
bubbles}
\usage{
gg_bubbles_DaNu.(data, title = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, angle_x = 0, ...)
}
\arguments{
\item{x}{A data.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_bubbles_DaNu.
bubbles
}
\section{ftypes}{
 Nu, Ye-Nu, Da-Nu, Ca-Nu, Ca
}
\examples{
add(1, 1)
add(10, 1)
}

