% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca.R
\name{gg_bullseye_Ca.}
\alias{gg_bullseye_Ca.}
\title{gg_bullseye_Ca.
Bullseye}
\usage{
gg_bullseye_Ca.(data, titleLabel = "", fillLabel = NULL,
  leg_pos = "right", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_bullseye_Ca.
Bullseye
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

