% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_bullseye_CatNum.}
\alias{gg_bullseye_CatNum.}
\title{Bullseye
Bullseye}
\usage{
gg_bullseye_CatNum.(data, titleLabel = "", subtitle = "", caption = "",
  fillLabel = NULL, leg_pos = "right", aggregation = "sum", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Bullseye
Bullseye
}
\section{ftypes}{
 Cat-Num
}

\examples{
add(1, 1)
add(10, 1)
}
