% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu-Nu.R
\name{gg_circle_CaNuNu.}
\alias{gg_circle_CaNuNu.}
\title{Point
Point}
\usage{
gg_circle_CaNuNu.(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, leg_pos = "right", size = 10, ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Point
Point
}
\section{ftypes}{
 Ca-Nu-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
