% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ye-Nu.R
\name{gg_circle_CaYeNu.}
\alias{gg_circle_CaYeNu.}
\title{gg_circle_CaYeNu.: title.
circle
ciculos}
\usage{
gg_circle_CaYeNu.(data, title = "", xlab = NULL, ylab = NULL,
  leg_pos = "right", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_circle_CaYeNu.: title.
circle
ciculos
}
\section{ftypes}{
 Ca-Ye-Nu,Ca-Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

