% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_pie_CatNum.R
\name{gg_donut_Cat}
\alias{gg_donut_Cat}
\title{Donut (categories)}
\usage{
gg_donut_Cat(data = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Ggplot2 visualization
}
\description{
Compare category's levels
}
\section{ctypes}{

Cat, Yea, Dat
}

\examples{
gg_donut_Cat(sample_data("Cat", nrow = 10))
}
