% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca.R
\name{gg_dot_bar_ver_Ca.}
\alias{gg_dot_bar_ver_Ca.}
\title{gg_dot_bar_ver_Ca.
Vertical Dot Bar}
\usage{
gg_dot_bar_ver_Ca.(data, titleLabel = "", xLabel = NULL, yLabel = "Count",
  fillLabel = NULL, leg_pos = "right")
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_dot_bar_ver_Ca.
Vertical Dot Bar
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

