% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R, R/Ca.R
\name{gg_gauge_dial_Ca.}
\alias{gg_gauge_dial_Ca.}
\title{gg_gauge_dial_Ca.
Gauge}
\usage{
gg_gauge_dial_Ca.(data)

gg_gauge_dial_Ca.(data)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}

\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.

The sum of \code{x} and \code{y}.
}
\description{
gg_gauge_dial_Ca.
Gauge

gg_gauge_dial_Ca.
Gauge
}
\section{ftypes}{
 Ca,Ca-Nu

 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
add(1, 1)
add(10, 1)
}

