% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_gauge_dial_CaNu.}
\alias{gg_gauge_dial_CaNu.}
\title{Dial gauge
Gauge}
\usage{
gg_gauge_dial_CaNu.(data, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Dial gauge
Gauge
}
\section{ftypes}{
 Ca-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
