% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nu-Nu.R
\name{gg_hist_NuNu.}
\alias{gg_hist_NuNu.}
\title{Vertical histogram
Histogram 2D}
\usage{
gg_hist_NuNu.(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, reverse = FALSE, angle_x = 0,
  leg_pos = "right", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical histogram
Histogram 2D
}
\section{ftypes}{
 Nu-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
