% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Da.R
\name{gg_histogram_CaDa.}
\alias{gg_histogram_CaDa.}
\title{gg_histogram_CaDa.: title.
histogram}
\usage{
gg_histogram_CaDa.(data, title = "", xlab = NULL, clab = NULL)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_histogram_CaDa.: title.
histogram
}
\section{ftypes}{
 Ca-Ye-Nu,Ca-Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

