% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Da.R
\name{gg_histogram_CaDa.}
\alias{gg_histogram_CaDa.}
\title{Histogram
histogram}
\usage{
gg_histogram_CaDa.(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Histogram
histogram
}
\section{ftypes}{
 Ca-Da
}

\examples{
add(1, 1)
add(10, 1)
}
