% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Da-Nu_funs.R
\name{gg_horizon_DaNu.}
\alias{gg_horizon_DaNu.}
\title{gg_horizon_DaNu.
Horizon}
\usage{
gg_horizon_DaNu.(data, titleLabel = "Report", xLabel = NULL,
  yLabel = NULL, leg_pos = "right", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_horizon_DaNu.
Horizon
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

