% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Da-Nu.R
\name{gg_kagi_DaNu.}
\alias{gg_kagi_DaNu.}
\title{gg_kagi_DaNu.
Kagi}
\usage{
gg_kagi_DaNu.(data, title = "", subtitle = "", caption = "",
  xlab = NULL, ylab = NULL, hline = NULL, angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_kagi_DaNu.
Kagi
}
\section{ftypes}{
 Da-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
