% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Da-Nu_funs.R
\name{gg_kagi_DaNu.}
\alias{gg_kagi_DaNu.}
\title{gg_kagi_DaNu. : title.
Kaggi}
\usage{
gg_kagi_DaNu.(data, title = "", xlab = NULL, ylab = NULL)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_kagi_DaNu. : title.
Kaggi
}
\section{ftypes}{
 Da-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

