% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R, R/possible_viz.R
\name{gg_line_Cat}
\alias{gg_line_Cat}
\title{Lines (categories)}
\usage{
gg_line_Cat(data, titleLabel = "", xLabel = NULL, subtitle = "",
  caption = "", yLabel = NULL, leg_pos = "right", angle_x = 0, ...)

gg_line_Cat(data, titleLabel = "", xLabel = NULL, subtitle = "",
  caption = "", yLabel = NULL, leg_pos = "right", angle_x = 0, ...)
}
\arguments{
\item{data}{A data.frame}

\item{x}{A number.}

\item{y}{A number.}
}
\value{
Ggplot2 visualization

The sum of \code{x} and \code{y}.
}
\description{
Compare category's levels
}
\section{ctypes}{

Cat
}

\section{ftypes}{
 Cat
}

\examples{
gg_line_Cat(sampleData("Cat", nrow = 10))
gg_line_Cat(sampleData("Cat"))
}
