% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{gg_line_Cat}
\alias{gg_line_Cat}
\title{Lines (categories)}
\usage{
gg_line_Cat(data = NULL, opts = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Ggplot2 visualization
}
\description{
Compare category's levels
}
\section{ctypes}{

Cat
}

\examples{
gg_line_Cat(sampleData("Cat", nrow = 10))
}
