% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{gg_line_Cat}
\alias{gg_line_Cat}
\title{Lines (categories)}
\usage{
gg_line_Cat(data, title = NULL, subtitle = NULL, caption = NULL,
  horLabel = NULL, verLabel = NULL, horLine = NULL, verLine = NULL,
  colors = NULL, colorText = "black", colorScale = "no",
  dropNa = FALSE, format = c("", ""), highlightValue = NULL,
  highlightValueColor = NULL, labelRatio = 1, labelWrap = 12,
  marks = c(".", ","), nDigits = 0, order = NULL,
  orientation = "ver", percentage = FALSE, shapeType = 19,
  sort = "no", sliceN = NULL, showText = TRUE, spline = FALSE,
  startAtZero = TRUE, theme = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Ggplot2 visualization
}
\description{
Compare category's levels
}
\section{ctypes}{

Cat
}

\examples{
gg_line_Cat(sampleData("Cat", nrow = 10))
}
