% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{gg_line_CatCat}
\alias{gg_line_CatCat}
\title{Lines (categories, ordered categories)}
\usage{
gg_line_CatCat(data, title = NULL, subtitle = NULL, caption = NULL,
  horLabel = NULL, verLabel = NULL, horLine = NULL, verLine = NULL,
  agg = "sum", colors = NULL, colorText = "black",
  colorScale = "discrete", dropNa = c(FALSE, FALSE), format = c("",
  ""), labelRatio = 1, labelWrap = c(12, 12),
  legendPosition = "right", legendTitle = NULL, marks = c(".", ","),
  nDigits = 2, order = NULL, orientation = "ver",
  percentage = FALSE, shapeType = 19, showText = TRUE,
  spline = FALSE, startAtZero = TRUE, theme = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Ggplot2 visualization
}
\description{
Compare quantities among two categories
}
\section{ctypes}{

Cat-Cat, Cat-Dat, Cat-Yea, Yea-Cat, Yea-Dat, Yea-Yea, Dat-Cat, Dat-Yea, Dat-Dat
}

\examples{
gg_line_CatCat(sampleData("Cat-Cat", nrow = 10))
}
