% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{gg_line_CatCatNum}
\alias{gg_line_CatCatNum}
\title{Lines (categories, ordered categories, numbers)}
\usage{
gg_line_CatCatNum(data, title = NULL, subtitle = NULL,
  caption = NULL, horLabel = NULL, verLabel = NULL, horLine = NULL,
  verLine = NULL, agg = "sum", agg_text = NULL, colors = NULL,
  colorText = "#5A6B72", colorScale = "discrete", dropNaV = c(FALSE,
  FALSE), prefix = NULL, suffix = NULL, labelRatio = 1,
  labelWrapV = c(12, 12), legendPosition = "bottom",
  legendTitle = NULL, marks = c(".", ","), nDigits = 0,
  order = NULL, orientation = "ver", percentage = FALSE,
  shapeType = 19, showText = TRUE, spline = FALSE,
  startAtZero = TRUE, theme = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Ggplot2 visualization
}
\description{
Compare quantities among two categories
}
\section{ctypes}{

Cat-Cat-Num, Cat-Dat-Num, Cat-Yea-Num, Yea-Cat-Num, Yea-Dat-Num, Yea-Yea-Num, Dat-Cat-Num, Dat-Yea-Num, Dat-Dat-Num
}

\examples{
gg_line_CatCatNum(sampleData("Cat-Cat-Num", nrow = 10))
}
