% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{gg_line_CatCatNum}
\alias{gg_line_CatCatNum}
\title{Lines (categories, ordered categories, numbers)}
\usage{
gg_line_CatCatNum(
  data = NULL,
  agg = "sum",
  agg_text = NULL,
  caption = NULL,
  colors = NULL,
  color_opacity = 0.7,
  color_scale = "discrete",
  drop_na = FALSE,
  drop_na_legend = FALSE,
  graph_type = "grouped",
  group_color = "transparent",
  highlight_value = NULL,
  highlight_value_color = "#F9B233",
  hor_label = NULL,
  hor_line = NULL,
  hor_line_label = NULL,
  label_ratio = 1,
  label_wrap = 12,
  label_wrap_legend = 12,
  legend_position = "bottom",
  legend_show = TRUE,
  legend_title = NULL,
  marks = c(".", ","),
  n_digits = NULL,
  order1 = NULL,
  order2 = NULL,
  orientation = "ver",
  percentage = FALSE,
  prefix = NULL,
  shape_type = 19,
  slice_n = NULL,
  sort = "no",
  spline = FALSE,
  start_zero = TRUE,
  subtitle = NULL,
  suffix = NULL,
  text_color = "#212428",
  text_show = TRUE,
  text_size = 3,
  theme = NULL,
  title = NULL,
  ver_label = NULL,
  ver_line = NULL,
  ver_line_label = NULL,
  opts = NULL,
  ...
)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Ggplot2 visualization
}
\description{
Compare quantities among two categories
}
\section{ctypes}{

Cat-Cat-Num, Cat-Dat-Num, Cat-Yea-Num, Yea-Cat-Num, Yea-Dat-Num, Yea-Yea-Num, Dat-Cat-Num, Dat-Yea-Num, Dat-Dat-Num
}

\examples{
gg_line_CatCatNum(sampleData("Cat-Cat-Num", nrow = 10))
}
