% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{gg_line_CatDatNum}
\alias{gg_line_CatDatNum}
\title{Lines (catrgories, dates, numbers)}
\usage{
gg_line_CatDatNum(data, title = NULL, subtitle = NULL,
  caption = NULL, horLabel = NULL, verLabel = NULL, horLine = NULL,
  verLine = NULL, agg = "sum", colors = NULL, colorText = "black",
  colorScale = "discrete", dropNa = c(FALSE, FALSE), format = c("",
  ""), labelRatio = 1, labelWrap = c(12, 12),
  legendPosition = "right", legendTitle = NULL, marks = c(".", ","),
  nDigits = 0, order = NULL, orientation = "ver",
  percentage = FALSE, shapeType = 19, showText = TRUE,
  spline = FALSE, startAtZero = TRUE, theme = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Ggplot2 visualization
}
\description{
Compare quantities among categories over dates
}
\section{ctypes}{

Cat-Dat-Num
}

\examples{
gg_line_CatDatNum(sampleData("Cat-Dat-Num", nrow = 10))
}
