% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{gg_line_CatNum}
\alias{gg_line_CatNum}
\title{Lines (categories, numbers)}
\usage{
gg_line_CatNum(data = NULL, opts = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Ggplot2 visualization
}
\description{
Compare aggregations among category's levels
}
\section{ctypes}{

Cat-Num, Dat-Num, Yea-Num
}

\examples{
gg_line_CatNum(sampleData("Cat-Num", nrow = 10))
}
