% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{gg_line_CatNum}
\alias{gg_line_CatNum}
\title{Lines (categories, numbers)}
\usage{
gg_line_CatNum(data, title = NULL, subtitle = NULL, caption = NULL,
  horLabel = NULL, verLabel = NULL, horLine = NULL, verLine = NULL,
  agg = "sum", agg_text = NULL, colors = NULL,
  colorText = "#5A6B72", colorScale = "no", dropNa = FALSE,
  prefix = NULL, suffix = NULL, highlightValue = NULL,
  highlightValueColor = NULL, labelRatio = 1, labelWrap = 12,
  marks = c(".", ","), nDigits = 0, order = NULL,
  orientation = "ver", percentage = FALSE, shapeType = 19,
  sort = "no", sliceN = NULL, showText = TRUE, spline = FALSE,
  startAtZero = TRUE, theme = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Ggplot2 visualization
}
\description{
Compare aggregations among category's levels
}
\section{ctypes}{

Cat-Num, Dat-Num, Yea-Num
}

\examples{
gg_line_CatNum(sampleData("Cat-Num", nrow = 10))
}
