% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{gg_line_CatNumP}
\alias{gg_line_CatNumP}
\title{Lines (ordered category, n numbers)}
\usage{
gg_line_CatNumP(data, title = NULL, subtitle = NULL, caption = NULL,
  horLabel = NULL, verLabel = NULL, horLine = NULL, verLine = NULL,
  agg = "sum", colors = NULL, colorText = "black",
  colorScale = "discrete", dropNaV = c(FALSE, FALSE), format = c("",
  ""), labelRatio = 1, labelWrapV = c(12, 12),
  legendPosition = "right", legendTitle = NULL, marks = c(".", ","),
  nDigits = 0, order = NULL, orientation = "ver",
  percentage = FALSE, shapeType = 19, showText = TRUE,
  spline = FALSE, startAtZero = TRUE, theme = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Ggplot2 visualization
}
\description{
Compare n quantities among category's levels
}
\section{ctypes}{

Cat-NumP
}

\examples{
gg_line_CatNumP(sampleData("Cat-NumP", nrow = 10))
}
