% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nu-Nu.R
\name{gg_line_multi_NumNum.}
\alias{gg_line_multi_NumNum.}
\title{Line for each numeric variable
Line for each Num}
\usage{
gg_line_multi_NumNum.(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Line for each numeric variable
Line for each Num
}
\section{ftypes}{
 Num-Num
}
\examples{
add(1, 1)
add(10, 1)
}

