% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_line_point_multi_CaNu.}
\alias{gg_line_point_multi_CaNu.}
\title{Grouped line + point
Grouped Line Color Point}
\usage{
gg_line_point_multi_CaNu.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = "Index", yLabel = NULL, leg_pos = "right",
  type = 1, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Grouped line + point
Grouped Line Color Point
}
\section{ftypes}{
 Ca-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
