% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possible_viz.R
\name{gg_line_point_multi_NumNum.}
\alias{gg_line_point_multi_NumNum.}
\title{Line + point
Line point plot}
\usage{
gg_line_point_multi_NumNum.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, shape_type = 19,
  angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Line + point
Line point plot
}
\section{ftypes}{
 Num-Num
}

\examples{
gg_line_point_multi_NumNum.(sampleData("Num-Num"))
}
