% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nu-Nu.R
\name{gg_line_point_NuNu.}
\alias{gg_line_point_NuNu.}
\title{Line + point
Line point plot}
\usage{
gg_line_point_NuNu.(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, shape_type = 19, angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Line + point
Line point plot
}
\section{ftypes}{
 Nu-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
