% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca.R
\name{gg_line_ver_Cat.}
\alias{gg_line_ver_Cat.}
\title{Vertical line
Vertical Line}
\usage{
gg_line_ver_Cat.(data, titleLabel = "", xLabel = NULL, subtitle = "",
  caption = "", yLabel = NULL, leg_pos = "right", angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical line
Vertical Line
}
\section{ftypes}{
 Cat
}
\examples{
add(1, 1)
add(10, 1)
}

