% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possible_viz.R
\name{gg_line_YeaNum.}
\alias{gg_line_YeaNum.}
\title{Horizontal line
Horizontal Lines}
\usage{
gg_line_YeaNum.(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, angle_x = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal line
Horizontal Lines
}
\section{ftypes}{
 Yea-Num
}

\examples{
gg_line_YeaNum.(sampleData("Yea-Num"))
}
