% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ye-Nu.R
\name{gg_lines_hor_CaYeNu.}
\alias{gg_lines_hor_CaYeNu.}
\title{gg_lines_hor_CaYeNu.: title.
Lines
Tiene múltiples líneas}
\usage{
gg_lines_hor_CaYeNu.(data, title = "", subtitle = "", caption = "",
  xlab = NULL, ylab = NULL, leg_pos = "right", angle = 90, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_lines_hor_CaYeNu.: title.
Lines
Tiene múltiples líneas
}
\section{ftypes}{
 Ca-Ye-Nu,Ca-Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

