% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ye-Nu.R
\name{gg_lines_hor_YeNu.}
\alias{gg_lines_hor_YeNu.}
\title{gg_lines_hor_YeNu.: title.
Barras stacked
Tiene múltiples líneas}
\usage{
gg_lines_hor_YeNu.(data, title = "", xlab = NULL, ylab = NULL)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_lines_hor_YeNu.: title.
Barras stacked
Tiene múltiples líneas
}
\section{ftypes}{
 Ca-Ye-Nu,Ca-Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

