% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Da-Nu_funs.R
\name{gg_lollipop_DaNu.}
\alias{gg_lollipop_DaNu.}
\title{gg_lollipop_DaNu.
lollipop}
\usage{
gg_lollipop_DaNu.(data, title = "", subtitle = "", caption = "",
  xlab = NULL, ylab = NULL, ...)
}
\arguments{
\item{x}{A data.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_lollipop_DaNu.
lollipop
}
\section{ftypes}{
 Da-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

