% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ye-Nu.R
\name{gg_lollipop_YeNu.}
\alias{gg_lollipop_YeNu.}
\title{gg_lollipop_YeNu.
lollipop
segment and point}
\usage{
gg_lollipop_YeNu.(data, title = "", xlab = NULL, ylab = NULL, size = 7)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_lollipop_YeNu.
lollipop
segment and point
}
\section{ftypes}{
 Ca-Ye-Nu,Ca-Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

