% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Da-Nu.R
\name{gg_multi_line_point_CaDaNu.}
\alias{gg_multi_line_point_CaDaNu.}
\title{gg_multi_line_point_CaDaNu.
Grouped Line Color Point}
\usage{
gg_multi_line_point_CaDaNu.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, fillLabel = NULL,
  leg_pos = "right", type = 1, size_text = 8, angle_x = 45, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_multi_line_point_CaDaNu.
Grouped Line Color Point
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

