% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca.R
\name{gg_pie_Ca.}
\alias{gg_pie_Ca.}
\title{gg_pie_Ca.
Pie}
\usage{
gg_pie_Ca.(data, titleLabel = "", fillLabel = NULL, leg_pos = "right")
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_pie_Ca.
Pie
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

