% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_pie_CaNu.}
\alias{gg_pie_CaNu.}
\title{gg_pie_CaNu.
Pie}
\usage{
gg_pie_CaNu.(data, titleLabel = "", subtitle = "", caption = "",
  text = TRUE, type = "percent", text_size = 3, leg_pos = "right",
  aggregation = "sum", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_pie_CaNu.
Pie
}
\section{ftypes}{
 Ca-Nu
}
\examples{
add(1, 1)z
add(10, 1)
}

