% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pie.R
\name{gg_pie_Cat}
\alias{gg_pie_Cat}
\title{Pie}
\usage{
gg_pie_Cat(data, title = NULL, subtitle = NULL, caption = NULL,
  colors = "#009EE3", colorText = "black", dropNa = FALSE,
  format = c("", ""), labelRatio = 0.1, labelWrap = 12,
  legPos = "right", legTitle = "", marks = c(".", ","),
  nDigits = 2, order = NULL, percentage = FALSE, sort = "no",
  sliceN = NULL, showText = TRUE, theme = NULL, ...)
}
\description{
Pie
}
