% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportions.R
\name{gg_pie_Cat}
\alias{gg_pie_Cat}
\title{Pie (categories)}
\usage{
gg_pie_Cat(data, title = NULL, subtitle = NULL, caption = NULL,
  colors = NULL, colorText = "#5A6B72", colorScale = "discrete",
  dropNa = FALSE, format = c("", ""), labelRatio = 0.1,
  labelWrap = 12, legendPosition = "right", legendTitle = NULL,
  marks = c(".", ","), nDigits = 0, order = NULL,
  orientation = "ver", percentage = FALSE, sort = "no",
  sliceN = NULL, showText = TRUE, theme = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Ggplot2 visualization
}
\description{
Compare category's levels
}
\section{ctypes}{

Cat, Yea, Dat
}

\examples{
gg_pie_Cat(sampleData("Cat", nrow = 10))
}
