% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportions.R
\name{gg_pie_Cat}
\alias{gg_pie_Cat}
\title{Pie (categories)}
\usage{
gg_pie_Cat(data = NULL, opts = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Ggplot2 visualization
}
\description{
Compare category's levels
}
\section{ctypes}{

Cat, Yea, Dat
}

\examples{
gg_pie_Cat(sampleData("Cat", nrow = 10))
}
