% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportions.R
\name{gg_pie_CatNum}
\alias{gg_pie_CatNum}
\title{Pie (categories, numbers)}
\usage{
gg_pie_CatNum(data, title = NULL, subtitle = NULL, caption = NULL,
  agg = "sum", colors = NULL, colorText = "black",
  colorScale = "discrete", dropNa = FALSE, format = c("", ""),
  labelRatio = 0.1, labelWrap = 12, legendPosition = "right",
  legendTitle = NULL, marks = c(".", ","), nDigits = 0,
  order = NULL, percentage = FALSE, sort = "no", sliceN = NULL,
  showText = TRUE, theme = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Ggplot2 visualization
}
\description{
Compare aggregations among category's levels
}
\section{ctypes}{

Cat-Num, Dat-Num, Yea-Num
}

\examples{
gg_pie_CatNum(sampleData("Cat-Num", nrow = 10))
}
