% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportions.R
\name{gg_pie_CatNum}
\alias{gg_pie_CatNum}
\title{Pie (categories, numbers)}
\usage{
gg_pie_CatNum(
  data = NULL,
  agg = "sum",
  agg_text = NULL,
  background = "transparent",
  caption = NULL,
  colors = NULL,
  color_scale = "discrete",
  drop_na = FALSE,
  highlight_value = NULL,
  highlight_value_color = "#F9B233",
  hor_label = NULL,
  hor_line = NULL,
  label_ratio = 1,
  label_wrap = 12,
  legend_position = "bottom",
  legend_show = TRUE,
  legend_title = NULL,
  marks = c(".", ","),
  n_digits = NULL,
  order = NULL,
  orientation = "ver",
  percentage = FALSE,
  prefix = NULL,
  slice_n = NULL,
  sort = "no",
  subtitle = NULL,
  suffix = NULL,
  text_color = "#5A6B72",
  text_show = TRUE,
  text_size = 3,
  theme = NULL,
  title = NULL,
  opts = NULL,
  ...
)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Ggplot2 visualization
}
\description{
Compare aggregations among category's levels
}
\section{ctypes}{

Cat-Num, Dat-Num, Yea-Num
}

\examples{
gg_pie_CatNum(sampleData("Cat-Num", nrow = 10))
}
