% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca.R
\name{gg_pie_facet_CatCat.}
\alias{gg_pie_facet_CatCat.}
\title{Pie facet
Facet Pie}
\usage{
gg_pie_facet_CatCat.(data, titleLabel = "", subtitle = "", caption = "",
  fillLabel = NULL, leg_pos = "right", text = TRUE, type = "count",
  color_text = "black", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Pie facet
Facet Pie
}
\section{ftypes}{
 Cat-Cat
}

\examples{
add(1, 1)
add(10, 1)
}
