% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Nu.R
\name{gg_pie_facet_CatCatNum.}
\alias{gg_pie_facet_CatCatNum.}
\title{Pie facet by second variable
Facet Pie}
\usage{
gg_pie_facet_CatCatNum.(data, titleLabel = "", subtitle = "",
  caption = "", fillLabel = NULL, leg_pos = "right",
  aggregation = "sum", text = TRUE, type = "count",
  color_text = "black", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Pie facet by second variable
Facet Pie
}
\section{ftypes}{
 Cat-Cat-Num
}
\examples{
add(1, 1)
add(10, 1)
}

