% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Ca.R
\name{gg_point_CaCaCa.}
\alias{gg_point_CaCaCa.}
\title{Scatter
Coloured Point}
\usage{
gg_point_CaCaCa.(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, fillLabel = NULL, angle_x = 0,
  shape_type = 19, leg_pos = "right", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Scatter
Coloured Point
}
\section{ftypes}{
 Ca-Ca-Ca
}

\examples{
add(1, 1)
add(10, 1)
}
