% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Ca.R
\name{gg_point_CaCaCa.}
\alias{gg_point_CaCaCa.}
\title{gg_point_CaCaCa.
Coloured Point}
\usage{
gg_point_CaCaCa.(data, titleLabel = "Report", xLabel = NULL,
  yLabel = NULL, fillLabel = NULL, size = 3, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_point_CaCaCa.
Coloured Point
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

