% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu-Nu.R
\name{gg_point_CaNuNu.}
\alias{gg_point_CaNuNu.}
\title{Scatter
point}
\usage{
gg_point_CaNuNu.(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, leg_pos = "right", shape_point = 1, ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Scatter
point
}
\section{ftypes}{
 Ca-Nu-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
