% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nu.R
\name{gg_point_flip_Nu.}
\alias{gg_point_flip_Nu.}
\title{gg_point_flip_Nu.
Line plot}
\usage{
gg_point_flip_Nu.(data, title = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, type = 0, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_point_flip_Nu.
Line plot
}
\section{ftypes}{
 Nu
}
\examples{
add(1, 1)
add(10, 1)
}

