% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_point_grouped_CaNu.}
\alias{gg_point_grouped_CaNu.}
\title{gg_point_grouped_CaNu.
Grouped Color Point}
\usage{
gg_point_grouped_CaNu.(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = "Index", yLabel = NULL, leg_pos = "right", type = 1, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_point_grouped_CaNu.
Grouped Color Point
}
\section{ftypes}{
 Ca-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
