% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nu-Nu.R
\name{gg_point_NuNu.}
\alias{gg_point_NuNu.}
\title{gg_point_NuNu.
Scatter plot}
\usage{
gg_point_NuNu.(data, title = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, shape = 1, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_point_NuNu.
Scatter plot
}
\section{ftypes}{
 Nu-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
