% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Da.R
\name{gg_pointline_ver_CaDa.}
\alias{gg_pointline_ver_CaDa.}
\title{Vertical line + point
pointlines}
\usage{
gg_pointline_ver_CaDa.(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, fillLabel = NULL, angle_x = 0,
  shape_type = 19, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical line + point
pointlines
}
\section{ftypes}{
 Ca-Da
}

\examples{
add(1, 1)
add(10, 1)
}
