% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Da.R
\name{gg_pointline_ver_CatDat.}
\alias{gg_pointline_ver_CatDat.}
\title{Vertical line + point
pointlines}
\usage{
gg_pointline_ver_CatDat.(data, titleLabel = "", subtitle = "",
  caption = "", xLabel = NULL, yLabel = NULL, fillLabel = NULL,
  angle_x = 0, shape_type = 19, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Vertical line + point
pointlines
}
\section{ftypes}{
 Cat-Dat
}
\examples{
add(1, 1)
add(10, 1)
}

