% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Nu.R
\name{gg_pyramid_CaCaNu.}
\alias{gg_pyramid_CaCaNu.}
\title{gg_pyramid_CaCaNu.
pyramid}
\usage{
gg_pyramid_CaCaNu.(data, titleLabel = "Report", fillLabel = NULL)
}
\arguments{
\item{x}{A category.}

\item{y}{A category.}

\item{z}{A category.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_pyramid_CaCaNu.
pyramid
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

