% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{gg_pyramid_CatCatNum.}
\alias{gg_pyramid_CatCatNum.}
\title{Pyramid
pyramid}
\usage{
gg_pyramid_CatCatNum.(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = NULL, aggregation = "sum", yLabel = NULL, fillLabel = NULL,
  leg_pos = "right", angle_x = 0, text = TRUE, type = "count",
  color_text = "black", ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A category.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Pyramid
pyramid
}
\section{ftypes}{
 Cat-Cat-Num
}

\examples{
add(1, 1)
add(10, 1)
}
