% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Da-Nu.R
\name{gg_scatter_hor_CaDaNu.}
\alias{gg_scatter_hor_CaDaNu.}
\title{gg_scatter_hor_CaDaNu.: title.
pointlines}
\usage{
gg_scatter_hor_CaDaNu.(data, title = "", xlab = NULL, ylab = NULL,
  clab = NULL, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_scatter_hor_CaDaNu.: title.
pointlines
}
\section{ftypes}{
 Ca-Ye-Nu,Ca-Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

