% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Da-Nu.R
\name{gg_scatter_hor_CaDaNu.}
\alias{gg_scatter_hor_CaDaNu.}
\title{Horizontal scatter
pointlines}
\usage{
gg_scatter_hor_CaDaNu.(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, clab = NULL, angle_x = 0,
  shape_type = 19, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Horizontal scatter
pointlines
}
\section{ftypes}{
 Ca-Da-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
