% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{gg_slope_CatNum.}
\alias{gg_slope_CatNum.}
\title{Slope
Slope}
\usage{
gg_slope_CatNum.(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, leg_pos = "right", overlap = TRUE,
  text_size = 6, size_point = 3, size_line = 1, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Slope
Slope
}
\section{ftypes}{
 Cat-Num
}

\examples{
add(1, 1)
add(10, 1)
}
