% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Da-Nu.R
\name{gg_steam_CaDaNu.}
\alias{gg_steam_CaDaNu.}
\title{gg_steam_CaDaNu.
Steam}
\usage{
gg_steam_CaDaNu.(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, fillLabel = NULL, leg_pos = "right",
  angle = 45, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_steam_CaDaNu.
Steam
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

