% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu-Nu.R
\name{gg_steam_CaNuNu.}
\alias{gg_steam_CaNuNu.}
\title{gg_steam_CaNuNu.
Steamgraph}
\usage{
gg_steam_CaNuNu.(data, titleLabel = "", subtitle = "", caption = "",
  xLabel = NULL, yLabel = NULL, leg_pos = "right", ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_steam_CaNuNu.
Steamgraph
}
\section{ftypes}{
 Ca-Nu-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
