% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ye-Nu.R
\name{gg_steam_CaYeNu.}
\alias{gg_steam_CaYeNu.}
\title{gg_steam_CaYeNu.
Steam}
\usage{
gg_steam_CaYeNu.(data, titleLabel = "", xLabel = NULL, yLabel = NULL,
  leg_pos = "right", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_steam_CaYeNu.
Steam
}
\section{ftypes}{
 Ca-Ye-Nu,Ca-Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

