% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca.R
\name{gg_treemap_Ca.}
\alias{gg_treemap_Ca.}
\title{gg_treemap_Ca.
Treemap fill by first Ca}
\usage{
gg_treemap_Ca.(data, titleLabel = "", subtitle = "", caption = "",
  fillLabel = NULL, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_treemap_Ca.
Treemap fill by first Ca
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

