% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap.R
\name{gg_treemap_Cat}
\alias{gg_treemap_Cat}
\title{Treemap (categories)}
\usage{
gg_treemap_Cat(data = NULL, opts = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare aggregations among category's levels
}
\section{ctypes}{

Cat, Yea, Dat
}

\examples{
gg_treemap_Cat(sampleData("Cat", nrow = 10))
}
