% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_treemap_CatNum.R
\name{gg_treemap_Cat}
\alias{gg_treemap_Cat}
\title{treemap Chart Cat}
\usage{
gg_treemap_Cat(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
This chart does not allow for chaning orientation
}
\examples{
gg_treemap_Cat(sample_data("Cat", nrow = 10))
}
