% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap.R
\name{gg_treemap_CatCat}
\alias{gg_treemap_CatCat}
\title{Treemap (categories, categories)}
\usage{
gg_treemap_CatCat(data = NULL, opts = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Ggplot visualization
}
\description{
Compare aggregations among category's levels
}
\section{ctypes}{

Cat-Cat, Cat-Yea, Cat-Dat,
}

\examples{
gg_treemap_CatCat(sampleData("Cat-Cat", nrow = 10))
}
