% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_treemap_CatCatNum.R
\name{gg_treemap_CatCatNum}
\alias{gg_treemap_CatCatNum}
\title{Treemap Chart Categories Categories Numeric}
\usage{
gg_treemap_CatCatNum(data, ...)
}
\arguments{
\item{data}{A data.frame}

\item{...}{
  Arguments passed on to \code{\link[dsvizopts:chart_viz_options]{dsvizopts::dsviz_default_opts}}
  \describe{
    \item{\code{}}{}
  }}

\item{orientation}{Doesn't do anything for this type of chart.}

\item{order}{doesn't do anything}
}
\description{
This chart does not allow for chaning orientation
}
\section{ctypes}{

Cat-Cat-Num, Cat-Cat
}

\examples{
gg_treemap_CatCatNum(sample_data("Cat-Yea-Num", nrow = 10))
}
