% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap.R
\name{gg_treemap_CatCatNum}
\alias{gg_treemap_CatCatNum}
\title{Treemap (categories, categories, numbers)}
\usage{
gg_treemap_CatCatNum(
  data = NULL,
  agg = "sum",
  agg_text = NULL,
  caption = NULL,
  colors = NULL,
  color_scale = "discrete",
  drop_na = FALSE,
  drop_na_legend = FALSE,
  group_color = "transparent",
  highlight_value = NULL,
  highlight_value_color = "#F9B233",
  label_wrap = 12,
  label_wrap_legend = 12,
  legend_position = "bottom",
  legend_show = TRUE,
  legend_title = NULL,
  marks = c(".", ","),
  n_digits = NULL,
  order1 = NULL,
  order2 = NULL,
  percentage = FALSE,
  prefix = NULL,
  slice_n = NULL,
  sort = "no",
  subtitle = NULL,
  suffix = NULL,
  text_color_v = c("#FFFFFF", "#212428"),
  text_position_v = c("bottomleft", "topleft"),
  text_show_v = c(TRUE, TRUE),
  text_size_v = c(15, 17),
  theme = NULL,
  title = NULL,
  opts = NULL,
  ...
)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Ggplot visualization
}
\description{
Compare aggregations among category's levels
}
\section{ctypes}{

Cat-Cat-Num, Cat-Yea-Num, Cat-Dat-Num,
}

\examples{
gg_treemap_CatCatNum(sampleData("Cat-Cat-Num", nrow = 10))
}
