% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap.R
\name{gg_treemap_CatNum}
\alias{gg_treemap_CatNum}
\title{Treemap (categories, numbers)}
\usage{
gg_treemap_CatNum(data, title = NULL, subtitle = NULL,
  caption = NULL, labelWrap = 12, colors = NULL,
  colorScale = "discrete", agg = "sum", marks = c(".", ","),
  nDigits = NULL, dropNa = FALSE, prefix = NULL, suffix = NULL,
  highlightValueColor = "#F9B233", percentage = FALSE, format = c("",
  ""), highlightValue = NULL, sliceN = NULL, showText = TRUE,
  showLegend = TRUE, legendPosition = "bottom", theme = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare aggregations among category's levels
}
\section{ctypes}{

Cat-Num, Yea-Num, Dat-Num,
}

\examples{
gg_treemap_CatNum(sampleData("Cat-Num", nrow = 10))
}
