% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_treemap_CatNum.R
\name{gg_treemap_CatNum}
\alias{gg_treemap_CatNum}
\title{treemap Chart Cat Numeric}
\usage{
gg_treemap_CatNum(data, ...)
}
\arguments{
\item{data}{A data.frame}

\item{...}{
  Arguments passed on to \code{\link[dsvizopts:chart_viz_options]{dsvizopts::dsviz_default_opts}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
This chart does not allow for chaning orientation
}
\section{ctypes}{

Cat-Num, Yea-Num
}

\examples{
gg_treemap_CatNum(sample_data("Cat-Num", nrow = 10))
}
