% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Nu.R
\name{gg_treemap_density_y_CaNu.}
\alias{gg_treemap_density_y_CaNu.}
\title{gg_treemap_density_y_CaNu.
Treemap Density by Nu}
\usage{
gg_treemap_density_y_CaNu.(data, titleLabel = "", subtitle = "",
  caption = "", reverse = FALSE, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_treemap_density_y_CaNu.
Treemap Density by Nu
}
\section{ftypes}{
 Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

