% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca-Ca-Nu.R
\name{gg_treemap_y_CaCaNu.}
\alias{gg_treemap_y_CaCaNu.}
\title{gg_treemap_y_CaCaNu.
Treemap Fill by second Ca}
\usage{
gg_treemap_y_CaCaNu.(data, titleLabel = "", subtitle = "", caption = "",
  fillLabel = NULL, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_treemap_y_CaCaNu.
Treemap Fill by second Ca
}
\section{ftypes}{
 Ca-Ca-Nu
}

\examples{
add(1, 1)
add(10, 1)
}
