% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca.R
\name{gg_waffle_Ca.}
\alias{gg_waffle_Ca.}
\title{Waffle
Waffle}
\usage{
gg_waffle_Ca.(data, square_size = 1, rows_number = 5, titleLabel = "",
  fillLabel = NULL, subtitle = "", caption = "", leg_pos = "right", ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Waffle
Waffle
}
\section{ftypes}{
 Ca
}

\examples{
add(1, 1)
add(10, 1)
}
