% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Da-Nu_funs.R
\name{gg_waterfall_DaNu.}
\alias{gg_waterfall_DaNu.}
\title{gg_waterfall_DaNu.
Waterfall}
\usage{
gg_waterfall_DaNu.(data, titleLabel = "Report", xLabel = NULL,
  yLabel = NULL)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_waterfall_DaNu.
Waterfall
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

