% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nu-Nu.R
\name{gg_waterfall_NuNu.}
\alias{gg_waterfall_NuNu.}
\title{gg_waterfall_NuNu.
Waterfall}
\usage{
gg_waterfall_NuNu.(data, titleLabel = "", xLabel = NULL, yLabel = NULL,
  ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_waterfall_NuNu.
Waterfall
}
\section{ftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

